### 2021-05-04 #3241 @whedon
[PRE REVIEW]: ImSwitch: Generalizing microscope control in Python
**Submitting author:** @kasasxav (<a href="http://orcid.org/0000-0002-9583-9022">Xavier Casas Moreno</a>)
**Repository:** <a href="https://github.com/kasasxav/ImSwitch" target ="_blank">https://github.com/kasasxav/ImSwitch</a>
**Version:** v1.0.0
