### 2021-05-04 0a2e270
Add nidaq in setupInfo
I included the counter and triggering of the nidaq scan in the config file since some versions don't support those settings.
