+++ imswitch/imcontrol/model/SetupInfo.py
@@ -70,6 +70,11 @@ class ScanInfo:
     TTLCycleDesignerParams: Dict[str, Any]  # properties to be read by TTLCycleDesigner
     sampleRate: int  # scan sample rate
 
+@dataclass(frozen=True)
+class NidaqInfo:
+    counterChannel: int  # Output for Counter for timing purposes
+    startTrigger: bool = False # Boolean for start triggering for sync
+
 
 @dataclass_json(undefined=Undefined.INCLUDE)
 @dataclass(frozen=True)
@@ -83,6 +88,8 @@ class SetupInfo:
     focusLock: Optional[FocusLockInfo] = None
 
     scan: ScanInfo = field(default_factory=ScanInfo)
+
+    nidaq: NidaqInfo = field(default_factory=dict)
 
     _catchAll: CatchAll = None
 
