### 2021-05-05 a2c26a9
Conda script for initialization
