+++ conda_script/readme.txt
@@ -1,0 +1,1 @@
+For running the .bat script you need conda and an environment called imswitch.
