+++ conda_script/readme.txt
@@ -1,1 +1,3 @@
 For running the .bat script you need conda and an environment called imswitch.
+
+From: https://gist.github.com/maximlt 
