+++ conda_script/readme.txt
@@ -1,3 +1,3 @@
 For running the .bat script you need conda and an environment called imswitch.
 
-From: https://gist.github.com/maximlt 
+From: https://gist.github.com/maximlt/531419545b039fa33f8845e5bc92edd6
