### 2021-05-05 3d5e9b2
change requirements for script CONDANAME and ENVNAME
