+++ conda_script/readme.txt
@@ -1,5 +1,5 @@
 For running the .bat script you need conda.
-To make it work you have to edit the .bat file CONDAPATH and ENVNAME.
+To make it work you have to edit the .bat file and change CONDAPATH and ENVNAME.
  
 From: https://gist.github.com/maximlt/531419545b039fa33f8845e5bc92edd6
        
