### 2021-05-05 f061134
Change counterChannel -> timerCounterChannel
