+++ imswitch/imcontrol/model/managers/NidaqManager.py
@@ -22,7 +22,7 @@ class NidaqManager(SignalInterface):
         self.__setupInfo = setupInfo
         self.tasks = {}
         self.busy = False
-        self.__counterChannel = setupInfo.nidaq.counterChannel
+        self.__timerCounterChannel = setupInfo.nidaq.timerCounterChannel
         self.__startTrigger = setupInfo.nidaq.startTrigger
 
     def __makeSortedTargets(self, sortingKey):
@@ -274,12 +274,12 @@ class NidaqManager(SignalInterface):
             # create task waiters and change constants for beginning scan
             self.aoTaskWaiter = WaitThread()
             self.doTaskWaiter = WaitThread()
-            if self.__counterChannel is not None :
+            if self.__timerCounterChannel is not None :
                 self.timerTaskWaiter = WaitThread()
                 # create timer counter output task, to control the acquisition timing (1 MHz)
                 sampsInScan = np.int(len(AOsignals[0] if len(AOsignals) > 0 else DOsignals[0]) * 10)
                 self.timerTask = self.__createChanCOTask(
-                    'TimerTask', channel=self.__counterChannel, rate=1e6, sampsInScan=sampsInScan,
+                    'TimerTask', channel=self.__timerCounterChannel, rate=1e6, sampsInScan=sampsInScan,
                     starttrig=self.__startTrigger, reference_trigger='ao/StartTrigger'
                 )
                 self.timerTaskWaiter.connect(self.timerTask)
@@ -321,7 +321,7 @@ class NidaqManager(SignalInterface):
                 )
                 self.tasks['do'] = self.doTask
             self.sigScanInitiate.emit(scanInfoDict)
-            if self.__counterChannel is not None:
+            if self.__timerCounterChannel is not None:
                 self.tasks['timer'].start()
                 self.timerTaskWaiter.start()
             if len(DOsignals) > 0:
