### 2021-05-05 320324c
Fix timerCounterChannel change not updated in SetupInfo.py
