### 2021-05-05 f436314
LaserController remove aotf condition in valueChanged
Not needed anymore if the sliders are already disabled
