### 2021-05-05 ab64aa5
Remove aotfLasers in LaserController, not needed
