### 2021-05-05 b4ec02e
Binary also disabled when digMod
