### 2021-05-05 1aadac8
Fix digModule defined twice
