### 2021-05-05 2086683
Disable not-digital lasers powers during scan
