### 2021-05-05 2829f87
Disable not-digital lasers powers during scan
