### 2021-05-05 5ab17bc
Emit sigScanStarted when scan starts
