+++ imswitch/imcontrol/controller/controllers/ScanController.py
@@ -163,6 +163,7 @@ class ScanController(SuperScanController):
             return
 
         self._master.nidaqManager.runScan(self.signalDic, self.scanInfoDict)
+        self._commChannel.sigScanStarted.emit()
 
     def scanDone(self):
         print('Scan done')
