### 2021-05-06 b37549c
Make ScanWidget scrollable
