### 2021-05-06 1978510
ScanWidget make also horizontal scroll
