### 2021-05-06 2a41ae6
Update STED devices
