### 2021-05-07 b16dd5e
Create shortcut for conda users
