### 2021-05-07 de1c2e3
Lasers disabled while scan, and re-structure LaserController
