+++ imswitch/imcontrol/model/interfaces/SLM.py
@@ -123,13 +123,13 @@ class SLMframe(wx.Frame):
     def SetMonitor(self, monitor):
         tryMonitor = monitor
 
-        while monitor > wx.Display.GetCount() - 1:
-            monitor -= 1  # Try other monitor
+        while tryMonitor > wx.Display.GetCount() - 1:
+            tryMonitor -= 1  # Try other monitor
 
         if tryMonitor < 0:
             raise ValueError('Invalid monitor (monitor %d).' % monitor)
 
-        self._x0, self._y0, self._resX, self._resY = wx.Display(monitor).GetGeometry()
+        self._x0, self._y0, self._resX, self._resY = wx.Display(tryMonitor).GetGeometry()
         
 
 class videoThread(threading.Thread):
