+++ .github/workflows/imswitch-test.yml
@@ -2,9 +2,9 @@ name: imswitch-test
 
 on:
   push:
-    branches: [ master ]
+    branches: [ dev ]
   pull_request:
-    branches: [ master ]
+    branches: [ dev ]
 
 jobs:
   build:
