### 2021-05-07 f0e4879
CoolLED manager, json and use-cases
