### 2021-05-07 ec89edd
Remove headline file creation
