### 2021-05-17 a908041
Make tests run for every branch, checking out appropriate branch
