+++ .github/workflows/imswitch-test.yml
@@ -1,10 +1,6 @@
 name: imswitch-test
 
-on:
-  push:
-    branches: [ dev ]
-  pull_request:
-    branches: [ dev ]
+on: [ push, pull_request ]
 
 jobs:
   build:
@@ -17,7 +13,7 @@ jobs:
     steps:
     - uses: actions/checkout@v2
       with:
-        ref: master
+        ref: ${{ github.ref }}
     - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v2
       with:
