### 2021-05-17 d132fd2
Fix missing import in RS232sManager
