+++ imswitch/imcontrol/model/managers/RS232sManager.py
@@ -1,3 +1,6 @@
+from .MultiManager import MultiManager
+
+
 class RS232sManager(MultiManager):
     """ LasersManager is an interface for dealing with RS232 devices. It is a
     MultiManager for RS232 devices. """
