+++ docs/imcontrol-setups.rst
@@ -17,8 +17,8 @@ camera.
 
 Available manager properties:
 
-- cameraListIndex -- the camera's index in the Hamamatsu camera list (list indexing starts at 0); set this to an invalid value, e.g. the string "mock" to load a mocker
-- hamamatsu -- dictionary of DCAM API properties
+- ``cameraListIndex`` -- the camera's index in the Hamamatsu camera list (list indexing starts at 0); set this to an invalid value, e.g. the string "mock" to load a mocker
+- ``hamamatsu`` -- dictionary of DCAM API properties
 
 
 APDManager
@@ -28,8 +28,8 @@ DetectorManager that deals with an avalanche photodiode connected to a counter i
 
 Available manager properties:
 
-- terminal -- the physical input terminal on the Nidaq to which the APD is connected
-- ctrInputLine -- the counter that the physical input terminal is connected to
+- ``terminal`` -- the physical input terminal on the Nidaq to which the APD is connected
+- ``ctrInputLine`` -- the counter that the physical input terminal is connected to
 
 
 TISManager
@@ -39,8 +39,8 @@ DetectorManager that deals with TheImagingSource cameras and the parameters for 
 
 Available manager properties:
 
-- cameraListIndex -- the camera's index in the TIS camera list (list indexing starts at 0); set this string to an invalid value to load a mocker
-- tis -- dictionary of TIS camera properties
+- ``cameraListIndex`` -- the camera's index in the TIS camera list (list indexing starts at 0); set this string to an invalid value to load a mocker
+- ``tis`` -- dictionary of TIS camera properties
 
 
 Laser managers
@@ -54,8 +54,8 @@ drivers available through Lantz.
 
 Available manager properties:
 
-- digitalDriver -- a string containing a Lantz driver name, e.g. "cobolt.cobolt0601.Cobolt0601"
-- digitalPorts -- a string array containing the COM ports to connect to, e.g. ["COM4"]
+- ``digitalDriver`` -- a string containing a Lantz driver name, e.g. "cobolt.cobolt0601.Cobolt0601"
+- ``digitalPorts`` -- a string array containing the COM ports to connect to, e.g. ["COM4"]
 
 
 NidaqAOLaserManager
@@ -73,8 +73,8 @@ LaserManager for controlling one channel of an AA Opto-Electronic acousto-optic 
 
 Available manager properties:
 
-- rs232device -- name of the defined rs232 communication channel through which the communication should take place
-- channel -- index of the channel in the acousto-optic device that should be controlled (indexing starts at 1)
+- ``rs232device`` -- name of the defined rs232 communication channel through which the communication should take place
+- ``channel`` -- index of the channel in the acousto-optic device that should be controlled (indexing starts at 1)
 
 CoolLEDLaserManager
 ---------------------
@@ -93,7 +93,7 @@ LaserManager for controlling a OneFive Katana pulsed laser (NKT Photonics).
 
 Available manager properties:
 
-- rs232device -- name of the defined rs232 communication channel through which the communication should take place 
+- ``rs232device`` -- name of the defined rs232 communication channel through which the communication should take place 
 
 
 Positioner managers
@@ -106,9 +106,9 @@ PositionerManager for analog NI-DAQ-controlled positioners.
 
 Available manager properties:
 
-- conversionFactor -- float
-- minVolt -- minimum voltage
-- maxVolt -- maximum voltage
+- ``conversionFactor`` -- float
+- ``minVolt`` -- minimum voltage
+- ``maxVolt`` -- maximum voltage
 
 
 PiezoconceptZManager
@@ -118,7 +118,7 @@ PositionerManager for control of a Piezoconcept Z-piezo through RS232-communicat
 
[…]
