### 2021-05-17 0060473
Clarify installation and requirements docs
