### 2021-05-17 a3c6a5e
Add module configuration documentation
