### 2021-05-17 325b033
Remove unneeded line requirements.txt (since python < 3.7 not supported)
