+++ requirements.txt
@@ -1,5 +1,4 @@
 colour-science~=0.3.15
-dataclasses; python_version < "3.7"
 dataclasses-json~=0.5.2
 dotmap~=1.3.23
 h5py~=3.1.0
