### 2021-05-18 1a259a5
Remove some technical stuff from documentation index page
