### 2021-05-23 8c37ca3
Generalize Lantz lasers and make digmod automatic
