+++ imswitch/imcontrol/view/widgets/LaserWidget.py
@@ -5,13 +5,10 @@ from .basewidgets import Widget
 
 
 class LaserWidget(Widget):
-    """ Laser widget containing digital modulation and normal control. """
+    """ Laser widget for setting laser powers etc. """
 
     sigEnableChanged = QtCore.Signal(str, bool)  # (laserName, enabled)
     sigValueChanged = QtCore.Signal(str, float)  # (laserName, value)
-
-    sigDigitalModToggled = QtCore.Signal(bool)  # (enabled)
-    sigDigitalValueChanged = QtCore.Signal(str, float)  # (laserName, value)
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -22,25 +19,18 @@ class LaserWidget(Widget):
         self.setLayout(self.scrollContainer)
 
         self.grid = QtWidgets.QGridLayout()
-        self.digModule = DigitalModule()
-        self.grid.addWidget(self.digModule, 4, 0, 2, -1)
 
         self.gridContainer = QtWidgets.QWidget()
         self.gridContainer.setLayout(self.grid)
 
         self.scrollArea = QtWidgets.QScrollArea()
         self.scrollArea.setFrameShape(QtWidgets.QFrame.NoFrame)
-        self.scrollArea.setHorizontalScrollBarPolicy(QtCore.Qt.ScrollBarAsNeeded)
-        self.scrollArea.setVerticalScrollBarPolicy(QtCore.Qt.ScrollBarAlwaysOff)
+        self.scrollArea.setHorizontalScrollBarPolicy(QtCore.Qt.ScrollBarAlwaysOff)
+        self.scrollArea.setVerticalScrollBarPolicy(QtCore.Qt.ScrollBarAsNeeded)
         self.scrollArea.setWidget(self.gridContainer)
         self.scrollArea.setWidgetResizable(True)
         self.scrollContainer.addWidget(self.scrollArea)
         self.gridContainer.installEventFilter(self)
-
-        # Connect signals
-        self.digModule.sigDigitalModToggled.connect(self.sigDigitalModToggled)
-        self.digModule.sigDigitalModToggled.connect(lambda: print(self.gridContainer.minimumSizeHint().height()))
-        self.digModule.sigDigitalValueChanged.connect(self.sigDigitalValueChanged)
 
     def addLaser(self, laserName, valueUnits, wavelength, valueRange=None, valueRangeStep=1):
         """ Adds a laser module widget. valueRange is either a tuple
@@ -58,14 +48,14 @@ class LaserWidget(Widget):
             lambda value: self.sigValueChanged.emit(laserName, value)
         )
 
-        self.grid.addWidget(control, 0, len(self.laserModules), 4, 1)
+        nameLabel = QtWidgets.QLabel(f'<h3>{laserName}<h3>')
+        nameLabel.setAlignment(QtCore.Qt.AlignVCenter)
+        color = guitools.colorutils.wavelengthToHex(wavelength)
+        nameLabel.setStyleSheet(f'font-size: 16px; padding-left: 8px; border-left: 4px solid {color}')
+
+        self.grid.addWidget(nameLabel, len(self.laserModules), 0)
+        self.grid.addWidget(control, len(self.laserModules), 1)
         self.laserModules[laserName] = control
-
-        self.digModule.addLaser(laserName, valueUnits, valueRange)
-
-    def isDigModActive(self):
-        """ Returns whether the digital modulation mode is active. """
-        return self.digModule.isActive()
 
     def isLaserActive(self, laserName):
         """ Returns whether the specified laser is powered on. """
@@ -75,15 +65,6 @@ class LaserWidget(Widget):
[…]
