### 2021-05-25 93d6c1e
Fix CoboltLaserManager _setModPower not working
