### 2021-05-25 3c1939f
Add presets feature to laser widget
