+++ imswitch/imcontrol/controller/controllers/LaserController.py
@@ -1,4 +1,6 @@
 from imswitch.imcommon.model import APIExport
+from imswitch.imcontrol.model import configfiletools
+from imswitch.imcontrol.view import guitools
 from .basecontrollers import ImConWidgetController
 
 
@@ -9,6 +11,7 @@ class LaserController(ImConWidgetController):
         super().__init__(*args, **kwargs)
 
         self.settingAttr = False
+        self.presetBeforeScan = None
 
         # Set up lasers
         for lName, lManager in self._master.lasersManager:
@@ -23,6 +26,13 @@ class LaserController(ImConWidgetController):
             self.setSharedAttr(lName, _enabledAttr, self._widget.isLaserActive(lName))
             self.setSharedAttr(lName, _valueAttr, self._widget.getValue(lName))
 
+        # Load presets
+        for laserPresetName in self._setupInfo.laserPresets:
+            self._widget.addPreset(laserPresetName)
+
+        self._widget.setCurrentPreset(None)  # Unselect
+        self._widget.setScanDefaultPreset(self._setupInfo.defaultLaserPresetForScan)
+
         # Connect CommunicationChannel signals
         self._commChannel.sharedAttrs.sigAttributeSet.connect(self.attrChanged)
         self._commChannel.sigScanStarted.connect(lambda: self.scanChanged(True))
@@ -31,6 +41,12 @@ class LaserController(ImConWidgetController):
         # Connect LaserWidget signals
         self._widget.sigEnableChanged.connect(self.toggleLaser)
         self._widget.sigValueChanged.connect(self.valueChanged)
+
+        self._widget.sigPresetSelected.connect(self.presetSelected)
+        self._widget.sigLoadPresetClicked.connect(self.loadPreset)
+        self._widget.sigSavePresetClicked.connect(self.savePreset)
+        self._widget.sigDeletePresetClicked.connect(self.deletePreset)
+        self._widget.sigPresetScanDefaultToggled.connect(self.presetScanDefaultToggled)
 
     def closeEvent(self):
         self._master.lasersManager.execOnAll(lambda l: l.setScanModeActive(False))
@@ -45,10 +61,133 @@ class LaserController(ImConWidgetController):
         """ Change magnitude. """
         self._master.lasersManager[laserName].setValue(magnitude)
         self._widget.setValue(laserName, magnitude)
+        self._widget.setCurrentPreset(None)
         self.setSharedAttr(laserName, _valueAttr, magnitude)
 
+    def presetSelected(self, presetName):
+        """ Handles what happens when a preset is selected in the preset list.
+        """
+        if presetName:
+            self._widget.setCurrentPreset(presetName)
+
+        self._widget.setScanDefaultPresetActive(
+            self._setupInfo.defaultLaserPresetForScan == presetName
+        )
+
+    def loadPreset(self):
+        """ Handles what happens when the user requests the selected preset to
+        be loaded. """
+        presetToLoad = self._widget.getCurrentPreset()
+        if not presetToLoad:
+            return
+
+        if presetToLoad not in self._setupInfo.laserPresets:
+            return
+
+        # Load values
+        self.applyPreset(self._setupInfo.laserPresets[presetToLoad])
+
+        # Keep preset selected in GUI
+        self._widget.setCurrentPreset(presetToLoad)
+
+    def savePreset(self):
+        """ Handles what happens when the user requests the current laser
+        values to be saved as a preset. """
+
+        name = guitools.askForTextInput(self._widget, 'Add laser preset',
+                                        'Enter a name for this preset:')
+
+        if not name:  # No name provided
+            return
+
+        add = True
+        alreadyExists = False
+        if name in self._setupInfo.laserPresets:
+            alreadyExists = True
+            add = guitools.askYesNoQuestion(
+                self._widget,
+                'Laser preset already exists',
[…]
