+++ imswitch/imcontrol/view/widgets/LaserWidget.py
@@ -10,35 +10,76 @@ class LaserWidget(Widget):
     sigEnableChanged = QtCore.Signal(str, bool)  # (laserName, enabled)
     sigValueChanged = QtCore.Signal(str, float)  # (laserName, value)
 
+    sigPresetSelected = QtCore.Signal(str)  # (presetName)
+    sigLoadPresetClicked = QtCore.Signal()
+    sigSavePresetClicked = QtCore.Signal()
+    sigDeletePresetClicked = QtCore.Signal()
+    sigPresetScanDefaultToggled = QtCore.Signal()
+
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.laserModules = {}
 
-        self.scrollContainer = QtWidgets.QGridLayout()
-        self.scrollContainer.setContentsMargins(0, 0, 0, 0)
-        self.setLayout(self.scrollContainer)
+        self.setMinimumHeight(320)
 
-        self.grid = QtWidgets.QGridLayout()
+        self.layout = QtWidgets.QGridLayout()
+        self.setLayout(self.layout)
 
-        self.gridContainer = QtWidgets.QWidget()
-        self.gridContainer.setLayout(self.grid)
+        # Lasers grid
+        self.lasersGrid = QtWidgets.QGridLayout()
+        self.lasersGrid.setContentsMargins(4, 4, 4, 4)
+
+        self.lasersGridContainer = QtWidgets.QWidget()
+        self.lasersGridContainer.setLayout(self.lasersGrid)
 
         self.scrollArea = QtWidgets.QScrollArea()
         self.scrollArea.setFrameShape(QtWidgets.QFrame.NoFrame)
         self.scrollArea.setHorizontalScrollBarPolicy(QtCore.Qt.ScrollBarAlwaysOff)
         self.scrollArea.setVerticalScrollBarPolicy(QtCore.Qt.ScrollBarAsNeeded)
-        self.scrollArea.setWidget(self.gridContainer)
+        self.scrollArea.setWidget(self.lasersGridContainer)
         self.scrollArea.setWidgetResizable(True)
-        self.scrollContainer.addWidget(self.scrollArea)
-        self.gridContainer.installEventFilter(self)
+        self.lasersGridContainer.installEventFilter(self)
+
+        self.layout.addWidget(self.scrollArea, 0, 0)
+
+        # Presets box
+        self.presetsBox = QtWidgets.QHBoxLayout()
+        self.presetsLabel = QtWidgets.QLabel('Presets: ')
+        self.presetsList = QtWidgets.QComboBox()
+        self.presetsList.currentIndexChanged.connect(
+            lambda i: self.sigPresetSelected.emit(self.presetsList.itemData(i))
+        )
+        self.loadPresetButton = guitools.BetterPushButton('Load selected')
+        self.loadPresetButton.clicked.connect(self.sigLoadPresetClicked)
+        self.savePresetButton = guitools.BetterPushButton('Save values…')
+        self.savePresetButton.clicked.connect(self.sigSavePresetClicked)
+        self.moreButton = QtWidgets.QToolButton()
+        self.moreButton.setText('More…')
+        self.moreButton.setPopupMode(QtWidgets.QToolButton.ToolButtonPopupMode.InstantPopup)
+        self.deletePresetAction = QtWidgets.QAction('Delete selected')
+        self.deletePresetAction.triggered.connect(self.sigDeletePresetClicked)
+        self.moreButton.addAction(self.deletePresetAction)
+        self.presetScanDefaultAction = QtWidgets.QAction('Make selected default for scanning')
+        self.presetScanDefaultAction.triggered.connect(self.sigPresetScanDefaultToggled)
[…]
