### 2021-05-26 3a7117f
Add comment to explain use of eventFilter
