### 2021-05-26 6867b70
Fix label alignment in new laser widget
