### 2021-05-27 464717b
Print message when setting CoboltLaserManager mod power
