### 2021-05-27 6c810f6
Rename mock lantz laser driver to Cobolt mock laser driver
Since it can only mock Cobolt lasers, not any lantz laser.
