+++ imswitch/imcontrol/model/drivers_mock/cobolt/cobolt0601.py
@@ -3,9 +3,9 @@ from lantz import Driver
 from lantz import Q_
 
 
-class MockLaser(Driver):
+class Cobolt0601_f2(Driver):
 
-    def __init__(self):
+    def __init__(self, *args, **kwargs):
         super().__init__()
 
         self.mW = Q_(1, 'mW')
