### 2021-05-27 4b90194
Fix missing property tag on mock Cobolt laser's mod_mode
