### 2021-05-27 5f1d13a
Fix infinite recursion in LinkedLantzLaser
