### 2021-05-27 561c3c6
Clarify documentation related to LantzLaserManager/CoboltLaserManager
