### 2021-05-27 f55f89f
Alphabetical order of managers in imcontrol setups documentation
