### 2021-05-27 0a9588b
Fix Cobolt laser manager behaviour with linked lasers
