### 2021-05-31 b24cc1f
Fix deprecated usage of Napari viewer.active_layer
