### 2021-05-31 384b10c
Make it possible to pick modules to load inside the program
