+++ imswitch/imcommon/controller/__init__.py
@@ -1,2 +1,3 @@
 from .basecontrollers import MainController, WidgetController, WidgetControllerFactory
 from .ModuleCommunicationChannel import ModuleCommunicationChannel
+from .MultiModuleWindowController import MultiModuleWindowController
