### 2021-06-17 d961733
Fix napari version to 0.4.7 (because of napari issue 2897)
