+++ requirements.txt
@@ -3,7 +3,7 @@ dataclasses-json~=0.5.2
 dotmap~=1.3.23
 h5py~=3.1.0
 lantzdev[qt]~=0.5.2
-napari[pyqt5]~=0.4.6
+napari[pyqt5]==0.4.7
 nidaqmx~=0.5.7
 numpy==1.19.3
 psutil~=5.7.3
