### 2021-06-19 5be5def
Fix crosshair not unlocking until after mouse is moved
