### 2021-06-19 5d53778
Add napari dock for shifting to imcontrol (resolves #30)
