### 2021-06-19 8bbdefd
Change Qt import in naparitools to qtpy
