### 2021-06-19 52da6d4
Add comment to NapariShiftWidget
