### 2021-06-19 392191f
Update levels btn as napari widget & add to imreconstruct (resolves #32)
