### 2021-06-19 968d311
Avoid asterisk imports in guitools
