### 2021-06-19 9281655
Make MockHamamatsu output more correct # of frames with internal trigger
