### 2021-06-20 cbefead
Avoid losing frames at the end of recordings (#33)
Involves calculating number of scan positions for scan recordings.
