+++ imswitch/imcontrol/controller/controllers/RecordingController.py
@@ -19,7 +19,7 @@ class RecordingController(ImConWidgetController):
 
         self.settingAttr = False
         self.recording = False
-        self.lapseCurrent = 0
+        self.lapseCurrent = -1
         self.lapseTotal = 0
 
         imreconstructRegistered = self._moduleCommChannel.isModuleRegistered('imreconstruct')
@@ -30,9 +30,8 @@ class RecordingController(ImConWidgetController):
 
         # Connect CommunicationChannel signals
         self._commChannel.sigRecordingEnded.connect(self.recordingEnded)
-        self._commChannel.sigUpdateRecFrameNum.connect(self._widget.updateRecFrameNum)
-        self._commChannel.sigUpdateRecTime.connect(self._widget.updateRecTime)
-        self._commChannel.sigScanEnded.connect(self.scanDone)
+        self._commChannel.sigUpdateRecFrameNum.connect(self.updateRecFrameNum)
+        self._commChannel.sigUpdateRecTime.connect(self.updateRecTime)
         self._commChannel.sharedAttrs.sigAttributeSet.connect(self.attrChanged)
 
         # Connect RecordingWidget signals
@@ -100,7 +99,9 @@ class RecordingController(ImConWidgetController):
                     *recordingArgs, recTime=self._widget.getTimeToRec()
                 )
             elif self.recMode == RecMode.ScanOnce:
-                self._master.recordingManager.startRecording(*recordingArgs)
+                self._master.recordingManager.startRecording(
+                    *recordingArgs, recFrames=self._commChannel.getNumScanPositions()
+                )
                 time.sleep(0.1)
                 self._commChannel.sigPrepareScan.emit()
             elif self.recMode == RecMode.ScanLapse:
@@ -114,36 +115,38 @@ class RecordingController(ImConWidgetController):
         else:
             self._master.recordingManager.endRecording()
 
-    def scanDone(self):
-        if self._widget.recButton.isChecked():
-            if self.recMode == RecMode.ScanOnce:
-                self._master.recordingManager.endRecording()
-            elif self.recMode == RecMode.ScanLapse:
-                self.lapseCurrent += 1
-                if self.lapseCurrent < self.lapseTotal:
-                    self._master.recordingManager.endRecording(emitSignal=False)
-                    self._widget.updateRecLapseNum(self.lapseCurrent)
-                    self.timer = Timer(singleShot=True)
-                    self.timer.timeout.connect(self.nextLapse)
-                    self.timer.start(int(self._widget.getTimelapseFreq() * 1000))
-                else:
-                    self._master.recordingManager.endRecording()
-
     def nextLapse(self):
         fileName = self.savename + "_" + str(self.lapseCurrent).zfill(len(str(self.lapseTotal)))
         self._master.recordingManager.startRecording(
-            self.detectorsBeingCaptured, self.recMode, fileName, self.saveMode, self.attrs
+            self.detectorsBeingCaptured, self.recMode, fileName, self.saveMode, self.attrs,
+            recFrames=self._commChannel.getNumScanPositions()
         )
         time.sleep(0.3)
         self._commChannel.sigPrepareScan.emit()
 
     def recordingEnded(self):
-        self.recording = False
-        self.lapseCurrent = 0
-        self._widget.updateRecFrameNum(0)
-        self._widget.updateRecTime(0)
-        self._widget.updateRecLapseNum(0)
-        self._widget.setRecButtonChecked(False)
+        if (self._widget.isRecButtonChecked() and self.recMode == RecMode.ScanLapse and
+                0 < self.lapseCurrent + 1 < self.lapseTotal):
+            self.lapseCurrent += 1
[…]
