+++ imswitch/imcontrol/model/managers/RecordingManager.py
@@ -132,32 +132,48 @@ class RecordingWorker(Worker):
                 files[detectorName].attrs[key] = value
 
         try:
-            if self.recMode == RecMode.SpecFrames:
-                if len(self.detectorNames) > 1:
-                    raise ValueError('Only one detector can be recorded in SpecFrames mode')
+            if self.recMode in [RecMode.SpecFrames, RecMode.ScanOnce, RecMode.ScanLapse]:
+                recFrames = self.recFrames
+                if recFrames is None:
+                    raise ValueError('recFrames must be specified in SpecFrames, ScanOnce or'
+                                     ' ScanLapse mode')
 
-                detectorName = self.detectorNames[0]
-                frames = self.recFrames
-                while self.__recordingManager.record and currentFrame[detectorName] < frames:
-                    newFrames = self._getNewFrames(detectorName)
-                    n = len(newFrames)
-                    if n > 0:
-                        it = currentFrame[detectorName]
-                        dataset = datasets[detectorName]
-                        if (it + n) <= frames:
-                            dataset.resize(n + it, axis=0)
-                            dataset[it:it + n, :, :] = newFrames
-                            currentFrame[detectorName] += n
-                        else:
-                            dataset.resize(frames, axis=0)
-                            dataset[it:frames, :, :] = newFrames[0:frames - it]
-                            currentFrame[detectorName] = frames
-                        self.__recordingManager.sigRecordingFrameNumUpdated.emit(it)
+                while self.__recordingManager.record and any([currentFrame[detectorName] < recFrames
+                                                              for detectorName in self.detectorNames]):
+                    for detectorName in self.detectorNames:
+                        if currentFrame[detectorName] >= recFrames:
+                            continue  # Reached requested number of frames with this detector, skip
+
+                        newFrames = self._getNewFrames(detectorName)
+                        n = len(newFrames)
+                        if n > 0:
+                            it = currentFrame[detectorName]
+                            dataset = datasets[detectorName]
+                            if (it + n) <= recFrames:
+                                dataset.resize(n + it, axis=0)
+                                dataset[it:it + n, :, :] = newFrames
+                                currentFrame[detectorName] += n
+                            else:
+                                dataset.resize(recFrames, axis=0)
+                                dataset[it:recFrames, :, :] = newFrames[0:recFrames - it]
+                                currentFrame[detectorName] = recFrames
+
+                            # Things get a bit weird if we have multiple detectors when we report
+                            # the current frame number, since the detectors may not be synchronized.
+                            # For now, we will report the lowest number.
+                            self.__recordingManager.sigRecordingFrameNumUpdated.emit(
+                                min(list(currentFrame.values()))
+                            )
                 self.__recordingManager.sigRecordingFrameNumUpdated.emit(0)
             elif self.recMode == RecMode.SpecTime:
+                recTime = self.recTime
+                if recTime is None:
+                    raise ValueError('recTime must be specified in SpecFrames or ScanLapse mode')
+
                 start = time.time()
                 currentRecTime = 0
-                while self.__recordingManager.record and currentRecTime < self.recTime:
+                shouldStop = False
+                while True:
                     for detectorName in self.detectorNames:
                         newFrames = self._getNewFrames(detectorName)
                         n = len(newFrames)
@@ -171,9 +187,16 @@ class RecordingWorker(Worker):
                                 np.around(currentRecTime, decimals=2)
                             )
                             currentRecTime = time.time() - start
+
+                    if shouldStop:
+                        break  # Enter loop one final time, then stop
+
+                    if not self.__recordingManager.record or currentRecTime >= recTime:
[…]
