### 2021-06-20 502c2b9
Disable recording widget fields while recording & add sigRecordingStarted
