+++ imswitch/imcontrol/view/widgets/RecordingWidget.py
@@ -95,8 +95,10 @@ class RecordingWidget(Widget):
                                    QtWidgets.QSizePolicy.Expanding)
         buttonGrid.addWidget(self.recButton, 0, 2)
 
+        layout = QtWidgets.QVBoxLayout()
+        self.setLayout(layout)
+
         recGrid = QtWidgets.QGridLayout()
-        self.setLayout(recGrid)
 
         recGrid.addWidget(recTitle, 0, 0, 1, 3)
         recGrid.addWidget(QtWidgets.QLabel('Detector to capture'), 1, 0)
@@ -125,7 +127,12 @@ class RecordingWidget(Widget):
         recGrid.addWidget(self.untilSTOPbtn, 10, 0, 1, -1)
         recGrid.addWidget(self.saveModeLabel, 12, 0)
         recGrid.addWidget(self.saveModeList, 12, 1, 1, -1)
-        recGrid.addWidget(buttonWidget, 13, 0, 1, -1)
+
+        self.recGridContainer = QtWidgets.QWidget()
+        self.recGridContainer.setLayout(recGrid)
+
+        layout.addWidget(self.recGridContainer)
+        layout.addWidget(buttonWidget)
 
         # Initial condition of fields and checkboxes.
         self.writable = True
@@ -226,6 +233,9 @@ class RecordingWidget(Widget):
 
     def checkUntilStop(self):
         self.untilSTOPbtn.setChecked(True)
+
+    def setFieldsEnabled(self, enabled):
+        self.recGridContainer.setEnabled(enabled)
 
     def setEnabledParams(self, numExpositions=False, timeToRec=False,
                          timelapseTime=False, timelapseFreq=False):
