### 2021-06-20 df36d72
Make custom napari widgets apply to all selected layers
