### 2021-06-20 6a17008
Add screenshot to module configuration docs
