### 2021-06-20 #3241 @MakerTobey
This looks like a very exciting tool. I'm neither a classical microscopy person nor a great Python programmer, but this software looks very professional to me. It does not only (at least at brief view) look like it has all the main features of a professional Nikon etc. platform, but it also has very exciting and advances hardware control interfaces to control lasers, fast photon detector modules, spatial light modulators and other units that are needed for advances microscopy (confocal, super resolution, scanning, etc.). I will definitely keep this in mind for my own future projects. Also: Thank you for using Phython, which is FOSS and in my eyes the best scientific platform to interface with custom hardware. I have not tried to execute the software (yet).
