### 2021-06-20 #3241 @Kevin-Mattheus-Moerman
@beniroquai @nicost @henrypinkard @untzag @nanthony21 @uellue would you be interested in reviewing this work for [the Journal of Open Source Software (JOSS)](https://joss.theoj.org/about): _ImSwitch: Generalizing microscope control in Python_

The [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) focuses on [the software](https://github.com/kasasxav/ImSwitch) and a [short paper](https://github.com/openjournals/joss-papers/blob/joss.03241/joss.03241/10.21105.joss.03241.pdf).

If you are interested in helping you can comment here. This here is a pre-review issue. The review process takes place in a review issue which I'll initiate when sufficient reviewers are recruited. Here is an example of a completed review issue: https://github.com/openjournals/joss-reviews/issues/2006, where you can see we have a streamlined GitHub based review process featuring a set of tick-boxes to guide the reviewers through the process.

Let me know if you have any questions.

Thanks!
