### 2021-06-21 bd78477
Add a loading screen while loading modules (#26)
