+++ imswitch/imscripting/__init__.py
@@ -1,7 +1,7 @@
 __title__ = 'Scripting'
 
 
-def getMainViewAndController(moduleCommChannel, multiModuleWindow, moduleMainControllers,
+def getMainViewAndController(moduleCommChannel, multiModuleWindowController, moduleMainControllers,
                              *_args, **_kwargs):
     from .controller import ImScrMainController
     from .view import ImScrMainView
@@ -11,7 +11,7 @@ def getMainViewAndController(moduleCommChannel, multiModuleWindow, moduleMainCon
         controller = ImScrMainController(
             view,
             moduleCommChannel=moduleCommChannel,
-            multiModuleWindow=multiModuleWindow,
+            multiModuleWindowController=multiModuleWindowController,
             moduleMainControllers=moduleMainControllers
         )
     except Exception as e:
