+++ imswitch/imcommon/view/MultiModuleWindow.py
@@ -1,38 +1,62 @@
 from pyqtgraph.Qt import QtCore, QtGui, QtWidgets
 
-from imswitch.imcommon.model import APIExport
-
 
 class MultiModuleWindow(QtWidgets.QMainWindow):
     sigPickModules = QtCore.Signal()
+    sigModuleAdded = QtCore.Signal(str, str)  # (moduleId, moduleName)
 
     def __init__(self, title, iconPath=None, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self._moduleIdNameMap = {}
+        self.moduleWidgets = {}
 
         self.setWindowTitle(title)
         if iconPath:
             self.setWindowIcon(QtGui.QIcon(iconPath))
 
-        # Create fallback menu bar
-        self._addItemsToMenuBar(self.menuBar())
-
         # Add tabs
         self.moduleTabs = QtWidgets.QTabWidget()
         self.moduleTabs.setTabPosition(QtWidgets.QTabWidget.TabPosition.West)
 
-        self.setCentralWidget(self.moduleTabs)
+        # Display loading screen until show(showLoadingScreen=False) is called
+        loadingLabel = QtWidgets.QLabel('<h1>Loading ImSwitch…</h1>')
+        loadingLabel.setAlignment(QtCore.Qt.AlignCenter)
+        self.loadingProgressBar = QtWidgets.QProgressBar()
+        self.loadingProgressBar.setMaximumWidth(480)
+        self.loadingProgressBar.setValue(0)
+
+        loadingLayout = QtWidgets.QVBoxLayout()
+        loadingLayout.setAlignment(QtCore.Qt.AlignCenter)
+        loadingLayout.setSpacing(32)
+        loadingLayout.addWidget(loadingLabel)
+        loadingLayout.addWidget(self.loadingProgressBar)
+
+        loadingContainer = QtWidgets.QWidget()
+        loadingContainer.setLayout(loadingLayout)
+        self.setCentralWidget(loadingContainer)
 
     def addModule(self, moduleId, moduleName, moduleWidget):
-        self._moduleIdNameMap[moduleId] = moduleName
+        self.moduleWidgets[moduleName] = moduleWidget
         self.moduleTabs.addTab(moduleWidget, moduleName)
 
         if hasattr(moduleWidget, 'menuBar') and callable(moduleWidget.menuBar):
             # Module widget has menu bar; add common items and hide fallback menu bar
-            self._addItemsToMenuBar(moduleWidget.menuBar())
+            self.addItemsToMenuBar(moduleWidget.menuBar())
             self.menuBar().hide()
 
-    def _addItemsToMenuBar(self, menuBar):
+        self.sigModuleAdded.emit(moduleId, moduleName)
+
+    def setSelectedModule(self, moduleName):
+        """ Sets the currently displayed module to the module with the
+        specified display name. """
+        for i in range(self.moduleTabs.count()):
+            if self.moduleTabs.tabText(i) == moduleName:
+                self.moduleTabs.setCurrentIndex(i)
+                return
+
+    def setLoadingProgress(self, progressFraction):
+        self.loadingProgressBar.setValue(progressFraction * 100)
+
+    def addItemsToMenuBar(self, menuBar):
         menuChildren = menuBar.findChildren(QtWidgets.QMenu, None, QtCore.Qt.FindDirectChildrenOnly)
         toolsMenu = None
         for menuChild in menuChildren:
@@ -50,19 +74,16 @@ class MultiModuleWindow(QtWidgets.QMainWindow):
         pickModulesAction.triggered.connect(self.sigPickModules)
         toolsMenu.addAction(pickModulesAction)
 
-    @APIExport
-    def setCurrentModule(self, moduleId):
[…]
