### 2021-06-21 0197da9
Remove unused UI items from ImageWidget
