### 2021-06-21 1b27e77
Hide progress bar text and change label text on loading screen
