+++ imswitch/imcontrol/model/interfaces/__init__.py
@@ -1,3 +1,3 @@
 from .hamamatsu import HamamatsuCamera, HamamatsuCameraMR
 from .hamamatsu_mock import MockHamamatsu
-from .lasers import LantzLaser
+from .lantzlasers import LantzLaser
