### 2021-06-23 03fabb4
Load mocker on lantz laser initialize() fail, make errors more clear
