### 2021-06-23 4d8f663
Fix inverted coordinates when showing ROI
