### 2021-06-23 bcaff98
Rename getCenterROI to getCenterViewbox
