+++ imswitch/imcontrol/controller/controllers/BeadRecController.py
@@ -20,7 +20,7 @@ class BeadRecController(ImConWidgetController):
             self.addROI()
 
             ROIsize = (64, 64)
-            ROIcenter = self._commChannel.getCenterROI()
+            ROIcenter = self._commChannel.getCenterViewbox()
 
             ROIpos = (ROIcenter[0] - 0.5 * ROIsize[0],
                       ROIcenter[1] - 0.5 * ROIsize[1])
