+++ imswitch/imcontrol/controller/controllers/AlignAverageController.py
@@ -36,7 +36,7 @@ class AlignAverageController(LiveUpdatedController):
             self.addROI()
 
             ROIsize = (64, 64)
-            ROIcenter = self._commChannel.getCenterROI()
+            ROIcenter = self._commChannel.getCenterViewbox()
 
             ROIpos = (ROIcenter[0] - 0.5 * ROIsize[0],
                       ROIcenter[1] - 0.5 * ROIsize[1])
