### 2021-06-24 d1735df
Fix unable to open files saved in DiskAndRAM mode before exiting
