### 2021-06-24 7fdfa5f
Fix Timelaps issue
Since now the scanning stops itself after all the frames are collected, the RecordingController should check for both scanning- and recordingEnded.
