### 2021-06-24 #3394 @untzag
Andreas Bodén and Ilaria Testa are listed as fourth and fifth authors, but I don't see them listed as contributors [1] to the software repository. @kasasxav could you clarify how these individuals contributed to ImSwitch? JOSS authorship guidelines might be relevant here [2].

[1] https://github.com/kasasxav/ImSwitch/graphs/contributors
[2] https://joss.readthedocs.io/en/latest/submitting.html#authorship
