### 2021-06-24 kasasxav/ImSwitch#40 @untzag
packaging
ImSwitch does not provide an installable Python package. Please provide `setup.py` or `pyproject.toml` so that users may actually install ImSwitch to their site packages. Running ImSwitch should be done using an entry point [1] so that the software is properly integrated into the operating system. I'm happy to provide pointers if you need help packaging.

(This issue must be resolved for me to approve publication in JOSS)

[1] https://packaging.python.org/specifications/entry-points/
