### 2021-06-24 #3394 @untzag
@kasasxav I feel strongly that ImSwitch should provide `setup.py` or `pyproject.toml` so that the software may be installed as a "proper" Python package. To me this issue block publication in JOSS.

See https://github.com/kasasxav/ImSwitch/issues/40, we can discuss there if you need some pointers on how to proceed with packaging.
