### 2021-06-24 kasasxav/ImSwitch#40 @untzag
One thing that you will need to do is not store config within your own repository. You should use the operating system's configuration directory for this. I recommend [appdirs](https://github.com/ActiveState/appdirs) which helps you get the config path in an OS-agnostic way.
