### 2021-06-24 kasasxav/ImSwitch#44 @untzag
community guidelines
Publication in JOSS requires community guidelines.

>Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

ImSwitch needs to add contribution guidelines. I'd like to see a contributing page in the documentation at the very least. Given the nature of the software, some more specific guidelines on how people might contribute additional hardware support and control features would be welcome.

(This issue must be resolved for me to approve publication in JOSS)
