### 2021-06-24 #3394 @untzag
@kasasxav JOSS requires documentation for contributors / community members. Please provide that prior to publication.

See https://github.com/kasasxav/ImSwitch/issues/44
