### 2021-06-24 #3394 @untzag
Due to the nature of review I've just opened a bunch of issues about problems I see with ImSwitch, which can seem quite negative...

I just want to say for the record that I like this project and I think it's definitely worthy of publication after some polish! I like the focus on modularity. The polish of the GUI is impressive. I could imagine some of the scientists that I work with using this software down the road. Good job authors!
