### 2021-06-24 kasasxav/ImSwitch#45 @untzag
consider documenting the ImSwitch data file format
All I can see from your documentation is that ImSwitch saves to HDF5 and that all the metadata is stored. What exactly does the HDF5 file look like? Is it one big array? One array per channel, or per detector? How does it record parameters scanned against? A brief description in the docs would go a long way to helping me know what to expect.
