### 2021-06-25 a79d2dc
Create contributing.rst
