+++ docs/contributing.rst
@@ -1,0 +1,37 @@
+*****************************
+How to contribute to ImSwitch
+*****************************
+
+We want to encourage users and developers to give active feedback on their experience and feel free to contribute. 
+
+First, we feel very strongly about being inclusive and respectful to other contributors. 
+Please follow the `Python Community guidelines <https://www.python.org/psf/conduct/>`_ if you wish to contribute to our project.
+
+There are different ways you can contribute:
+
+User feedback
+==============
+We would like to hear about your experience when using ImSwitch and suggestions for improvement. You can do that by starting a thread in the `Discussion section <https://github.com/kasasxav/ImSwitch/discussions/>`_ of the repository. 
+
+Improving documentation
+========================
+We would like to include your use-case into the documentation!
+Feel free to open a discussion thread about what you wish to include and offer help with regarding the documentation.
+
+Reporting bugs
+===============
+If you experience a bug, you can directly report it in the `Issues section <https://github.com/kasasxav/ImSwitch/issues/>`_ and add a bug label. Describe how to reproduce the bug and as much information as possible to fix it more efficiently. 
+
+Have you written a code that fixes the bug? You can open a new Pull Request or write your suggestion in Discussions.
+
+Suggest a new feature 
+======================
+If you are missing features and want to develop ImSwitch further, you can start a Discussion and brainstorm your suggestions. Feel free to open a pull request, but we believe it’s essential to have feedback from the community if you want to add new functionality.
+
+Adding device support
+======================
+For adding support to new devices, we encourage you to follow the software architecture diagram provided in the `paper <https://github.com/kasasxav/ImSwitch/blob/master/paper/paper.md>`_ and create the necessary classes (driver, device manager, etc.). You can start a discussion if you want to get advice and help to get started.
+
+Automated tests
+================
+We want to keep including automated tests into the development process, so if you have contributed to the project by fixing a bug or providing new functionality, we encourage you to write code that tests your contribution as well.
