### 2021-06-25 8a7523f
Update contributing.rst
