### 2021-06-25 8f06e51
Update contributing.rst
