### 2021-06-25 48cbf5b
Update contributing.rst
