### 2021-06-25 10c9a78
Update contributing.rst
