### 2021-06-25 f73b5bd
Include HDF5 in index
