### 2021-06-25 8b0f09d
Update contributing.rst
