+++ docs/contributing.rst
@@ -9,24 +9,24 @@ Please follow the `Python Community guidelines <https://www.python.org/psf/condu
 
 There are different ways you can contribute:
 
+Reporting bugs
+===============
+If you encounter a bug, you can directly report it in the `issues section <https://github.com/kasasxav/ImSwitch/issues/>`_. Please describe how to reproduce the bug, and include as much information as possible that can be helpful for fixing it.
+
+Have you written code that fixes the bug? You can open a new pull request or write your suggestion in the issue.
+
 User feedback
 ==============
-We would like to hear about your experience when using ImSwitch and suggestions for improvement. You can do that by starting a thread in the `discussion section <https://github.com/kasasxav/ImSwitch/discussions/>`_ of the repository. 
+We would like to hear about your experience when using ImSwitch and suggestions for improvement. You can do that by starting a thread in the `discussion section <https://github.com/kasasxav/ImSwitch/discussions/>`_ of the repository.
+
+Suggest a new feature 
+======================
+If you are missing features and want to develop ImSwitch further, you can start a discussion and brainstorm your suggestions. Feel free to open a pull request, but we believe it’s essential to have feedback from the community if you want to add new functionality.
 
 Improving documentation
 ========================
 We would like to include your use-case into the documentation!
 Feel free to open a discussion thread about what you wish to include or improve.
-
-Reporting bugs
-===============
-If you experience a bug, you can directly report it in the `issues section <https://github.com/kasasxav/ImSwitch/issues/>`_ and add a bug label. Describe how to reproduce the bug and as much information as possible to fix it more efficiently. 
-
-Have you written a code that fixes the bug? You can open a new pull request or write your suggestion in discussions.
-
-Suggest a new feature 
-======================
-If you are missing features and want to develop ImSwitch further, you can start a discussion and brainstorm your suggestions. Feel free to open a pull request, but we believe it’s essential to have feedback from the community if you want to add new functionality.
 
 Adding device support
 ======================
