### 2021-06-25 kasasxav/ImSwitch#45 @kasasxav
Hi @untzag, thank you for your feedback! We have now included a section in the documentation for explaining the data file and metadata. Please let us know if it is clear enough or if you have further feedback.
