### 2021-06-25 kasasxav/ImSwitch#44 @kasasxav
Thank you for your comment! We have included a contributing page in the documentation and linked it in the readme. Please let us know if you have further feedback regarding this issue.
