+++ setup.cfg
@@ -1,0 +1,28 @@
+[metadata]
+name = ImSwitch-pkg
+version = 0.0.1
+author = Xavier Casas Moreno
+author_email = xaviercm@kth.se
+description = Microscopy control
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/kasasxav/ImSwitch
+project_urls =
+    Bug Tracker = https://github.com/kasasxav/ImSwitch/issues
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+    = imswitch
+packages = find:
+python_requires = >=3.6
+
+[options.packages.find]
+where = imswitch
+
+[options.entry_points]
+console_scripts =
+    imswitch-start = imswitch.__main__:main
