### 2021-06-28 00f749d
Disable laser slider reactions to scroll wheel
