+++ imswitch/imcontrol/view/widgets/LaserWidget.py
@@ -221,7 +221,7 @@ class LaserModule(QtWidgets.QWidget):
         self.maxpower = QtWidgets.QLabel()
         self.maxpower.setAlignment(QtCore.Qt.AlignLeft | QtCore.Qt.AlignVCenter)
 
-        self.slider = QtWidgets.QSlider(QtCore.Qt.Horizontal, self)
+        self.slider = guitools.BetterSlider(QtCore.Qt.Horizontal, self, allowScrollChanges=False)
         self.slider.setFocusPolicy(QtCore.Qt.NoFocus)
 
         if not isBinary:
