### 2021-06-28 e23f7b5
Make requirements less strict
