+++ requirements.txt
@@ -1,18 +1,18 @@
-colour-science~=0.3.15
-dataclasses-json~=0.5.2
-dotmap~=1.3.23
-h5py~=3.1.0
-lantzdev[qt]~=0.5.2
-napari[pyqt5]==0.4.7
-nidaqmx~=0.5.7
-numpy==1.19.3
-psutil~=5.7.3
+colour-science >= 0.3, < 0.4
+dataclasses-json >= 0.5, < 0.6
+dotmap >= 1.3, < 2
+h5py >= 3, < 4
+lantzdev[qt] >= 0.5.2, < 0.7
+napari[pyqt5] == 0.4.7  # pinned because of napari issue 2897
+nidaqmx >= 0.5.7, < 0.6
+numpy >= 1.19, < 2
+psutil >= 5.4.8, < 6
 git+https://github.com/morefigs/py-ic-imaging-control#egg=py-ic-imaging-control; platform_system == "Windows"
-PyQt5~=5.15.2
-pyqtgraph==0.11.0
-pyserial~=3.5
-QDarkStyle~=2.8.1
-QScintilla~=2.12.1
-scikit-image~=0.18.1
-tifffile
-wxPython~=4.1.1
+PyQt5 >= 5.15.2, < 6
+pyqtgraph == 0.11.0  # pinned because we use some internal code that has changed since 0.11.0
+pyserial >= 3.5, < 4
+QDarkStyle >= 2.8, < 4
+QScintilla >= 2.12, < 3
+scikit-image >= 0.18, < 0.19
+tifffile >= 2020.11.26
+wxPython >= 4.1, < 5
