### 2021-06-28 e26ecd1
Add PhotometricsManager for testing
Still incomplete and not tested, but I generated a basic manager adapting HamamatsuManager with the Photometrics SDK.
