+++ imswitch/imcontrol/model/managers/detectors/PhotometricsManager.py
@@ -1,0 +1,195 @@
+from .DetectorManager import (
+    DetectorManager, DetectorNumberParameter, DetectorListParameter
+)
+
+class PhotometricsManager(DetectorManager):
+    """ DetectorManager that deals with the Hamamatsu parameters and frame
+    extraction for a Hamamatsu camera.
+
+    Available manager properties:
+    * cameraListIndex -- the camera's index in the Hamamatsu camera list (list indexing starts at
+                         0); set this to an invalid value, e.g. the string "mock" to load a mocker
+    * hamamatsu -- dictionary of DCAM API properties
+    """
+
+    def __init__(self, detectorInfo, name, **_kwargs):
+        self._camera = getCameraObj(detectorInfo.managerProperties['cameraListIndex'])
+        self._binning = 1
+
+        for propertyName, propertyValue in detectorInfo.managerProperties['hamamatsu'].items():
+            self._camera.setParam( propertyName, propertyValue)
+
+        fullShape = self._camera.sensor_size
+
+        model = self._camera.pvc_get_pvcam_version()
+
+        # Prepare parameters
+        parameters = {
+            'Set exposure time': DetectorNumberParameter(group='Timings', value=0,
+                                                         valueUnits='s', editable=True),
+            'Real exposure time': DetectorNumberParameter(group='Timings', value=0,
+                                                          valueUnits='s', editable=False),
+            'Internal frame interval': DetectorNumberParameter(group='Timings', value=0,
+                                                               valueUnits='s', editable=False),
+            'Readout time': DetectorNumberParameter(group='Timings', value=0,
+                                                    valueUnits='s', editable=False),
+            'Internal frame rate': DetectorNumberParameter(group='Timings', value=0,
+                                                           valueUnits='fps', editable=False),
+            'Trigger source': DetectorListParameter(group='Acquisition mode',
+                                                    value='Internal trigger',
+                                                    options=['Internal trigger',
+                                                             'External "start-trigger"',
+                                                             'External "frame-trigger"'],
+                                                    editable=True),
+            'Camera pixel size': DetectorNumberParameter(group='Miscellaneous', value=0.1,
+                                                         valueUnits='µm', editable=True)
+        }
+
+        super().__init__(detectorInfo, name, fullShape=fullShape, supportedBinnings=[1, 2, 4],
+                         model=model, parameters=parameters, croppable=True)
+        self._updatePropertiesFromCamera()
+        super().setParameter('Set exposure time', self.parameters['Real exposure time'].value)
+
+    @property
+    def pixelSizeUm(self):
+        umxpx = self.parameters['Camera pixel size'].value
+        return [1, umxpx, umxpx]
+
+    def getLatestFrame(self):
+        return self._camera.poll_frame()
+
+    def getChunk(self):
+        return self.getLatestFrame()
+
+    def flushBuffers(self):
+        #self._camera.updateIndices()
+        print("Nothing")
+
+    def crop(self, hpos, vpos, hsize, vsize):
+        """Method to crop the frame read out by the camera. """
+        roi = (hpos, hpos+hsize, vpos, vpos+vsize)
+        self._performSafeCameraAction(
+            lambda: self._camera.set_param('roi', self._camera.set_param('roi', roi))
+        )
+       
+        # This should be the only place where self.frameStart is changed
+        self._frameStart = (hpos, vpos)
+        # Only place self.shapes is changed
+        self._shape = (hsize, vsize)
+
+    def setBinning(self, binning):
[…]
