### 2021-06-28 6ad664e
Adapt stylesheet to work well with qdarkstyle >= 3.0.0
