### 2021-06-28 d1c872a
add json slm info dataclass
add json slm info dataclass, reading from a json slm info file, and initiate other changes needed to do the change from pickle to json serializing
