### 2021-06-28 7c53033
Fix QtCore.Signal being unavailable
Import qtpy instead of PyQt5
