+++ imswitch/imreconstruct/view/ScanParamsDialog.py
@@ -1,4 +1,4 @@
-from pyqtgraph.Qt import QtCore, QtWidgets
+from qtpy import QtCore, QtWidgets
 
 from .guitools import BetterPushButton
 
