+++ imswitch/imcommon/view/guitools/PickModulesDialog.py
@@ -1,4 +1,4 @@
-from pyqtgraph.Qt import QtCore, QtWidgets
+from qtpy import QtCore, QtWidgets
 
 
 class PickModulesDialog(QtWidgets.QDialog):
