+++ imswitch/imreconstruct/view/MultiDataFrame.py
@@ -1,4 +1,4 @@
-from pyqtgraph.Qt import QtCore, QtGui, QtWidgets
+from qtpy import QtCore, QtGui, QtWidgets
 
 from .guitools import BetterPushButton
 
