+++ imswitch/imcommon/view/guitools/dialogtools.py
@@ -1,4 +1,4 @@
-from pyqtgraph.Qt import QtCore, QtWidgets
+from qtpy import QtCore, QtWidgets
 
 
 def askYesNoQuestion(widget, title, question):
