+++ imswitch/imcommon/view/guitools/BetterSlider.py
@@ -1,4 +1,4 @@
-from pyqtgraph.Qt import QtCore, QtWidgets
+from qtpy import QtCore, QtWidgets
 
 
 class BetterSlider(QtWidgets.QSlider):
