### 2021-06-28 e3296a5
Remove OptimizedImageItem
The performance improvements made in pyqtgraph 0.12.2 may eliminate the need for our custom changes.
