+++ imswitch/imcommon/view/guitools/OptimizedImageItem.py
@@ -1,293 +1,0 @@
-"""
-OptimizedImageItem is a version of pyqtgraph's ImageItem that uses threading and only renders the
-part of the image that is visible.
-
-Based on pyqtgraph 0.11.0 code.
-"""
-
-import math
-import numpy as np
-import pyqtgraph as pg
-from qtpy import QtCore, QtGui
-from pyqtgraph import debug, functions as fn, Point
-
-try:
-    from collections.abc import Callable
-except ImportError:
-    # fallback for python < 3.3
-    from collections import Callable
-
-
-class OptimizedImageItem(pg.ImageItem):
-    sigImageReadyForARGB = QtCore.Signal()
-    sigImageDisplayed = QtCore.Signal()
-
-    def __init__(self, *args, **kargs):
-        self.shouldPaint = False
-        self.onlyRenderVisible = True
-        super().__init__(*args, **kargs)
-
-        # Prepare image computation worker
-        self.imageARGBWorker = ImageARGBWorker()
-        self.imageARGBWorker.qimageProduced.connect(self.updateQImage)
-        self.imageARGBThread = QtCore.QThread()
-        self.imageARGBWorker.moveToThread(self.imageARGBThread)
-        self.sigImageReadyForARGB.connect(self.imageARGBWorker.produceARGB)
-        self.imageARGBThread.start()
-
-    def setAutoDownsample(self, ads):
-        self.autoDownsample = ads
-
-    def setOnlyRenderVisible(self, orv, render=True):
-        self.onlyRenderVisible = orv
-        if render:
-            self.render()
-
-    def setImage(self, image=None, autoLevels=None, **kargs):
-        profile = debug.Profiler()
-
-        gotNewData = False
-        if image is None:
-            if self.image is None:
-                return
-        else:
-            gotNewData = True
-            shapeChanged = (self.image is None or image.shape != self.image.shape)
-            image = image.view(np.ndarray)
-            if self.image is None or image.dtype != self.image.dtype:
-                self._effectiveLut = None
-            self.image = image
-            if self.image.shape[0] > 2**15-1 or self.image.shape[1] > 2**15-1:
-                if 'autoDownsample' not in kargs:
-                    kargs['autoDownsample'] = True
-            if shapeChanged:
-                self.prepareGeometryChange()
-                self.informViewBoundsChanged()
-
-        profile()
-
-        if autoLevels is None:
-            if 'levels' in kargs:
-                autoLevels = False
-            else:
-                autoLevels = True
-        if autoLevels:
-            img = self.image
-            while img.size > 2**16:
-                img = img[::2, ::2]
-            mn, mx = np.nanmin(img), np.nanmax(img)
-            # mn and mx can still be NaN if the data is all-NaN
-            if mn == mx or np.isnan(mn) or np.isnan(mx):
-                mn = 0
-                mx = 255
-            kargs['levels'] = [mn,mx]
-
-        profile()
-
-        self.setOpts(update=False, **kargs)
-
-        profile()
-
-        self.render()
-
-        profile()
-
-        if gotNewData:
-            self.sigImageChanged.emit()
-
-    def render(self):
-        # Convert data to QImage for display.
-
-        profile = debug.Profiler()
-        if self.image is None or self.image.size == 0:
-            return
-
-        # Request a lookup table if this image has only one channel
-        if self.image.ndim == 2 or self.image.shape[2] == 1:
-            if isinstance(self.lut, Callable):
-                lut = self.lut(self.image)
-            else:
-                lut = self.lut
-        else:
-            lut = None
-
-        if self.autoDownsample:
-            # reduce dimensions of image based on screen resolution
[…]
