+++ imswitch/imcontrol/view/widgets/FFTWidget.py
@@ -38,7 +38,7 @@ class FFTWidget(Widget):
         self.cwidget = pg.GraphicsLayoutWidget()
         self.vb = self.cwidget.addViewBox(row=1, col=1)
         self.vb.setMouseMode(pg.ViewBox.RectMode)
-        self.img = guitools.OptimizedImageItem(axisOrder='row-major')
+        self.img = pg.ImageItem(axisOrder='row-major')
         self.img.translate(-0.5, -0.5)
         self.vb.addItem(self.img)
         self.vb.setAspectLocked(True)
