### 2021-06-28 #3394 @kasasxav
Hi @untzag, thank you very much for your feedback. Thank you also for the nice words about the project.

We are working on addressing the issues you mentioned. As for [kasasxav/ImSwitch#44](https://github.com/kasasxav/ImSwitch/issues/44) and [kasasxav/ImSwitch#45](https://github.com/kasasxav/ImSwitch/issues/45), we have now included the required documentation.

As for the authors contributions, I prepared write it below:
