### 2021-06-28 #3394 @kasasxav
I understand your concern @untzag, and I will try to clarify it for you. I believe that in our project, there have been two phases:
First, we designed the architecture of the software, applied the MVP into our project, and decided how we wanted to create the Python classes and connections between the different modules of ImSwitch. In this initial step, it was mainly Andreas Bodén and me who discussed it. Some of the ideas that we brought together and have been important in the development of the project are:

- The class CommunicationChannel, which contains Qt signals, allows different presenters to communicate with each other easily.

- The entity `MasterController` controls `Managers` for each driver.

Andreas also developed different classes like the first versions of the `ScanDesigners` (https://github.com/kasasxav/ImSwitch/commit/13af1bd62dafe11a25e6ea9c72393c5eea979bdf) and `NidaqHelper` (https://github.com/kasasxav/ImSwitch/commit/8f1d2ebb12ece49ea5eb3be1d8879e32f60c2781). The Image Reconstruction module was also based on his Python code, but since we re-arranged it using an MVP and Staffan introduced it the code, it is not listed in GitHub. We took it from his private repository.

The second phase is when Staffan and Jonatan joined the team. We have focused on providing functionality and improving the design, so it has been quite productive in developing code and making a stable release. But it has been possible because we had solid ground from the beginning and a clear idea about how the code should be.

Ilaria Testa has provided direction and guidance to the project. In her career, she has been very involved in the super-resolution community and worked with a wide variety of microscopes and control software. She has given feedback regarding what users expect of software and what is needed when building a microscope. For example, we have brainstormed together the idea of annotating the images with parameters from the scan, such as laser powers, region of interest and exposure time. We have also periodically discussed what direction in the development of the code we wanted to take, taking into account future projects and what is needed in the community. For example, we had the idea of writing scripts to control ImSwitch inspired by high-throughput microscopy and long-term recordings (>12 hours).
