+++ imswitch/imcontrol/controller/controllers/SLMController.py
@@ -52,10 +52,7 @@ class SLMController(ImConWidgetController):
             lambda: self.setMask(MaskMode.Split))
 
         # Connect SLMWidget parameter tree updates
-        self.applySlmParam = self._widget.slmParameterTree.p.param('Apply')
-        self.applySlmParam.sigStateChanged.connect(self.applyGeneral)
-        self.applyAberParam = self._widget.aberParameterTree.p.param('Apply')
-        self.applyAberParam.sigStateChanged.connect(self.applyAberrations)
+        self._widget.applyChangesButton.clicked.connect(self.applyParams)
 
     # Button pressed functions
     def moveMask(self, direction):
@@ -68,9 +65,9 @@ class SLMController(ImConWidgetController):
 
     def saveParams(self):
         obj = self._widget.controlPanel.objlensComboBox.currentText()
-        general_paramtree = self._widget.slmParameterTree
-        aber_paramtree = self._widget.aberParameterTree
-        center_coords = self._master.slmManager.getCenters()
+        #general_paramtree = self._widget.slmParameterTree
+        #aber_paramtree = self._widget.aberParameterTree
+        #center_coords = self._master.slmManager.getCenters()
 
         if obj == 'No objective':
             print('You have to choose an objective from the drop down menu.')
@@ -82,21 +79,52 @@ class SLMController(ImConWidgetController):
         else:
             raise ValueError(f'Unsupported objective "{obj}"')
 
-        state_general = general_paramtree.p.saveState()
-        state_aber = aber_paramtree.p.saveState()
-        state_pos = center_coords
-        #TODO: fix this to save as .json instead of pickling
-        with open(os.path.join(self.slmDir, filename), 'wb') as f:
-            pickler = pickle.Pickler(f)
-            pickler.dump(state_general)
-            pickler.dump(state_pos)
-            pickler.dump(state_aber)
-        # print(f'Saved SLM parameters for {obj} objective.')
+        slm_info_dict = self.getInfoDict(self._widget.slmParameterTree.p, self._widget.aberParameterTree.p, self._master.slmManager.getCenters())
+        print(slm_info_dict)
+        slm_info = SLMInfo()
+        slm_info = slm_info.from_dict_impl(slm_info_dict)
+        with open(os.path.join(self.slmDir, filename), 'w') as f:
+            f.write(slm_info.to_json(indent=4))
+        print(f'Saved SLM parameters for {obj} objective.')
+
+    def getInfoDict(self, generalParams, aberParams, centers):
+        state_general = {
+            "radius": generalParams.getValues()["general"][1]["radius"][0],
+            "sigma": generalParams.getValues()["general"][1]["sigma"][0]
+        }
+        state_aber = {
+            "left": {
+                "tilt": aberParams.getValues()["left"][1]["tilt"][0],
+                "tip": aberParams.getValues()["left"][1]["tip"][0],
+                "defocus": aberParams.getValues()["left"][1]["defocus"][0],
+                "spherical": aberParams.getValues()["left"][1]["spherical"][0],
+                "verticalComa": aberParams.getValues()["left"][1]["verticalComa"][0],
+                "horizontalComa": aberParams.getValues()["left"][1]["horizontalComa"][0],
+                "verticalAstigmatism": aberParams.getValues()["left"][1]["verticalAstigmatism"][0],
+                "obliqueAstigmatism": aberParams.getValues()["left"][1]["obliqueAstigmatism"][0]
+            },
+            "right": {
+                "tilt": aberParams.getValues()["right"][1]["tilt"][0],
+                "tip": aberParams.getValues()["right"][1]["tip"][0],
[…]
