### 2021-06-29 1a26d92
Some changes in PhotometricsManager
