### 2021-06-29 d279bbc
Update PhotometricsManager.py
