+++ imswitch/imcontrol/model/managers/detectors/PhotometricsManager.py
@@ -18,13 +18,13 @@ class PhotometricsManager(DetectorManager):
         self._camera = getCameraObj(detectorInfo.managerProperties['cameraListIndex'])
         self._binning = 1
 
-        #for propertyName, propertyValue in detectorInfo.managerProperties['hamamatsu'].items():
-        #    print(propertyName)
-        #    self._camera.set_param( propertyName, propertyValue)
-        #self._camera.exp_time = detectorInfo.managerProperties['hamamatsu']['exp_time']
+        for propertyName, propertyValue in detectorInfo.managerProperties['hamamatsu'].items():
+            self._camera.setParam( propertyName, propertyValue)
+
         fullShape = self._camera.sensor_size
 
-        model = self._camera.name
+        model = self._camera.pvc_get_pvcam_version()
+
         # Prepare parameters
         parameters = {
             'Set exposure time': DetectorNumberParameter(group='Timings', value=0,
@@ -41,8 +41,12 @@ class PhotometricsManager(DetectorManager):
                                                     value='Internal trigger',
                                                     options=['Internal trigger',
                                                              'External "start-trigger"',
-                                                             'External "frame-trigger"'],
-                                                    editable=True),
+                                                             'External "frame-trigger"']),
+            'Readout port': DetectorListParameter(group='ports',
+                                                    value='Sensitivity',
+                                                    options=['Sensitivity',
+                                                             'Speed"',
+                                                             'Dynamic range"'],                                        editable=True),
             'Camera pixel size': DetectorNumberParameter(group='Miscellaneous', value=0.1,
                                                          valueUnits='µm', editable=True)
         }
@@ -70,9 +74,9 @@ class PhotometricsManager(DetectorManager):
     def crop(self, hpos, vpos, hsize, vsize):
         """Method to crop the frame read out by the camera. """
         roi = (hpos, hpos+hsize, vpos, vpos+vsize)
-        def setRoi():
+        def cropAction():
             self._camera.roi = roi
-        self._performSafeCameraAction(setRoi)
+        self._performSafeCameraAction(cropAction)
         # This should be the only place where self.frameStart is changed
         self._frameStart = (hpos, vpos)
         # Only place self.shapes is changed
@@ -83,9 +87,9 @@ class PhotometricsManager(DetectorManager):
 
         binstring = f'{binning}x{binning}'
         coded = binstring.encode('ascii')
-        def setBinning():
+        def binningAction():
             self._camera.binning = binning
-        self._performSafeCameraAction(setBinning)
+        self._performSafeCameraAction(binningAction)
 
     def setParameter(self, name, value):
         super().setParameter(name, value)
@@ -95,7 +99,8 @@ class PhotometricsManager(DetectorManager):
             self._updatePropertiesFromCamera()
         elif name == 'Trigger source':
             self._setTriggerSource(value)
-
+        elif name == 'Readout port':
+            self._setReadoutPort(value)
         return self.parameters
 
     def startAcquisition(self):
@@ -109,28 +114,33 @@ class PhotometricsManager(DetectorManager):
 
     def _setTriggerSource(self, source):
         print("Change trigger source")
-        #if source == 'Internal trigger':
-        #    self._performSafeCameraAction(
-        #        lambda: self._camera.setPropertyValue('trigger_source', 1)
-        #    )
+        def triggerAction(mode):
+            self._camera.exp_mode = mode
+        if source == 'Internal trigger':
[…]
