### 2021-06-29 17f3688
Some more changes for the Photometrics manager and config file
