+++ imswitch/imcontrol/model/managers/detectors/PhotometricsManager.py
@@ -18,35 +18,31 @@ class PhotometricsManager(DetectorManager):
         self._camera = getCameraObj(detectorInfo.managerProperties['cameraListIndex'])
         self._binning = 1
 
-        for propertyName, propertyValue in detectorInfo.managerProperties['hamamatsu'].items():
-            self._camera.setParam( propertyName, propertyValue)
+        #for propertyName, propertyValue in detectorInfo.managerProperties['hamamatsu'].items():
+        #    self._camera.setParam( propertyName, propertyValue)
 
         fullShape = self._camera.sensor_size
 
-        model = self._camera.pvc_get_pvcam_version()
+        model = self._camera.name
 
         # Prepare parameters
         parameters = {
             'Set exposure time': DetectorNumberParameter(group='Timings', value=0,
-                                                         valueUnits='s', editable=True),
+                                                         valueUnits='ms', editable=True),
             'Real exposure time': DetectorNumberParameter(group='Timings', value=0,
-                                                          valueUnits='s', editable=False),
-            'Internal frame interval': DetectorNumberParameter(group='Timings', value=0,
-                                                               valueUnits='s', editable=False),
-            'Readout time': DetectorNumberParameter(group='Timings', value=0,
-                                                    valueUnits='s', editable=False),
-            'Internal frame rate': DetectorNumberParameter(group='Timings', value=0,
-                                                           valueUnits='fps', editable=False),
+                                                          valueUnits='ms', editable=False),
+            'Readout time' : DetectorNumberParameter(group='Timings', value=0,
+                                                          valueUnits='ms', editable=False),
             'Trigger source': DetectorListParameter(group='Acquisition mode',
                                                     value='Internal trigger',
                                                     options=['Internal trigger',
                                                              'External "start-trigger"',
-                                                             'External "frame-trigger"']),
+                                                             'External "frame-trigger"'], editable=True),
             'Readout port': DetectorListParameter(group='ports',
                                                     value='Sensitivity',
                                                     options=['Sensitivity',
                                                              'Speed"',
-                                                             'Dynamic range"'],                                        editable=True),
+                                                             'Dynamic range"'], editable=True),
             'Camera pixel size': DetectorNumberParameter(group='Miscellaneous', value=0.1,
                                                          valueUnits='µm', editable=True)
         }
@@ -107,66 +103,67 @@ class PhotometricsManager(DetectorManager):
         self._camera.start_live()
 
     def stopAcquisition(self):
+        self._camera.abort()
         self._camera.finish()
 
     def _setExposure(self, time):
-        self._camera.exp_time = time
+        self._camera.exp_time = int(time)
 
     def _setTriggerSource(self, source):
         print("Change trigger source")
-        def triggerAction(mode):
-            self._camera.exp_mode = mode
+        def triggerAction():
+            self._camera.exp_mode = trigger_value
         if source == 'Internal trigger':
-            self._performSafeCameraAction(lambda: triggerAction(1792))
+            trigger_value = 1792
+            self._performSafeCameraAction(triggerAction)
 
         elif source == 'External "start-trigger"':
-            self._performSafeCameraAction(lambda: triggerAction(2304))
+            trigger_value = 2304
+            self._performSafeCameraAction(triggerAction)
 
         elif source == 'External "frame-trigger"':
[…]
