### 2021-06-29 df09615
Change package name and version in setup.cfg to ImSwitch 1.1.0
