+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
-name = ImSwitch-pkg
-version = 0.0.1
+name = ImSwitch
+version = 1.1.0
 author = Xavier Casas Moreno
 author_email = xaviercm@kth.se
 description = Microscopy control
