### 2021-06-29 32fc4c7
Change minimum Python version to 3.7 in setup.cfg
