+++ setup.cfg
@@ -18,7 +18,7 @@ classifiers =
 package_dir =
     = imswitch
 packages = find:
-python_requires = >=3.6
+python_requires = >=3.7
 
 [options.packages.find]
 where = imswitch
