### 2021-06-29 826e085
Show error messages in GUI when modules don't load
