### 2021-06-30 94dddff
Restructure data/user files; ImSwitch can now be started after install
Configs, scripts, and other user files are now saved outside the program structure. Data/library/resource files accessed by the software are now located in a _data folder.
