+++ imswitch/imcontrol/controller/controllers/ScanController.py
@@ -6,8 +6,7 @@ from traceback import print_exc
 
 import numpy as np
 
-from imswitch.imcommon import constants
-from imswitch.imcommon.model import APIExport
+from imswitch.imcommon.model import APIExport, dirtools
 from imswitch.imcontrol.view import guitools
 from .basecontrollers import SuperScanController
 
@@ -31,7 +30,7 @@ class ScanController(SuperScanController):
             self._master.scanManager.TTLTimeUnits
         )
 
-        self.scanDir = os.path.join(constants.rootFolderPath, 'save/scans')
+        self.scanDir = os.path.join(dirtools.UserFileDirs.Root, 'imcontrol_scans')
         if not os.path.exists(self.scanDir):
             os.makedirs(self.scanDir)
 
