+++ MANIFEST.in
@@ -1,0 +1,1 @@
+graft imswitch/_data
