+++ setup.cfg
@@ -15,14 +15,30 @@ classifiers =
     Operating System :: OS Independent
 
 [options]
-package_dir =
-    = imswitch
 packages = find:
+include_package_data = True
 python_requires = >=3.7
-
-[options.packages.find]
-where = imswitch
+install_requires =
+    colour-science >= 0.3
+    dataclasses-json >= 0.5
+    dotmap >= 1.3
+    h5py >= 3
+    lantzdev[qt] >= 0.5.2
+    napari[pyqt5] >= 0.4.7
+    nidaqmx >= 0.5.7
+    numpy >= 1.19
+    psutil >= 5.4.8
+    py-ic-imaging-control @ git+https://github.com/morefigs/py-ic-imaging-control#egg=py-ic-imaging-control;platform_system=='Windows'
+    PyQt5 >= 5.15.2
+    pyqtgraph == 0.11.0
+    pyserial >= 3.5
+    QDarkStyle >= 2.8
+    QScintilla >= 2.12
+    qtpy >= 1.9
+    scikit-image >= 0.18
+    tifffile >= 2020.11.26
+    wxPython >= 4.1
 
 [options.entry_points]
 console_scripts =
-    imswitch-start = imswitch.__main__:main
+    imswitch = imswitch.__main__:main
