### 2021-06-30 cddc301
Exclude data folder from flake8 in GitHub actions
