+++ .github/workflows/imswitch-test.yml
@@ -33,10 +33,10 @@ jobs:
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --exclude=scripts/* \
+        flake8 . --count --exclude=imswitch/_data/* \
         --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --count --exclude=scripts/* \
+        flake8 . --count --exclude=imswitch/_data/* \
         --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
       run: |
