### 2021-06-30 54f030d
Add menu bar action for opening user files folder
