### 2021-06-30 e9a9bf8
Complete and tested PhotometricsManager
