### 2021-06-30 fef0570
add finalize() to PhotometricsManager
