### 2021-06-30 4655f66
Run imswitch-test GitHub action on installed package
