### 2021-06-30 509c4d4
Run imswitch-test GitHub action on schedule as well (every day)
