+++ .github/workflows/imswitch-test.yml
@@ -1,6 +1,10 @@
 name: imswitch-test
 
-on: [ push, pull_request ]
+on:
+  push:
+  pull_request:
+  schedule:
+    - cron: '30 7 * * *'
 
 jobs:
   build:
