### 2021-06-30 7ac28f5
Fix all detectors stopping when one detector encounters an error
