### 2021-06-30 7addff9
Make ModuleLoadErrorView a QMainWindow so it can have a menu bar
