### 2021-06-30 2d6d8b0
Try preventing thread blocking in PhotometricsManager
