### 2021-07-01 947dd36
When copying example files, ignore FileExistsError
