### 2021-07-01 64e351c
sigScanStarting before runScan
