### 2021-07-01 e52bffa
sigScanStarting before runScan
