### 2021-07-01 5be5b6c
SLM: serialized slm parameters to json
finalized switching from pickling slm parameters to serializing them and saving in .json. additional fixes of how parameters are treated and changes to the widget interface.
