+++ imswitch/imcontrol/model/managers/SLMManager.py
@@ -3,7 +3,6 @@ import glob
 import enum
 import math
 
-from lantz import Q_
 from PIL import Image
 import numpy as np
 
@@ -12,10 +11,8 @@ from scipy import signal as sg
 class SLMManager:
     def __init__(self, slmInfo, *args, **kwargs):
         super().__init__(*args, **kwargs)
-
         if slmInfo is None:
             return
-
         self.__slm = getSLMObj(slmInfo.monitorIdx)
         self.__slmInfo = slmInfo
         self.__wavelength = self.__slmInfo.wavelength
@@ -23,7 +20,6 @@ class SLMManager:
         self.__angleMount = self.__slmInfo.angleMount
         self.__slmSize = (self.__slmInfo.width, self.__slmInfo.height)
         self.__correctionPatternsDir = self.__slmInfo.correctionPatternsDir
-        #print(self.__slmSize)
         self.__maskLeft = Mask(self.__slmSize[1], int(self.__slmSize[0]/2), self.__wavelength)
         self.__maskRight = Mask(self.__slmSize[1], int(self.__slmSize[0]/2), self.__wavelength)
         self.__masks = [self.__maskLeft, self.__maskRight]
@@ -35,12 +31,15 @@ class SLMManager:
         self.__masksAber = [self.__maskAberLeft, self.__maskAberRight]
         self.__masksTilt = [self.__maskTiltLeft, self.__maskTiltRight]
 
-        self.updateSLMDisplay()
+        self.updateSLMDisplay(maskChange=True, tiltChange=True, aberChange=True)
 
-    def saveState(self, state_general, state_pos, state_aber):
-        self.state_general = state_general
-        self.state_pos = state_pos
-        self.state_aber = state_aber
+    def saveState(self, state_general=None, state_pos=None, state_aber=None):
+        if state_general != None:
+            self.state_general = state_general
+        if state_pos != None:
+            self.state_pos = state_pos
+        if state_aber != None:
+            self.state_aber = state_aber
 
     def initCorrectionMask(self):
         # Add correction mask with correction pattern
@@ -71,23 +70,23 @@ class SLMManager:
         self.__maskAberRight = Mask(self.__slmSize[1], int(self.__slmSize[0]/2), self.__wavelength)
         self.__maskAberRight.setBlack()
 
-    def setMask(self, mask, angle, sigma, maskMode):
+    def setMask(self, mask, maskMode):
         if self.__masks[mask].mask_type == MaskMode.Black and maskMode != MaskMode.Black:
             self.__masksTilt[mask].setTilt(self.__angleMount, self.__pixelsize)
         if maskMode == maskMode.Donut:
             self.__masks[mask].setDonut()
         elif maskMode == maskMode.Tophat:
-            self.__masks[mask].setTophat(sigma)
+            self.__masks[mask].setTophat()
         elif maskMode == maskMode.Half:
-            self.__masks[mask].setHalf(angle)
+            self.__masks[mask].setHalf()
         elif maskMode == maskMode.Gauss:
             self.__masks[mask].setGauss()
         elif maskMode == maskMode.Hex:
-            self.__masks[mask].setHex(angle)
+            self.__masks[mask].setHex()
         elif maskMode == maskMode.Quad:
-            self.__masks[mask].setQuad(angle)
+            self.__masks[mask].setQuad()
         elif maskMode == maskMode.Split:
-            self.__masks[mask].setSplit(angle)
+            self.__masks[mask].setSplit()
         elif maskMode == maskMode.Black:
             self.__masks[mask].setBlack()
[…]
