### 2021-07-02 93934d2
Update documentation to reflect packaging and user file location changes
