+++ docs/modules.rst
@@ -20,6 +20,3 @@ are not enabled will not show up in the program.
 
 .. image:: ./images/module-selection.png
     :align: center
-
-Behind the scenes, the list of enabled modules is defined in the /config/modules.json file, which is
-automatically created the first time the program starts.
