### 2021-07-02 db66f46
Remove old Hamamatsu stuff from PhotometricsManager
