### 2021-07-02 81b65a1
Add PhotometricsManager to docs
