### 2021-07-02 4c3b674
Embed py-ic-imaging-control
Since it's not published on PyPI
