+++ imswitch/imcontrol/model/interfaces/pyicic/IC_Camera.py
@@ -1,0 +1,569 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from __future__ import absolute_import
+from __future__ import division
+from builtins import object
+from ctypes import *
+import time
+
+from .IC_GrabberDLL import IC_GrabberDLL
+from .IC_Exception import IC_Exception
+from .IC_Property import IC_Property
+from . import IC_Structures as structs
+
+
+GrabberHandlePtr = POINTER(structs.GrabberHandle)
+
+# "typedefs"
+IMG_FILETYPE = ['FILETYPE_BMP',
+                'FILETYPE_JPG']
+COLOR_FORMAT = ['Y800',
+                'RGB24',
+                'RGB32',
+                'UYVY',
+                'Y16',
+                'NONE']
+
+# c function type for frame callback
+# outside of class so it can be called by unbound function
+C_FRAME_READY_CALLBACK = CFUNCTYPE(None, GrabberHandlePtr, POINTER(c_ubyte), c_ulong, c_void_p)
+
+class IC_Camera:
+    
+    @property
+    def callback_registered(self):
+        return self._callback_registered
+  
+    def __init__(self, unique_device_name):
+        
+        self._unique_device_name = unique_device_name
+        
+        self._handle = IC_GrabberDLL.create_grabber()
+        if not self._handle:
+            raise IC_Exception(todo)
+        
+        self._callback_registered = False
+        self._frame = {'num'    :   -1,
+                       'ready'  :   False}
+
+    def __getattr__(self, attr):
+    
+        if attr in IC_Property.get_all_property_names():
+            return IC_Property(self._handle, attr)
+        else:
+            raise AttributeError
+    
+    # not needed if we use props directly
+    #def __setattr__(self, attr, val):
+    #    
+    #    if attr.startswith('_'):
+    #        super(IC_Camera, self).__setattr__(attr, val)
+    #
+    #    # if it's an actual device property
+    #    elif attr in self.get_all_property_names():
+    #        IC_Property(self._handle, attr).value = val
+    #
+    #    # otherwise just set the attribute value as normal
+    #    else:
+    #        super(IC_Camera, self).__setattr__(attr, val)
+        
+    def open(self):
+        """
+        Open the camera device, required for most functions.
+        """
+        err = IC_GrabberDLL.open_device_by_unique_name(self._handle,
+                                                       self._unique_device_name.encode('ascii'))
+        if err != 1:
+            raise IC_Exception(err)
+    
+    def close(self):
+        """
+        Close the camera device.
+        """
+        IC_GrabberDLL.close_device(self._handle)
+ 
+ 
+        ## don't use, returns wrong number..?
+        #def get_serial_number(self):
+        #    
+        #    #serial = create_string_buffer(20)
+        #    serial = (c_char * 20)()
+        #    
+        #    IC_GrabberDLL.get_serial_number(self._handle,
+        #                                    serial)
+        #
+        #    return serial.value
+    
+    
+    def is_open(self):
+        """
+        Check if the camera device is currently open.
+        
+        :returns: boolean -- True if camera is open.
+        """
+        return bool(IC_GrabberDLL.is_dev_valid(self._handle))
+    
+    def show_property_dialog(self):
+        """
+        Show property dialog for device.
+        """
+        err = IC_GrabberDLL.show_property_dialog(self._handle)
+        if err != 1:
+            raise IC_Exception(err)
+    
+    def list_property_names(self):
+        return IC_Property.get_all_property_names()
+    
+    # use props instead, e.g. cam.gain.range
+    #def get_property_range(self, property_name):
+    #    return IC_Property(self._handle, property_name).range
+    #
[…]
