+++ setup.cfg
@@ -28,7 +28,6 @@ install_requires =
     nidaqmx >= 0.5.7
     numpy >= 1.19
     psutil >= 5.4.8
-    py-ic-imaging-control @ git+https://github.com/morefigs/py-ic-imaging-control#egg=py-ic-imaging-control;platform_system=='Windows'
     PyQt5 >= 5.15.2
     pyqtgraph == 0.11.0
     pyserial >= 3.5
