### 2021-07-03 2327ca4
addItemToViewer in NapariBaseWidget
