### 2021-07-03 2da3798
Implement ULens using NapariBaseWidget
