### 2021-07-03 90872fb
Trying napari point layer
