### 2021-07-03 0ec3fa5
Add descriptions in example scripts
