### 2021-07-03 e73abc8
Settle for pyqtgraph >= 0.12.1 for now
It brings enough performance improvements to be acceptable.
