+++ setup.cfg
@@ -29,7 +29,7 @@ install_requires =
     numpy >= 1.19
     psutil >= 5.4.8
     PyQt5 >= 5.15.2
-    pyqtgraph == 0.11.0
+    pyqtgraph >= 0.12.1
     pyserial >= 3.5
     QDarkStyle >= 2.8
     QScintilla >= 2.12
