### 2021-07-03 02f3e6f
Get data from selected layer
