### 2021-07-05 c6fb303
Restructure some dialog code
