+++ imswitch/imcontrol/controller/controllers/LaserController.py
@@ -1,7 +1,7 @@
 from imswitch.imcommon.model import APIExport
 from imswitch.imcontrol.model import configfiletools
 from imswitch.imcontrol.view import guitools
-from .basecontrollers import ImConWidgetController
+from ..basecontrollers import ImConWidgetController
 
 
 class LaserController(ImConWidgetController):
