+++ imswitch/imcontrol/view/guitools/__init__.py
@@ -1,3 +1,2 @@
 from imswitch.imcommon.view.guitools import *  # noqa
-from .PickSetupDialog import PickSetupDialog
 from .ViewSetupInfo import ROIInfo, LaserPresetInfo, ScanDefaultsInfo, ViewSetupInfo
