### 2021-07-05 b6c48f7
Rename "examples" directory to "user_defaults"
Since we don't just store examples there.
