### 2021-07-05 a1b1d83
Add "documentation" and "about" items to help menu
