### 2021-07-05 d3b0068
Run automated testing on Windows and macOS as well
