+++ .github/workflows/imswitch-test.yml
@@ -9,9 +9,11 @@ on:
 jobs:
   build:
 
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.os }}
     strategy:
+      fail-fast: false
       matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
         python-version: [3.7, 3.8, 3.9]
 
     steps:
@@ -24,16 +26,19 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
-        sudo apt install -y xvfb libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0 libxcb-keysyms1 \
-            libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xfixes0 build-essential \
-            libgtk-3-dev
+        if [ "$RUNNER_OS" == "Linux" ]; then
+          sudo apt install -y xvfb libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0 libxcb-keysyms1 \
+              libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xfixes0 build-essential \
+              libgtk-3-dev
 
-        pip install -U -f https://extras.wxpython.org/wxPython4/extras/linux/gtk3/ubuntu-18.04 \
-            wxPython
+          pip install -U -f https://extras.wxpython.org/wxPython4/extras/linux/gtk3/ubuntu-18.04 \
+              wxPython
+        fi
 
         python3 -m pip install --upgrade pip
         python3 -m pip install .
         python3 -m pip install flake8 pytest pytest-qt
+      shell: bash
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
@@ -42,6 +47,12 @@ jobs:
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exclude=imswitch/_data/* \
         --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+      shell: bash
     - name: Test with pytest
       run: |
-        xvfb-run --server-args "-screen 0 1920x1080x24" python3 -m pytest --pyargs imswitch -v
+        if [ "$RUNNER_OS" == "Linux" ]; then
+          xvfb-run --server-args "-screen 0 1920x1080x24" python3 -m pytest --pyargs imswitch -v
+        else
+          python3 -m pytest --pyargs imswitch -v
+        fi
+      shell: bash
