### 2021-07-05 00ae883
Add GitHub action for bundling
