+++ .github/workflows/imswitch-bundle.yml
@@ -1,0 +1,46 @@
+name: imswitch-bundle
+
+on:
+  workflow_dispatch:
+  push:
+    tags:
+      - '*'
+
+jobs:
+  build:
+
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [windows-latest]
+
+    steps:
+    - uses: actions/checkout@v2
+      with:
+        ref: ${{ github.ref }}
+    - name: Set up Python 3.9
+      uses: actions/setup-python@v2
+      with:
+        python-version: 3.9
+    - name: Install dependencies
+      run: |
+        if [ "$RUNNER_OS" == "Linux" ]; then
+          sudo apt install -y  build-essential libgtk-3-dev
+
+          pip install -U -f https://extras.wxpython.org/wxPython4/extras/linux/gtk3/ubuntu-18.04 \
+              wxPython
+        fi
+
+        python3 -m pip install --upgrade pip
+        python3 -m pip install .
+        python3 -m pip install pyinstaller
+      shell: bash
+    - name: Build bundle
+      run: |
+        pyinstaller imswitch.spec
+    - name: Upload artifact
+      uses: actions/upload-artifact@v2
+      with:
+        name: imswitch-${{ matrix.os }}
+        path: dist
