### 2021-07-05 e19b6a0
Change save reconstruction shortcut to Ctrl+D to avoid napari conflict
