### 2021-07-06 b01260a
Fix API docs generator assuming certain imcontrol settings
