### 2021-07-06 c400cd8
Recalculate curves in repeat mode
Discussion #37
