### 2021-07-06 82cc445
Separate documentation and testing in readme
