### 2021-07-06 374f321
Fix freezing in recordings in some cases
