### 2021-07-06 030a62b
Make update check dialog wider
