### 2021-07-06 07976a9
Fix missing word wrap in update check dialog
