### 2021-07-06 b38845d
Document more properties in DetectorManager
