### 2021-07-06 152302a
Pin napari in setup.cfg (because of napari issue 2897)
