+++ setup.cfg
@@ -25,7 +25,7 @@ install_requires =
     h5py >= 3
     lantzdev[qt] >= 0.5.2
     luddite >= 1
-    napari[pyqt5] >= 0.4.7
+    napari[pyqt5] == 0.4.7  # pinned because of napari issue 2897
     nidaqmx >= 0.5.7
     numpy >= 1.19
     packaging >= 19
