### 2021-07-06 f89cfce
Fix snaps/recordings not working with non-square images
