### 2021-07-06 0d4930c
Fix incorrect fullShape in HamamatsuManager
