### 2021-07-07 2ba8c96
Adapt install instructions for new bundle/PyPI methods of installing
