### 2021-07-07 25dd6b4
Add changelog to documentation
