+++ docs/changelog.rst
@@ -1,0 +1,9 @@
+*********
+Changelog
+*********
+
+
+1.0.0
+=====
+
+Initial release.
