### 2021-07-07 827edab
Restructure table of contents in documentation
