### 2021-07-07 389fcbd
Add 1.1.0 changelog
