### 2021-07-07 kasasxav/ImSwitch#40 @stafak
Hey @untzag! This is now done -- [pyproject.toml](https://github.com/kasasxav/ImSwitch/blob/master/pyproject.toml) and [setup.cfg](https://github.com/kasasxav/ImSwitch/blob/master/setup.cfg) have been added, and the user files are now placed in a more appropriate location.
