+++ requirements.txt
@@ -1,21 +1,1 @@
-colour-science >= 0.3, < 0.4
-dataclasses-json >= 0.5, < 0.6
-dotmap >= 1.3, < 2
-h5py >= 3, < 4
-lantzdev[qt] >= 0.5.2, < 0.7
-luddite >= 1, < 2
-napari[pyqt5] == 0.4.7  # pinned because of napari issue 2897
-nidaqmx >= 0.5.7, < 0.6
-numpy >= 1.19, < 2
-packaging >= 19
-psutil >= 5.4.8, < 6
-PyQt5 >= 5.15.2, < 6
-pyqtgraph >= 0.12.1, < 0.13
-pyserial >= 3.5, < 4
-QDarkStyle >= 2.8, < 4
-QScintilla >= 2.12, < 3
-qtpy >= 1.9, < 2
-requests >= 2.25, < 3
-scikit-image >= 0.18, < 0.19
-tifffile >= 2020.11.26
-wxPython >= 4.1, < 5
+.
