### 2021-07-08 kasasxav/ImSwitch#40 @untzag
Great, I've installed ImSwitch locally using pip and from PyPI. Very cool. Thanks!
