### 2021-07-08 kasasxav/ImSwitch#44 @untzag
Thanks @kasasxav @stafak, looks good to me.
