### 2021-07-08 #3394 @untzag
> @kasasxav JOSS requires documentation for contributors / community members. Please provide that prior to publication.
>
> See [kasasxav/ImSwitch#44](https://github.com/kasasxav/ImSwitch/issues/44)

The authors have satisfied this requirement to my satisfaction, although other reviewers may of-course disagree.
