### 2021-07-08 #3394 @untzag
All of my check-boxes have been checked.

I haven't been able to test ImSwitch with real hardware, but I've looked at the source code and I'm satisfied that everything seems reasonable. From what I can see, ImSwitch's contribution isn't raw hardware support but rather orchestration and GUI with a hardware abstraction model. It seems flexible enough to take on a variety of camera-based workloads. I think in cases like these where specialty hardware is involved we can take the author's at their word that he software does work on real production experimental systems.

I personally recommend ImSwitch for publication in JOSS. Look forward to reading what other reviewers think :smile:
