### 2021-07-09 b377eae
Remove (non-working) links from scripting API reference
