+++ docs/api/api.imcontrol.rst
@@ -2,157 +2,157 @@
    :format: html
 
 
-class **api.imcontrol**\ (\ `builtins.object <builtins.html#object>`_\ )  
-----------------------------------------------------------------------------
+class **api.imcontrol** (builtins.object)  
+-----------------------------------------------
 
 These functions are available in the api.imcontrol object.  
 
 Methods defined here:  
 
-**getDetectorNames**\ () from
-`imswitch.imcontrol.controller.controllers.SettingsController.SettingsController <imswitch.imcontrol.controller.controllers.SettingsController.html#SettingsController>`_
+**getDetectorNames** () from
+imswitch.imcontrol.controller.controllers.SettingsController.SettingsController
 
 Returns the device names of all detectors. These device names can\ :raw-html-m2r:`<br>`
 be passed to other detector-related functions.
 
-**getLaserNames**\ () from
-`imswitch.imcontrol.controller.controllers.LaserController.LaserController <imswitch.imcontrol.controller.controllers.LaserController.html#LaserController>`_
+**getLaserNames** () from
+imswitch.imcontrol.controller.controllers.LaserController.LaserController
 
 Returns the device names of all lasers. These device names can be\ :raw-html-m2r:`<br>`
 passed to other laser-related functions.
 
-**getPositionerNames**\ () from
-`imswitch.imcontrol.controller.controllers.PositionerController.PositionerController <imswitch.imcontrol.controller.controllers.PositionerController.html#PositionerController>`_
+**getPositionerNames** () from
+imswitch.imcontrol.controller.controllers.PositionerController.PositionerController
 
 Returns the device names of all positioners. These device names can\ :raw-html-m2r:`<br>`
 be passed to other positioner-related functions.
 
-**getPositionerPositions**\ () from
-`imswitch.imcontrol.controller.controllers.PositionerController.PositionerController <imswitch.imcontrol.controller.controllers.PositionerController.html#PositionerController>`_
+**getPositionerPositions** () from
+imswitch.imcontrol.controller.controllers.PositionerController.PositionerController
 
 Returns the positions of all positioners.
 
-**loadScanParamsFromFile**\ (filePath) from
-`imswitch.imcontrol.controller.controllers.ScanController.ScanController <imswitch.imcontrol.controller.controllers.ScanController.html#ScanController>`_
+**loadScanParamsFromFile** (filePath) from
+imswitch.imcontrol.controller.controllers.ScanController.ScanController
 
 Loads scanning parameters from the specified file.
 
-**movePositioner**\ (positionerName, axis, dist) from
-`imswitch.imcontrol.controller.controllers.PositionerController.PositionerController <imswitch.imcontrol.controller.controllers.PositionerController.html#PositionerController>`_
+**movePositioner** (positionerName, axis, dist) from
+imswitch.imcontrol.controller.controllers.PositionerController.PositionerController
 
 Moves the specified positioner axis by the specified number of\ :raw-html-m2r:`<br>`
 micrometers.
 
-**runScan**\ () from
-`imswitch.imcontrol.controller.controllers.ScanController.ScanController <imswitch.imcontrol.controller.controllers.ScanController.html#ScanController>`_
+**runScan** (hasStarted=False) from
+imswitch.imcontrol.controller.controllers.ScanController.ScanController
 
 Runs a scan with the set scanning parameters.
 
-**saveScanParamsToFile**\ (filePath) from
-`imswitch.imcontrol.controller.controllers.ScanController.ScanController <imswitch.imcontrol.controller.controllers.ScanController.html#ScanController>`_
+**saveScanParamsToFile** (filePath) from
[…]
