+++ docs/api/_actions.rst
@@ -2,21 +2,21 @@
    :format: html
 
 
-class **_actions**\ (\ `builtins.object <builtins.html#object>`_\ )  
------------------------------------------------------------------------
+class **_actions** (builtins.object)  
+------------------------------------------
 
 These functions are available at the global level.  
 
 Methods defined here:  
 
-**getWaitForSignal**\ (self, signal, pollIntervalSeconds=1)
+**getWaitForSignal** (self, signal, pollIntervalSeconds=1)
 
 Returns a function that will wait for the specified signal to emit.\ :raw-html-m2r:`<br>`
 The returned function will continuously check whether the signal has\ :raw-html-m2r:`<br>`
 been emitted since its creation. The polling interval defaults to one\ :raw-html-m2r:`<br>`
 second, and can be customized.
 
-**importScript**\ (self, path)
+**importScript** (self, path)
 
 Imports the script at the specified path (either absolute or\ :raw-html-m2r:`<br>`
 relative to the main script) and returns it as a module variable.
