### 2021-07-09 8204354
Fix FFTController accessing view stuff directly
