### 2021-07-09 dc85e99
Rename pyqtgraphtools items to make it more clear they are for pyqtgraph
