### 2021-07-09 47158db
Clarify recording example script only works when setup allows recording
