### 2021-07-11 #3394 @beniroquai
All my boxes are checked too.
Nevertheless, I would really like to implement some of our hardware adapters in this framework since I not only find it visually appealing but like its performance and integration into e.g. Napari. The paper together with the nicely documented instructions give precise details for users.
I would like to give it 1-2 more days until I have an opinion about its ease of letting other integrating adapters :-)

Other than that: I like the project a lot and recommend ImSwitch for a publication in JOSS too!
