### 2021-07-12 a36da00
Fix napari overriding menu bar in macOS
