+++ imswitch/imcommon/view/guitools/__init__.py
@@ -2,6 +2,7 @@ from .BetterPushButton import BetterPushButton
 from .BetterSlider import BetterSlider
 from .colorutils import wavelengthToHex
 from .dialogtools import askYesNoQuestion, askForFilePath, askForFolderPath, askForTextInput
+from .EmbeddedNapari import EmbeddedNapari
 from .imagetools import bestLevels, minmaxLevels
 from .naparitools import (
     addNapariGrayclipColormap, NapariUpdateLevelsWidget, NapariShiftWidget, VispyROIVisual,
