+++ imswitch/imcontrol/model/interfaces/SLM.py
@@ -4,18 +4,19 @@ Created on Sun Dec 06 20:14:02 2015
 Downloaded from:
 https://github.com/wavefrontshaping/slmPy
 
+Converted to PyQt5 for ImSwitch.
+
 @author: Sebastien Popoff
 """
 
-try:
-    import wx
-except ImportError:
-    raise(ImportError,"The wxPython module is required to run this program.")
-import threading
-import numpy as np
 import time
 
-EVT_NEW_IMAGE = wx.PyEventBinder(wx.NewEventType(), 0)
+import numpy as np
+import skimage
+import skimage.transform
+from qtpy import QtCore, QtGui, QtWidgets
+
+from imswitch.imcommon.framework import Mutex
 
 
 class SLMdisplay:
@@ -25,13 +26,15 @@ class SLMdisplay:
         self.monitor = monitor
         # Create the thread in which the window app will run
         # It needs its thread to continuously refresh the window
-        self.vt =  videoThread(self)      
-        self.eventLock = threading.Lock()
+        self.eventLock = Mutex()
         if (self.isImageLock):
-            self.eventLock = threading.Lock()
+            self.eventLock = Mutex()
+
+        self.frame = SLMframe(monitor = self.monitor, isImageLock = self.isImageLock)
+        self.frame.show()
         
     def getSize(self):
-        return self.vt.frame._resX, self.vt.frame._resY
+        return self.frame._resX, self.frame._resY
 
     def updateArray(self, mask):
         """
@@ -51,114 +54,83 @@ class SLMdisplay:
         if len(self.array.shape) == 2:
             bw_array = self.array.copy()
             bw_array.shape = h, w, 1
-            color_array = np.concatenate((bw_array,bw_array,bw_array), axis=2)
-            data = color_array.tostring()
-        else :      
-            data = self.array.tostring()   
-        img = wx.ImageFromBuffer(width=w, height=h, dataBuffer=data)
-        # Create the event
-        event = ImageEvent()
-        event.img = img
-        event.eventLock = self.eventLock
-        
+            img = np.concatenate((bw_array,bw_array,bw_array), axis=2)
+        else:
+            img = self.array
+
         # Wait for the lock to be released (if isImageLock = True)
         # to be sure that the previous image has been displayed
         # before displaying the next one - it avoids skipping images
         if (self.isImageLock):
-            event.eventLock.acquire()
+            self.eventLock.lock()
         # Wait (can bug when closing/opening the SLM too quickly otherwise)
         time.sleep(0.5)
         # Trigger the event (update image)
-        self.vt.frame.AddPendingEvent(event)
+        self.frame.sigNewImage.emit(img, False, None, self.eventLock)
         
     def close(self):
-        self.vt.frame.Close()
+        self.frame.close()
 
 
-class ImageEvent(wx.PyCommandEvent):
-    def __init__(self, eventType=EVT_NEW_IMAGE.evtType[0], id=0):
-        wx.PyCommandEvent.__init__(self, eventType, id)
-        self.img = None
-        self.color = False
-        self.oldImageLock = None
-        self.eventLock = None
+class SLMframe(QtWidgets.QLabel):
+    sigNewImage = QtCore.Signal(object, bool, object, object)  # (img, color, oldImageLock, eventLock)
 
-
-class SLMframe(wx.Frame):
     """Frame used to display full screen image."""
     def __init__(self, monitor, isImageLock = True):   
         self.isImageLock = isImageLock
         # Create the frame
-        #wx.Frame.__init__(self,None,-1,'SLM window',pos = (self._x0, self._y0), size = (self._resX, self._resY)) 
+        super().__init__()
         self.SetMonitor(monitor)
[…]
