### 2021-07-12 64a08e5
Link fix
The references should point to this folder, right? https://github.com/kasasxav/ImSwitch/tree/master/imswitch/_data/user_defaults/imcontrol_setups
