### 2021-07-12 kasasxav/ImSwitch#59 @uellue
GUI seems incompatible with retina display on Windows, makes software unusable
## How to reproduce

* Use most recent release, Windows download .zip on Windows 10 system with high-resolution display (3840x2160, 250 % scale) and follow instructions
* Choose STED example

## What happens

A window without interaction possibility covers the full screen (see screenshot). Other parts of the GUI are behind that window. Keyboard shortcuts have to be used to access the task bar and bring one of the other windows to the foreground.

![2021-07-12 (1)](https://user-images.githubusercontent.com/25689052/125314670-afacab00-e336-11eb-9fd5-0641945e39bc.png)

If a second monitor with 1920x1080 and 100 % scale is connected, the scale of the main  GUI seems off, at least if it is on the low resolution monitor.

## What should happen

A nice window layout and a way to close or move aside windows without resorting to keyboard shortcuts. :-)
