### 2021-07-12 #3394 @uellue
First of all, congratulations! A nice paper and software. A few items before I can check all the boxes:

## Functionality

Unfortunately I ran into issues with the GUI on my system: https://github.com/kasasxav/ImSwitch/issues/59 If possible, I'd abstain from feedback for this item and follow the other reviewers. :-)

## State of the field

* The general term "microscopy" is used. Currently, the software seems to be focused on light-based microscopy. There is no mention of and no support for electron microscopy or scanning probe microscopy, for example. In principle ImSwitch should also be suitable for this, right? Maybe it would be good to mention more prominently the microscopy techniques that ImSwitch could support in principle and the techniques for which there is currently support implemented?
* The issues described in the paper sound very familiar from my own work in transmission electron microscopy. We have chosen Tango as our preferred solution for automation and control across an assortion of devices in complex TEM set-ups. It seems to solve the described issues pretty well and includes GUI components that seem conceptually similar to ImSwitch: https://www.tango-controls.org/developers/gui/. Maybe Tango and EPICS could be included in the "state of the field"?

## Documentation: Automated tests and community guidelines

Here one could use a bit more documentation, possibly better coverage and perhaps automated coverage reports.

* I could only find tests for the `imcontrol` module. Did I miss something, how are the other parts covered? Do you analyze code coverage? I find codecov.io rather helpful. :-)
* The "contributing" guidelines ask for implementing tests, which is good! However, there could be more information on how to run tests locally and where tests should be added.

## Functionality documentation

After reading the documentation with example configurations and the Hardware Control Configuration section I feel like I would need a lot more detailed reference documentation until I could customize ImSwitch for my own use. Since customization is pitched as the main purpose of ImSwitch, this part of the documentation should be extended. In particular, all items and options that I can use in a JSON parameter file should be documented clearly and concisely. Or did I miss this somehow in the documentation?

Furthermore, more instructions on how to implement extensions such as support for new hardware or analysis methods should be included, or it should be easier to find.

## General feedback

This is not relevant for the JOSS review. Reading the functionality documentation, I was wondering if the JSON files to customize ImSwitch for different techniques could lead to the [inner platform effect](https://en.wikipedia.org/wiki/Inner-platform_effect) over time. When including more hardware and supporting more techniques, the syntax will likely become more and more complex. Maybe it could be better to use a Python script instead of a JSON file to specify a microscope set-up? That way one could use an autogenerated API reference from docstrings as well, instead of having a "language break" between JSON and Python. That would make addressing my previous comment easier as well.
