### 2021-07-12 #3394 @uellue
> I don't have easy access to a Windows machine at the moment---perhaps another reviewer could try installing using that route.

@untzag the bundle is confirmed to work for me, minus https://github.com/kasasxav/ImSwitch/issues/59 which I think is unrelated to the installation method.
