### 2021-07-12 #3394 @uellue
@kasasxav reading my review comment again, it may sound like a long list of complaints. :-D Just to make it clear, ImSwitch really looks like a nice project, and thank you for releasing it as Open Source! The raised points are to be understood as suggestions to make it even better and to address points that will likely be important to reach a wider audience so that ImSwitch can achieve its full potential. 👍
