### 2021-07-13 kasasxav/ImSwitch#60 @kasasxav
Add documentation about tests
How to run tests locally and where to add them
