### 2021-07-13 kasasxav/ImSwitch#61 @kasasxav
Documenting config file
Consider doing it automatically based on either comments or decorators
