### 2021-07-13 9a2af16
Prevent freezing when bead reconstruction is active
