### 2021-07-14 4d289ed
Use nearest neighbor scaling for SLM image (as before Qt conversion)
