### 2021-07-14 21b2e2c
Restructure SLM display
