+++ imswitch/imcontrol/view/__init__.py
@@ -1,3 +1,4 @@
 from .guitools import ViewSetupInfo
 from .ImConMainView import ImConMainView
 from .PickSetupDialog import PickSetupDialog
+from .SLMDisplay import SLMDisplay
