### 2021-07-14 04120a7
Add "show" button for SLM display, don't show SLM display on start
