### 2021-07-15 7f24b39
Initial work on auto-generating manager documentation for setups
