+++ imswitch/imcontrol/model/managers/detectors/HamamatsuManager.py
@@ -8,9 +8,11 @@ class HamamatsuManager(DetectorManager):
     extraction for a Hamamatsu camera.
 
     Available manager properties:
-    * cameraListIndex -- the camera's index in the Hamamatsu camera list (list indexing starts at
-                         0); set this to an invalid value, e.g. the string "mock" to load a mocker
-    * hamamatsu -- dictionary of DCAM API properties
+
+    - ``cameraListIndex`` -- the camera's index in the Hamamatsu camera list
+      (list indexing starts at 0); set this to an invalid value, e.g. the
+      string "mock" to load a mocker
+    - ``hamamatsu`` -- dictionary of DCAM API properties to pass to the driver
     """
 
     def __init__(self, detectorInfo, name, **_kwargs):
