+++ imswitch/imcontrol/model/managers/lasers/AAAOTFLaserManager.py
@@ -2,6 +2,17 @@ from .LaserManager import LaserManager
 
 
 class AAAOTFLaserManager(LaserManager):
+    """ LaserManager for controlling one channel of an AA Opto-Electronic
+    acousto-optic modulator/tunable filter through RS232 communication.
+
+    Available manager properties:
+
+    - ``rs232device`` -- name of the defined rs232 communication channel
+      through which the communication should take place
+    - ``channel`` -- index of the channel in the acousto-optic device that
+      should be controlled (indexing starts at 1)
+    """
+
     def __init__(self, laserInfo, name, **kwargs):
         print(type(laserInfo))
         self._channel = int(laserInfo.managerProperties['channel'])
