+++ imswitch/imcontrol/model/managers/detectors/PhotometricsManager.py
@@ -9,8 +9,9 @@ class PhotometricsManager(DetectorManager):
     """ DetectorManager that deals with frame extraction for a Photometrics camera.
 
     Available manager properties:
-    * cameraListIndex -- the camera's index in the Photometrics camera list (list indexing starts at
-                         0)
+
+    - ``cameraListIndex`` -- the camera's index in the Photometrics camera list
+      (list indexing starts at 0)
     """
 
     def __init__(self, detectorInfo, name, **_kwargs):
