+++ imswitch/imcontrol/model/managers/detectors/TISManager.py
@@ -4,6 +4,17 @@ import numpy as np
 
 
 class TISManager(DetectorManager):
+    """ DetectorManager that deals with TheImagingSource cameras and the
+    parameters for frame extraction from them.
+
+    Available manager properties:
+
+    - ``cameraListIndex`` -- the camera's index in the TIS camera list (list
+      indexing starts at 0); set this string to an invalid value, e.g. the
+      string "mock" to load a mocker
+    - ``tis`` -- dictionary of TIS camera properties
+    """
+
     def __init__(self, detectorInfo, name, **_kwargs):
         self._camera = getTISObj(detectorInfo.managerProperties['cameraListIndex'])
 
