### 2021-07-15 fa0d2d3
Fix SLM display not being shown on the correct screen
