### 2021-07-15 a5a0a2a
Allow user to close SLM display by pressing escape while it is in focus
