### 2021-07-16 b44bded
Fix napari icon showing up instead of ImSwitch icon in some cases
