### 2021-07-16 4280f83
Fix SLM display position sometimes being off by 8 pixels on Windows
