### 2021-07-16 7a92e76
Allow user to change SLM monitor in the SLM widget
