+++ imswitch/imcontrol/view/widgets/SLMWidget.py
@@ -11,6 +11,7 @@ class SLMWidget(Widget):
     """ Widget containing slm interface. """
 
     sigSLMDisplayToggled = QtCore.Signal(bool)  # (enabled)
+    sigSLMMonitorChanged = QtCore.Signal(int)  # (monitor)
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -112,10 +113,20 @@ class SLMWidget(Widget):
         abertreeDock.addWidget(self.aberParameterTree)
         self.paramtreeDockArea.addDock(abertreeDock, 'above', pmtreeDock)
         
-        # Button for showing SLM display
+        # Button for showing SLM display and spinbox for monitor selection
+        self.slmDisplayLayout = QtWidgets.QHBoxLayout()
+
         self.slmDisplayButton = guitools.BetterPushButton('Show SLM display (fullscreen)')
         self.slmDisplayButton.setCheckable(True)
         self.slmDisplayButton.toggled.connect(self.sigSLMDisplayToggled)
+        self.slmDisplayLayout.addWidget(self.slmDisplayButton, 1)
+
+        self.slmMonitorLabel = QtWidgets.QLabel('Screen:')
+        self.slmDisplayLayout.addWidget(self.slmMonitorLabel)
+
+        self.slmMonitorBox = QtWidgets.QSpinBox()
+        self.slmMonitorBox.valueChanged.connect(self.sigSLMMonitorChanged)
+        self.slmDisplayLayout.addWidget(self.slmMonitorBox)
 
         # Button to apply changes
         self.applyChangesButton = guitools.BetterPushButton('Apply changes')
@@ -221,13 +232,14 @@ class SLMWidget(Widget):
         self.grid.addWidget(self.slmFrame, 0, 0, 1, 2)
         self.grid.addWidget(self.paramtreeDockArea, 1, 0, 2, 1)
         self.grid.addWidget(self.applyChangesButton, 3, 0, 1, 1)
-        self.grid.addWidget(self.slmDisplayButton, 3, 1, 1, 1)
+        self.grid.addLayout(self.slmDisplayLayout, 3, 1, 1, 1)
         self.grid.addWidget(self.controlPanel, 1, 1, 2, 1)
 
     def initSLMDisplay(self, monitor):
         from imswitch.imcontrol.view import SLMDisplay
         self.slmDisplay = SLMDisplay(self, monitor)
         self.slmDisplay.sigClosed.connect(lambda: self.sigSLMDisplayToggled.emit(False))
+        self.slmMonitorBox.setValue(monitor)
 
     def updateSLMDisplay(self, imgArr):
         self.slmDisplay.updateImage(imgArr)
@@ -235,6 +247,9 @@ class SLMWidget(Widget):
     def setSLMDisplayVisible(self, visible):
         self.slmDisplay.setVisible(visible)
         self.slmDisplayButton.setChecked(visible)
+
+    def setSLMDisplayMonitor(self, monitor):
+        self.slmDisplay.setMonitor(monitor, updateImage=True)
 
 
 # Copyright (C) 2020, 2021 TestaLab
