### 2021-07-16 237d659
Remove old wxPython stuff
