+++ .github/workflows/imswitch-bundle.yml
@@ -25,13 +25,6 @@ jobs:
         python-version: 3.9
     - name: Install dependencies
       run: |
-        if [ "$RUNNER_OS" == "Linux" ]; then
-          sudo apt install -y  build-essential libgtk-3-dev
-
-          pip install -U -f https://extras.wxpython.org/wxPython4/extras/linux/gtk3/ubuntu-18.04 \
-              wxPython
-        fi
-
         python3 -m pip install --upgrade pip
         python3 -m pip install .
         python3 -m pip install pyinstaller
