+++ .github/workflows/imswitch-test.yml
@@ -28,8 +28,7 @@ jobs:
       run: |
         if [ "$RUNNER_OS" == "Linux" ]; then
           sudo apt install -y xvfb libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0 libxcb-keysyms1 \
-              libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xfixes0 build-essential \
-              libgtk-3-dev
+              libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xfixes0
         fi
 
         python3 -m pip install --upgrade pip
