### 2021-07-16 kasasxav/ImSwitch#59 @stafak
Hi! :)

I think the issue with the full screen display should be solved now with the latest commits to master. It's supposed to show up on another screen, but when there only is one monitor connected and turned on it falls back to that monitor, which results in a not very nice situation as you noticed. We've made the following changes to the full screen display:

- It will now only show up after clicking "show SLM display" in the SLM widget
- It is now possible to close it by pressing the escape key while it is in focus
- A warning will show up if it is about to show up in front of the main ImSwitch window (which also explains how to close it)

> If a second monitor with 1920x1080 and 100 % scale is connected, the scale of the main GUI seems off, at least if it is on the low resolution monitor.

Unfortunately, this seems to be a Qt problem that there is no easy solution for. I searched the internet for information on this, and the general recommendation seems to be to let Windows handle the scaling instead of Qt, but this leads to bluriness which we believe would not be acceptable in a microscopy program. If a user wants to let Windows handle the scaling anyway, they can do so by [changing the high DPI scaling override setting](https://www.windowscentral.com/how-change-high-dpi-settings-classic-apps-windows-10-april-2018-update) for ImSwitch.exe.
