### 2021-07-19 b3812e7
Reorganize lantz-related laser things
