+++ imswitch/imcontrol/model/managers/lasers/AAAOTFLaserManager.py
@@ -10,9 +10,7 @@ class AAAOTFLaserManager(LaserManager):
         self.blankingOn()
         self.internalControl()
 
-        super().__init__(
-            laserInfo, name, isBinary=False, isDigital=True, valueUnits='arb'
-        )
+        super().__init__(laserInfo, name, isBinary=False, valueUnits='arb')
 
     def setEnabled(self, enabled):
         """Turn on (1) or off (0) laser emission"""
