### 2021-07-19 9982d87
Add docs for creating device managers
