+++ imswitch/imcontrol/model/managers/detectors/DetectorManager.py
@@ -1,7 +1,7 @@
 import traceback
-from abc import abstractmethod
+from abc import ABC, abstractmethod
 from dataclasses import dataclass
-from typing import Any, Callable, List
+from typing import Any, Dict, List, Optional, Tuple
 
 import numpy as np
 from imswitch.imcommon.framework import Signal, SignalInterface
@@ -9,38 +9,79 @@ from imswitch.imcommon.framework import Signal, SignalInterface
 
 @dataclass
 class DetectorAction:
+    """ An action that is made available for the user to execute. """
+
     group: str
-    func: Callable[[], Any]
+    """ The group to place the action in (does not need to be
+    pre-defined). """
+
+    func: callable
+    """ The function that is called when the action is executed. """
 
 
 @dataclass
-class DetectorParameter:
+class DetectorParameter(ABC):
+    """ Abstract base class for detector parameters that are made available for
+    the user to view/edit. """
+
     group: str
+    """ The group to place the parameter in (does not need to be
+    pre-defined). """
+
     value: Any
+    """ The value of the parameter. """
+
     editable: bool
+    """ Whether it is possible to edit the value of the parameter. """
 
 
 @dataclass
 class DetectorNumberParameter(DetectorParameter):
+    """ A detector parameter with a numerical value. """
+
     value: float
+    """ The value of the parameter. """
+
     valueUnits: str
+    """ Parameter value units, e.g. "nm" or "fps". """
 
 
 @dataclass
 class DetectorListParameter(DetectorParameter):
+    """ A detector parameter with a value from a list of options. """
+
     value: str
+    """ The value of the parameter. """
+
     options: List[str]
+    """ The available values to pick from. """
 
 
 class DetectorManager(SignalInterface):
-    """ Abstract class for a manager for controlling detectors. Intended to be
-    extended for each type of detector. """
+    """ Abstract base class for managers that control detectors. Each type of
+    detector corresponds to a manager derived from this class. """
 
     sigImageUpdated = Signal(np.ndarray, bool)
 
     @abstractmethod
-    def __init__(self, detectorInfo, name, fullShape, supportedBinnings, model, *,
-                 parameters=None, actions=None, croppable=True):
+    def __init__(self, detectorInfo, name: str, fullShape: Tuple[int, int],
+                 supportedBinnings: List[int], model: str, *,
+                 parameters: Optional[Dict[str, DetectorParameter]] = None,
+                 actions: Optional[Dict[str, DetectorAction]] = None,
+                 croppable: bool = True) -> None:
+        """
+        Args:
+            detectorInfo: See setup file documentation.
+            name: The unique name that the device is identified with in the
+              setup file.
+            fullShape: Maximum image size as a tuple ``(width, height)``.
+            supportedBinnings: Supported binnings as a list.
+            model: Detector device model name.
+            parameters: Parameters to make available to the user to view/edit.
+            actions: Actions to make available to the user to execute.
+            croppable: Whether the detector image can be cropped.
+        """
+
         super().__init__()
 
         self._detectorInfo = detectorInfo
@@ -67,6 +108,7 @@ class DetectorManager(SignalInterface):
         self.setBinning(supportedBinnings[0])
 
     def updateLatestFrame(self, init):
+        """ :meta private: """
         try:
             self.__image = self.getLatestFrame()
         except Exception:
@@ -74,11 +116,11 @@ class DetectorManager(SignalInterface):
         else:
             self.sigImageUpdated.emit(self.__image, init)
 
-    def setParameter(self, name, value):
-        """Sets a parameter value and returns the updated list of parameters.
+    def setParameter(self, name: str, value: Any) -> Dict[str, DetectorParameter]:
+        """ Sets a parameter value and returns the updated list of parameters.
[…]
