+++ imswitch/imcontrol/model/managers/detectors/PhotometricsManager.py
@@ -14,7 +14,7 @@ class PhotometricsManager(DetectorManager):
       (list indexing starts at 0)
     """
 
-    def __init__(self, detectorInfo, name, **_kwargs):
+    def __init__(self, detectorInfo, name, **_lowLevelManagers):
         self._camera = getCameraObj(detectorInfo.managerProperties['cameraListIndex'])
         self._binning = 1
 
