+++ imswitch/imcontrol/model/managers/lasers/NidaqLaserManager.py
@@ -7,8 +7,8 @@ class NidaqLaserManager(LaserManager):
     Available manager properties: None
     """
 
-    def __init__(self, laserInfo, name, **kwargs):
-        self._nidaqManager = kwargs['nidaqManager']
+    def __init__(self, laserInfo, name, **lowLevelManagers):
+        self._nidaqManager = lowLevelManagers['nidaqManager']
         super().__init__(laserInfo, name, isBinary=laserInfo.analogChannel is None, valueUnits='V')
 
     def setEnabled(self, enabled):

[…]
