### 2021-07-19 a723db3
Fix missing project name and copyright in Sphinx
