+++ docs/conf.py
@@ -12,6 +12,8 @@ html_theme = 'sphinx_rtd_theme'
 html_static_path = ['_static']
 html_css_files = ['css/custom.css']
 
+project = 'ImSwitch'
+copyright = '2020, 2021 TestaLab'
 
 
 class ClassCondensedHeaderDocumenter(autodoc.ClassDocumenter):
