### 2021-07-19 98e4021
Fix imcontrol API documentation not building
