### 2021-07-19 9b78b0f
Fix missing bullet lists in HDF5 file documentation
