+++ docs/Hdf5datafile.rst
@@ -13,8 +13,8 @@ Datasets
 Each image recording is saved in a dataset named ``data``, with dimensions Z × Y × X, where Z is the number of frames while Y and X are the vertical and horizontal axes respectively.
 Two extra parameters are stored in the dataset:
 
-* ``detector_name``: name of the detector (camera or point-detector) that provided the images.
-* ``element_size_um``: pixel size of the image, this parameter will be automatically read by ImageJ when opening the file.
+- ``detector_name``: name of the detector (camera or point-detector) that provided the images.
+- ``element_size_um``: pixel size of the image, this parameter will be automatically read by ImageJ when opening the file.
 
 
 Object attributes
@@ -28,20 +28,20 @@ For example, for HamamatsuManager: Binning, model, camera pixel size, readout ti
 
 The detector's attributes follow the form:
 
-``Detector:NameDetector:DetectorProperty``
+- ``Detector:NameDetector:DetectorProperty``
 
 Lasers
 -------
 The power and whether it was ON/OFF for each laser is stored in the form:
 
-``Laser:LaserName:Enabled`` (boolean)
-``Laser:LaserName:Value``
+- ``Laser:LaserName:Enabled`` (boolean)
+- ``Laser:LaserName:Value``
 
 Positioners
 ------------
 The value for each positioner is stored to encode in which area the image was taken:
 
-``Positioner:PostionerName:PostionerAxis:Position``
+- ``Positioner:PostionerName:PostionerAxis:Position``
 
 Recording and scanning
 ------------------------
@@ -49,6 +49,6 @@ The parameters of the recording and scanning are attributes as well. They includ
 and ScanWidget, regarding the pulse scheme, the stage positions and step sizes, the type of recording, number of frames, etc.
 It can vary depending on the setup used, but they generally follow the form shown below:
 
-``Rec:PropertyName``
-``ScanStage:PropertyName``
-``ScanTTL:PropertyName``
+- ``Rec:PropertyName``
+- ``ScanStage:PropertyName``
+- ``ScanTTL:PropertyName``
