### 2021-07-19 45b1ec8
Use appropriate sphinx style documentation for scripting API reference
