### 2021-07-19 e10d530
Update contributing.rst
