+++ docs/contributing.rst
@@ -9,29 +9,48 @@ Please follow the `Python Community guidelines <https://www.python.org/psf/condu
 
 There are different ways you can contribute:
 
+
 Reporting bugs
 ===============
-If you encounter a bug, you can directly report it in the `issues section <https://github.com/kasasxav/ImSwitch/issues/>`_. Please describe how to reproduce the bug, and include as much information as possible that can be helpful for fixing it.
+If you encounter a bug,
+you can directly report it in the `issues section <https://github.com/kasasxav/ImSwitch/issues/>`_.
+Please describe how to reproduce the bug
+and include as much information as possible that can be helpful for fixing it.
 
-Have you written code that fixes the bug? You can open a new pull request or write your suggestion in the issue.
+Have you written code that fixes the bug?
+You can open a new pull request or include your suggested fix in the issue.
+
 
 User feedback
 ==============
-We would like to hear about your experience when using ImSwitch and suggestions for improvement. You can do that by starting a thread in the `discussion section <https://github.com/kasasxav/ImSwitch/discussions/>`_ of the repository.
+We would like to hear about your experience when using ImSwitch and suggestions for improvement.
+You can do that by starting a thread in the `discussion section <https://github.com/kasasxav/ImSwitch/discussions/>`_ on GitHub.
+
 
 Suggest a new feature 
 ======================
-If you are missing features and want to develop ImSwitch further, you can start a discussion and brainstorm your suggestions. Feel free to open a pull request, but we believe it’s essential to have feedback from the community if you want to add new functionality.
+If you are missing features and want to develop ImSwitch further,
+you can start a discussion and brainstorm your suggestions.
+Feel free to open a pull request,
+but we believe it’s essential to have feedback from the community if you want to add new functionality.
+
 
 Improving documentation
 ========================
 We would like to include your use-case into the documentation!
 Feel free to open a discussion thread about what you wish to include or improve.
 
+
 Adding device support
 ======================
-For adding support to new devices, we encourage you to follow the software architecture diagram provided in the `paper <https://github.com/kasasxav/ImSwitch/blob/master/paper/paper.md>`_ and create the necessary classes (driver, device manager, etc.). You can start a discussion if you want to get advice and help to get started.
+For adding support to new devices,
+we encourage you to follow the software architecture diagram provided in the `paper <https://github.com/kasasxav/ImSwitch/blob/master/paper/paper.md>`_ and create the necessary classes
+(driver, device manager, etc.).
+You can start a discussion if you want to get advice and help to get started.
+
 
 Automated tests
 ================
-We want to keep including automated tests into the development process, so if you have contributed to the project by fixing a bug or providing new functionality, we encourage you to write code that tests your contribution as well.
+We want to keep including automated tests into the development process,
+so if you have contributed to the project by fixing a bug or providing new functionality,
+we encourage you to write code that tests your contribution as well.
