### 2021-07-19 17167c3
Document where tests are located and where to run them
