### 2021-07-19 8c6c49d
Better support for decimals in laser values
