+++ imswitch/imcontrol/view/widgets/LaserWidget.py
@@ -74,12 +74,14 @@ class LaserWidget(Widget):
 
         self.layout.addLayout(self.presetsBox, 1, 0)
 
-    def addLaser(self, laserName, valueUnits, wavelength, valueRange=None, valueRangeStep=1):
+    def addLaser(self, laserName, valueUnits, valueDecimals, wavelength, valueRange=None,
+                 valueRangeStep=1):
         """ Adds a laser module widget. valueRange is either a tuple
         (min, max), or None (if the laser can only be turned on/off). """
 
         control = LaserModule(
-            units=valueUnits, valueRange=valueRange, tickInterval=5, singleStep=valueRangeStep,
+            valueUnits=valueUnits, valueDecimals=valueDecimals, valueRange=valueRange,
+            tickInterval=5, singleStep=valueRangeStep,
             initialPower=valueRange[0] if valueRange is not None else 0
         )
         control.sigEnableChanged.connect(
@@ -199,20 +201,23 @@ class LaserModule(QtWidgets.QWidget):
     sigEnableChanged = QtCore.Signal(bool)  # (enabled)
     sigValueChanged = QtCore.Signal(float)  # (value)
 
-    def __init__(self, units, valueRange, tickInterval, singleStep, initialPower, *args, **kwargs):
+    def __init__(self, valueUnits, valueDecimals, valueRange, tickInterval, singleStep,
+                 initialPower, *args, **kwargs):
         super().__init__(*args, **kwargs)
+        self.valueDecimals = valueDecimals
+
         isBinary = valueRange is None
 
         # Graphical elements
         self.setPointLabel = QtWidgets.QLabel('Setpoint')
         self.setPointLabel.setAlignment(QtCore.Qt.AlignCenter)
-        self.setPointEdit = QtWidgets.QLineEdit(str(float(initialPower)))
+        self.setPointEdit = QtWidgets.QLineEdit(str(initialPower))
         self.setPointEdit.setFixedWidth(50)
         self.setPointEdit.setAlignment(QtCore.Qt.AlignRight | QtCore.Qt.AlignVCenter)
 
         self.powerLabel = QtWidgets.QLabel('Power')
         self.powerLabel.setAlignment(QtCore.Qt.AlignCenter)
-        self.powerIndicator = QtWidgets.QLabel(str(float(initialPower)))
+        self.powerIndicator = QtWidgets.QLabel(str(initialPower))
         self.powerIndicator.setFixedWidth(50)
         self.powerIndicator.setAlignment(QtCore.Qt.AlignRight | QtCore.Qt.AlignVCenter)
 
@@ -221,7 +226,12 @@ class LaserModule(QtWidgets.QWidget):
         self.maxpower = QtWidgets.QLabel()
         self.maxpower.setAlignment(QtCore.Qt.AlignLeft | QtCore.Qt.AlignVCenter)
 
-        self.slider = guitools.BetterSlider(QtCore.Qt.Horizontal, self, allowScrollChanges=False)
+        if valueDecimals > 0:
+            self.slider = guitools.FloatSlider(QtCore.Qt.Horizontal, self, allowScrollChanges=False,
+                                               decimals=valueDecimals)
+        else:
+            self.slider = guitools.BetterSlider(QtCore.Qt.Horizontal, self,
+                                                allowScrollChanges=False)
         self.slider.setFocusPolicy(QtCore.Qt.NoFocus)
 
         if not isBinary:
@@ -243,10 +253,10 @@ class LaserModule(QtWidgets.QWidget):
 
         self.powerGrid.addWidget(self.setPointLabel, 0, 0, 1, 2)
[…]
