### 2021-07-19 6b3c732
Close RS232 ports on exit
