### 2021-07-19 4c09e8f
Add link to new manager docs from contributing page
