### 2021-07-19 2f7c368
Add "How device managers are implemented" header to device support docs
