### 2021-07-19 74aac1d
Link to paper for multi-manager explanation from device support docs
