### 2021-07-19 9cef07d
Explain that tests are run automatically in GitHub
