+++ paper/paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 The constant development of novel microscopy methods with an increased number of dedicated hardware devices poses significant challenges to software development. On the one hand, software should control complex instruments, provide flexibility to adapt between different microscope modalities, and be open and resilient to modification and extension by users and developers. On the other hand, the community needs software that can satisfy the requirements of the users, such as a user-friendly interface and robustness of the code. In this context, we present ImSwitch, based on the model-view-presenter (MVP) design pattern [@mvp], with an architecture that uses polymorphism to provide a generalized solution to microscope control. Consequently, ImSwitch makes it possible to adapt between different modalities and aims at satisfying the needs of both users and developers. We have also included a scripting module for microscope automation applications and a structure to efficiently share data between different modules, such as hardware control and image processing.
 
 # Statement of need
-As described in @janelia, there is a need to have a generalized solution for microscope control, since the usual approach in microscopy labs is to create their own custom-built software. A complete list is available in @zenodo_janelia. ImSwitch is designed to be compatible with many different microscope modalities and customizable to the specific design of individual custom-built microscopes, all while using the same code. We would like to involve the community in further developing ImSwitch in this direction, believing that it is possible to integrate current state-of-the-art solutions into one unified software.
+As described in @janelia, there is a need to have a generalized solution for microscope control, since the usual approach in microscopy labs is to create their own custom-built software. A complete list is available in @zenodo_janelia. ImSwitch is designed to be compatible with many different microscope modalities and customizable to the specific design of individual custom-built microscopes, all while using the same code. We would like to involve the community in further developing ImSwitch in this direction, believing that it is possible to integrate current state-of-the-art solutions into one unified software. Currently, ImSwitch provides support for light microscopy techniques but could be extended to other instruments requiring multiple hardware synchronization.
 
 # Architecture and implementation
 ImSwitch consists of software modules, e.g., hardware control and image processing, that exchange data through the ModuleCommunicationChannel (\autoref{fig:diagram}). Modules can also expose an internal Application Programming Interface (API) to the scripting platform, allowing user-defined scripts to interact with them without requiring the user to modify the code itself to meet their demands. Each of the modules follows the MVP architecture, which has three primary layers: model, view, and presenter. The hardware control module has the following structure:
@@ -56,6 +56,8 @@ As a proof of concept ImSwitch has been implemented to control the various micro
 µManager [@uManager] is a free and open-source software for microscope control written in Java and C++, with pervasive driver support to different devices and widely used by microscopists. However, µManager does not adapt fully to complex microscopes that require scanning and triggering or other intricate operations. Moreover, software modification and extension are difficult due to the programming languages and the architecture. [Python bindings](https://pypi.org/project/pymmcore/) were recently made available for [MMCore](https://valelab4.ucsf.edu/~MM/doc/MMCore/html/class_c_m_m_core.html), which is the device control layer of µManager, making it compatible with ImSwitch implementation (through the model device sub-layer). Pycro-Manager [@pycroManager] provides access to µManager from Python and combines it with image processing and multiple applications. However, custom plugins must still be written for the µManager Java layer, making it complicated to use in complex devices without a pre-existing specialized implementation. This limits the user-friendliness for microscope users. 
 
 Microscope [@microscope] is a recent alternative for providing a resilient device layer in Python for controlling microscope devices. It is possible to script microscope experiments and use Cockpit [@cockpit] as a GUI. However, the GUI is still limited to simpler microscopes. The most substantial focus of Microscope is to provide a standard interface to control microscope device types. Therefore, it could be integrated with ImSwitch to work towards software generalization by increasing compatibility and including even more devices.
[…]
