### 2021-07-20 2c7e873
Changes in paper
