+++ paper/paper.bib
@@ -131,3 +131,22 @@ year = {2021}
   author={M. Potel},
   year={1996}
 }
+@article{epics,
+title = {EPICS architecture},
+author = {Dalesio, L R and Kozubal, A J and Kraimer, M R},
+doi = {},
+url = {https://www.osti.gov/biblio/6110347}, journal = {},
+number = ,
+volume = ,
+place = {United States},
+year = {1991},
+month = {1}
+}
+@misc{tango,
+  author    = {{Tango Controls}},
+  title     = {Tango Controls},
+    year = {2015},
+    url = {https://www.tango-controls.org},
+    urldate = {2021-07-20}
+}
+
