### 2021-07-20 9800920
Add issue template for bug reports
