+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,31 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: bug
+assignees: ''
+
+---
+
+**Bug description**
+A clear and concise description of what the bug is.
+
+**How to reproduce**
+Steps to reproduce the behavior:
+1.
+2.
+3.
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**Environment**
+ - OS: [e.g. "Windows 10"]
+ - ImSwitch version: [e.g. "1.1.0", or a branch name or commit hash for development versions]
+ - How you downloaded ImSwitch: [either "pip" or "standalone bundle"]
+
+**Additional context**
+Add any other context about the problem here, if applicable.
