+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,6 +1,6 @@
 ---
 name: Bug report
-about: Create a report to help us improve
+about: Submit a report to help us improve ImSwitch
 title: ''
 labels: bug
 assignees: ''
