### 2021-07-21 f5f25e0
Improve hardware configuration file documentation
