+++ imswitch/imcontrol/model/managers/positioners/MHXYStageManager.py
@@ -5,7 +5,7 @@ class MHXYStageManager(PositionerManager):
     """ PositionerManager for control of a Marzhauser XY-stage through RS232
     communication.
 
-    Available manager properties:
+    Manager properties:
 
     - ``rs232device`` -- name of the defined rs232 communication channel
       through which the communication should take place
