+++ docs/conf.py
@@ -6,7 +6,7 @@ from sphinx.ext import autodoc
 
 sys.path.insert(0, os.path.abspath('..'))
 
-extensions = ['sphinx.ext.autodoc', 'sphinx.ext.napoleon']
+extensions = ['sphinx.ext.autodoc', 'sphinx.ext.napoleon', 'sphinx.ext.autosectionlabel']
 
 html_theme = 'sphinx_rtd_theme'
 html_static_path = ['_static']
