+++ imswitch/imcontrol/model/managers/lasers/AAAOTFLaserManager.py
@@ -5,7 +5,7 @@ class AAAOTFLaserManager(LaserManager):
     """ LaserManager for controlling one channel of an AA Opto-Electronic
     acousto-optic modulator/tunable filter through RS232 communication.
 
-    Available manager properties:
+    Manager properties:
 
     - ``rs232device`` -- name of the defined rs232 communication channel
       through which the communication should take place
