### 2021-07-21 0693e42
Merge low-level-managers documentation into adding-device-support
