+++ docs/index.rst
@@ -35,16 +35,10 @@ well as for developers (scripting and API modules, and hardware control and JSON
     :hidden:
     :caption: Internals and specifications
 
-    Hdf5datafile
     modules
+    Hdf5datafile
     imcontrol-setups
-
-.. toctree::
-    :hidden:
-    :caption: Extending ImSwitch
-
     adding-device-support
-    low-level-managers
 
 .. toctree::
     :glob:
