+++ docs/low-level-managers.rst
@@ -1,20 +1,0 @@
-****************************
-Available low-level managers
-****************************
-
-
-lowLevelManagers['nidaqManager']
-================================
-
-.. autoclass:: imswitch.imcontrol.model.managers.NidaqManager.NidaqManager
-   :members: setAnalog, setDigital
-
-
-lowLevelManagers['rs232sManager']
-=================================
-
-.. autoclass:: imswitch.imcontrol.model.managers.RS232sManager.RS232sManager
-   :members:
-
-.. autoclass:: imswitch.imcontrol.model.managers.rs232.RS232Manager.RS232Manager
-   :members:
