### 2021-07-21 c3d886f
Normalize capitalization in docs headers
