### 2021-07-21 kasasxav/ImSwitch#60 @stafak
Done in 17167c3ddf44a53e9b834cbebd84657c599aa7ed.
