### 2021-07-21 kasasxav/ImSwitch#61 @stafak
Done in e6ccd81a571e998e11c23ed94a7cc68e0daef111.
