### 2021-07-21 b3fa6f4
Remove unused power column from laser widget
