### 2021-07-21 0f8942a
Change default recording path to ImSwitch user folder
