+++ paper/paper.md
@@ -57,7 +57,7 @@ As a proof of concept ImSwitch has been implemented to control the various micro
 
 Microscope [@microscope] is a recent alternative for providing a resilient device layer in Python for controlling microscope devices. It is possible to script microscope experiments and use Cockpit [@cockpit] as a GUI. However, the GUI is still limited to simpler microscopes. The most substantial focus of Microscope is to provide a standard interface to control microscope device types. Therefore, it could be integrated with ImSwitch to work towards software generalization by increasing compatibility and including even more devices.
 
-TANGO Controls [@tango] is a software solution for distributed control systems that provides a communication protocol and an object-oriented architecture for controlling devices of a range of applications. Similar to Tango, EPICS [@epics] focuses on supporting applications that operate complex devices such as particle accelerators and telescopes. While their design and architecture are conceptually similar to ImSwitch, our focus resides on microscopy applications and, in particular, image-based control systems and related hardware when performing experiments and handling data (image acquisition, reconstruction, and analysis). We also provide general tools for microscope users and builders of custom-made microscopes.
+TANGO Controls [@tango] is a software solution for distributed control systems that provides a communication protocol and an object-oriented architecture for controlling devices of a range of applications. Similar to TANGO Controls, EPICS [@epics] focuses on supporting applications that operate complex devices such as particle accelerators and telescopes. While their design and architecture are conceptually similar to ImSwitch, our focus resides on microscopy applications and, in particular, image-based control systems and related hardware when performing experiments and handling data (image acquisition, reconstruction, and analysis). We also provide general tools for microscope users and builders of custom-made microscopes.
 
 # Acknowledgements
 We would like to acknowledge the authors of [Tempesta](https://github.com/TestaLab/Tempesta), a further development of Tormenta [@tormenta], for their contributions to earlier software and inspiration for a user-friendly interface. We are also thankful to Francesca Pennacchietti for testing and providing feedback.
