### 2021-07-21 #3394 @kasasxav
Hi @uelle, thank you for your feedback!

## Functionality
The issue [kasasxav/ImSwitch#59](https://github.com/kasasxav/ImSwitch/issues/59) is now addressed, thank you for bringing it up.

## State of the field
We totally agree with your points, ImSwitch aims at providing a solution for microscopy in general but in practice we support specifically light-based microscopy. I have added a sentence in the Summary of the paper now:

"Currently, ImSwitch provides support for light microscopy techniques but could be extended to other microscopy modalities requiring multiple hardware synchronization."

And we have also cited tango controls and EPICS:

“TANGO Controls [@tango] is a software solution for distributed control systems that provides a communication protocol and an object-oriented architecture for controlling devices of a range of applications. Similar to TANGO Controls, EPICS [@epics] focuses on supporting applications that operate complex devices such as particle accelerators and telescopes. While their design and architecture are conceptually similar to ImSwitch, our focus resides on microscopy applications and, in particular, image-based control systems and related hardware when performing experiments and handling data (image acquisition, reconstruction, and analysis). We also provide general tools for microscope users and builders of custom-made microscopes.”

## Documentation: automated tests and community guidelines
We have added to the documentation how to run tests and where they should be added [kasasxav/ImSwitch#60](https://github.com/kasasxav/ImSwitch/issues/60), the config file and the different fields [kasasxav/ImSwitch#61](https://github.com/kasasxav/ImSwitch/issues/61), and how to extend ImSwitch [kasasxav/ImSwitch#63](https://github.com/kasasxav/ImSwitch/issues/63). I think it looks much more complete now and hopefully clear for both developers that want to extend ImSwitch and users. Thank you again for the feedback.

ImSwitch is a software that aims at facilitating hardware control in microscopy applications. For this reason, in this project we have focused mainly in the `imcontrol` module, and covered it with automatic tests to make sure that the core functionality works efficiently. The other modules `imreconstruct` and `imscripting` have not been covered for the following reasons:

-	The `imscripting` module calls functions directly exported from the `imcontrol` module (covered already with tests). We offer this platform so that users and developers can write, test and run their own scripts. If there is an interest and use cases we might increase coverage for those in the future, but we believe it's not parf of the scope of this work.

-	`imreconstruct` was mainly developed for the specific need of our lab. Therefore, we have implemented our reconstruction methods for the techniques we use and suggested that other labs can implement their own modules. The idea behind showing `imreconstruct` is to explain that there is a possibility to dynamically share data and allow for other applications to be developed. We will keep working in the future to provide a more generalized module.

## General feedback
Thank you for your suggestion. We use dataclasses that directly fetch the configuration parameters from the JSON files. We have autogenerated a documentation for the JSON files based on docstrings.

We will keep in mind your other suggestions for the future.
