### 2021-07-21 #3394 @kasasxav
You can check the latest documentation at https://imswitch.readthedocs.io/en/latest/
