### 2021-07-22 c8daadc
Remove unneeded slider if/else
