### 2021-07-22 2faab29
Create the directory before saving snap or opening recording folder
