### 2021-07-22 b6f7316
Create shortcuts: decorator and functionality
