+++ imswitch/imcommon/model/__init__.py
@@ -1,3 +1,4 @@
 from .api import APIExport, generateAPI
+from .shortcut import shortcut, generateShortcuts
 from .SharedAttributes import SharedAttributes
 from .VFileCollection import DataItem, VFileCollection
