### 2021-07-22 6d85436
Add shortcuts
