+++ imswitch/imcontrol/view/widgets/RecordingWidget.py
@@ -6,6 +6,8 @@ from qtpy import QtCore, QtWidgets
 from imswitch.imcontrol.view import guitools as guitools
 from .basewidgets import Widget
 
+from imswitch.imcommon.model.shortcut import shortcut
+
 
 class RecordingWidget(Widget):
     """ Widget to control image or sequence recording.
@@ -273,6 +275,9 @@ class RecordingWidget(Widget):
     def updateRecLapseNum(self, lapseNum):
         self.currentLapse.setText(str(lapseNum) + ' /')
         
+    @shortcut('Shift+r', "Record")
+    def toggleRecButton(self):
+        self.recButton.toggle()
 
 # Copyright (C) 2020, 2021 TestaLab
 # This file is part of ImSwitch.
