+++ imswitch/imcommon/view/guitools/naparitools.py
@@ -52,6 +52,7 @@ class NapariBaseWidget(QtWidgets.QWidget):
             napariViewer.window.qt_viewer.dockLayerList
         )
         napariViewer.window.qt_viewer.dockLayerList.show()
+        return widget
 
 
 class NapariUpdateLevelsWidget(NapariBaseWidget):
@@ -81,6 +82,8 @@ class NapariUpdateLevelsWidget(NapariBaseWidget):
         for layer in self.viewer.layers.selected:
             layer.contrast_limits = minmaxLevels(layer.data)
 
+    def toggleButton(self):
+        self._on_update_levels()
 
 class NapariShiftWidget(NapariBaseWidget):
     """ Napari widget for shifting the currently selected layer by a
