+++ imswitch/imcontrol/view/widgets/ViewWidget.py
@@ -1,3 +1,4 @@
+from imswitch.imcommon.model.shortcut import shortcut
 from qtpy import QtCore, QtWidgets
 
 from imswitch.imcontrol.view import guitools as guitools
@@ -65,6 +66,17 @@ class ViewWidget(Widget):
         """ Sets whether the LiveView crosshair is visible. """
         self.gridButton.setChecked(visible)
 
+    @shortcut('Shift+l', "Liveview")
+    def toggleLiveviewButton(self):
+        self.liveviewButton.toggle()
+
+    @shortcut('Shift+g', "Grid")
+    def toggleGridButton(self):
+        self.gridButton.toggle()
+    
+    @shortcut('Shift+c', "Crosshair")
+    def toggleCrosshairButton(self):
+        self.crosshairButton.toggle()
 
 # Copyright (C) 2020, 2021 TestaLab
 # This file is part of ImSwitch.
