### 2021-07-22 41b573f
Add next detector shortcut
