+++ imswitch/imcontrol/view/widgets/SettingsWidget.py
@@ -4,6 +4,7 @@ from pyqtgraph.parametertree import ParameterTree, Parameter
 from imswitch.imcontrol.view import guitools as guitools
 from .basewidgets import Widget
 
+from imswitch.imcommon.model import shortcut
 
 class CamParamTree(ParameterTree):
     """ Making the ParameterTree for configuration of the detector during imaging
@@ -220,6 +221,9 @@ class SettingsWidget(Widget):
     def hideROI(self):
         self.ROI.hide()
         
+    @shortcut("Shift+n", "Next detector")
+    def toggleNextButton(self):
+        self.nextDetectorButton.toggle()
 
 # Copyright (C) 2020, 2021 TestaLab
 # This file is part of ImSwitch.
