### 2021-07-22 6f5d1bf
Simplify updateLevels shortcut
