### 2021-07-22 57710c9
Shortcuts for laser presets
