### 2021-07-23 6ea62ea
NextDetector click instead of toggle
