### 2021-07-23 cbaed36
Add typing to API documentation
