+++ imswitch/imcontrol/controller/controllers/RecordingController.py
@@ -1,5 +1,6 @@
 import os
 import time
+from typing import Optional, Union
 
 from imswitch.imcommon.framework import Timer
 from imswitch.imcommon.model import filetools, ostools, APIExport
@@ -280,60 +281,60 @@ class RecordingController(ImConWidgetController):
                 self.setSharedAttr(_freqAttr, self._widget.getTimelapseFreq())
 
     @APIExport
-    def snapImage(self):
+    def snapImage(self) -> None:
         """ Take a snap and save it to a .tiff file at the set file path. """
         self.snap()
 
     @APIExport
-    def startRecording(self):
+    def startRecording(self) -> None:
         """ Starts recording with the set settings to the set file path. """
         self._widget.setRecButtonChecked(True)
 
     @APIExport
-    def stopRecording(self):
+    def stopRecording(self) -> None:
         """ Stops recording. """
         self._widget.setRecButtonChecked(False)
 
     @APIExport
-    def setRecModeSpecFrames(self, numFrames):
+    def setRecModeSpecFrames(self, numFrames: int) -> None:
         """ Sets the recording mode to record a specific number of frames. """
         self.specFrames()
         self._widget.setNumExpositions(numFrames)
 
     @APIExport
-    def setRecModeSpecTime(self, secondsToRec):
+    def setRecModeSpecTime(self, secondsToRec: Union[int, float]) -> None:
         """ Sets the recording mode to record for a specific amount of time.
         """
         self.specTime()
         self._widget.setTimeToRec(secondsToRec)
 
     @APIExport
-    def setRecModeScanOnce(self):
+    def setRecModeScanOnce(self) -> None:
         """ Sets the recording mode to record a single scan. """
         self.recScanOnce()
 
     @APIExport
-    def setRecModeScanTimelapse(self, secondsToRec, freqSeconds):
+    def setRecModeScanTimelapse(self, secondsToRec: float, freqSeconds: float) -> None:
         """ Sets the recording mode to record a timelapse of scans. """
         self.recScanLapse()
         self._widget.setTimelapseTime(secondsToRec)
         self._widget.setTimelapseFreq(freqSeconds)
 
     @APIExport
-    def setRecModeUntilStop(self):
+    def setRecModeUntilStop(self) -> None:
         """ Sets the recording mode to record until recording is manually
         stopped. """
         self.untilStop()
 
     @APIExport
-    def setDetectorToRecord(self, detectorName):
+    def setDetectorToRecord(self, detectorName: Union[str, int]) -> None:
         """ Sets which detectors to record. One can also pass -1 as the
         argument to record the current detector, or -2 to record all detectors.
         """
         self._widget.setDetectorToCapture(detectorName)
 
     @APIExport
-    def setRecFilename(self, filename):
+    def setRecFilename(self, filename: Optional[str]) -> None:
         """ Sets the name of the file to record to. This only sets the name of
         the file, not the full path. One can also pass None as the argument to
         use a default time-based filename. """
@@ -343,7 +344,7 @@ class RecordingController(ImConWidgetController):
             self._widget.setCustomFilenameEnabled(False)
 
     @APIExport
-    def setRecFolder(self, folderPath):
+    def setRecFolder(self, folderPath: str) -> None:
         """ Sets the folder to save recordings into. """
         self._widget.setRecFolder(folderPath)
 
