+++ imswitch/imcontrol/controller/controllers/LaserController.py
@@ -1,3 +1,5 @@
+from typing import List, Union
+
 from imswitch.imcommon.model import APIExport
 from imswitch.imcontrol.model import configfiletools
 from imswitch.imcontrol.view import guitools
@@ -207,18 +209,18 @@ class LaserController(ImConWidgetController):
             self.settingAttr = False
 
     @APIExport
-    def getLaserNames(self):
+    def getLaserNames(self) -> List[str]:
         """ Returns the device names of all lasers. These device names can be
         passed to other laser-related functions. """
         return self._master.lasersManager.getAllDeviceNames()
 
     @APIExport
-    def setLaserActive(self, laserName, active):
+    def setLaserActive(self, laserName: str, active: bool) -> None:
         """ Sets whether the specified laser is powered on. """
         self._widget.setLaserActive(laserName, active)
 
     @APIExport
-    def setLaserValue(self, laserName, value):
+    def setLaserValue(self, laserName: str, value: Union[int, float]) -> None:
         """ Sets the value of the specified laser, in the units that the laser
         uses. """
         self._widget.setValue(laserName, value)
