+++ imswitch/imscripting/model/actions.py
@@ -2,7 +2,7 @@ import importlib.util
 import os
 import time
 
-from imswitch.imcommon.framework import FrameworkUtils
+from imswitch.imcommon.framework import Signal, FrameworkUtils
 from imswitch.imcommon.model import APIExport, generateAPI
 
 
@@ -14,7 +14,7 @@ class _Actions:
         self._scriptPath = scriptPath
 
     @APIExport
-    def importScript(self, path):
+    def importScript(self, path: str) -> None:
         """ Imports the script at the specified path (either absolute or
         relative to the main script) and returns it as a module variable. """
 
@@ -27,7 +27,7 @@ class _Actions:
         return script
 
     @APIExport
-    def getWaitForSignal(self, signal, pollIntervalSeconds=1):
+    def getWaitForSignal(self, signal: Signal, pollIntervalSeconds: float = 1.0) -> None:
         """ Returns a function that will wait for the specified signal to emit.
         The returned function will continuously check whether the signal has
         been emitted since its creation. The polling interval defaults to one
