### 2021-07-24 356b063
Add comment in shortcut.py
