### 2021-07-24 3d99cdb
add NapariBaseWidget in init
