### 2021-07-24 ba88a0e
Send napariViewer to NapariBaseWidgets
