### 2021-07-24 02cd032
Remove function from ULensesWidget
