### 2021-07-26 052eeef
Fix incorrect return typing for global-level API functions
