### 2021-07-26 6d83176
Allow user to select detectors to record from a checkbox list (#52)
