+++ imswitch/imcontrol/view/widgets/RecordingWidget.py
@@ -3,7 +3,7 @@ import time
 
 from qtpy import QtCore, QtWidgets
 
-from imswitch.imcontrol.view import guitools as guitools
+from imswitch.imcontrol.view import guitools
 from .basewidgets import Widget
 
 
@@ -11,7 +11,8 @@ class RecordingWidget(Widget):
     """ Widget to control image or sequence recording.
     Recording only possible when liveview active. """
 
-    sigDetectorChanged = QtCore.Signal()
+    sigDetectorModeChanged = QtCore.Signal()
+    sigDetectorSpecificChanged = QtCore.Signal()
     sigOpenRecFolderClicked = QtCore.Signal()
     sigSpecFileToggled = QtCore.Signal(bool)  # (enabled)
     sigSnapRequested = QtCore.Signal()
@@ -30,7 +31,10 @@ class RecordingWidget(Widget):
         recTitle.setTextFormat(QtCore.Qt.RichText)
 
         # Detector list
-        self.detectorList = QtWidgets.QComboBox()
+        self.detectorModeList = QtWidgets.QComboBox()
+        self.detectorList = guitools.CheckableComboBox()
+        self.detectorList.setItemTypeName(singular='detector', plural='detectors')
+        self.detectorList.setVisible(False)
 
         # Folder and filename fields
         baseOutputFolder = self._options.recording.outputFolder
@@ -99,34 +103,57 @@ class RecordingWidget(Widget):
         self.setLayout(layout)
 
         recGrid = QtWidgets.QGridLayout()
+        gridRow = 0
 
-        recGrid.addWidget(recTitle, 0, 0, 1, 3)
-        recGrid.addWidget(QtWidgets.QLabel('Detector to capture'), 1, 0)
-        recGrid.addWidget(self.detectorList, 1, 1, 1, 4)
-        recGrid.addWidget(QtWidgets.QLabel('Folder'), 2, 0)
-        recGrid.addWidget(self.folderEdit, 2, 1, 1, 3)
-        recGrid.addWidget(self.openFolderButton, 2, 4)
-        recGrid.addWidget(self.filenameEdit, 3, 1, 1, 3)
+        recGrid.addWidget(recTitle, gridRow, 0, 1, 3)
+        gridRow += 1
 
-        recGrid.addWidget(self.specifyfile, 3, 0)
+        recGrid.addWidget(QtWidgets.QLabel('Detector to capture'), gridRow, 0)
+        recGrid.addWidget(self.detectorModeList, gridRow, 1, 1, 4)
+        gridRow += 1
 
-        recGrid.addWidget(modeTitle, 4, 0)
-        recGrid.addWidget(self.specifyFrames, 5, 0, 1, 5)
-        recGrid.addWidget(self.currentFrame, 5, 1)
-        recGrid.addWidget(self.numExpositionsEdit, 5, 2)
-        recGrid.addWidget(self.specifyTime, 6, 0, 1, 5)
-        recGrid.addWidget(self.currentTime, 6, 1)
-        recGrid.addWidget(self.timeToRec, 6, 2)
-        recGrid.addWidget(self.tRemaining, 6, 3, 1, 2)
-        recGrid.addWidget(self.recScanOnceBtn, 7, 0, 1, 5)
-        recGrid.addWidget(self.recScanLapseBtn, 8, 0, 1, 5)
-        recGrid.addWidget(self.currentLapse, 8, 1)
-        recGrid.addWidget(self.timeLapseEdit, 8, 2)
-        recGrid.addWidget(self.freqLabel, 8, 3)
-        recGrid.addWidget(self.freqEdit, 8, 4)
[…]
