+++ imswitch/imcommon/view/guitools/__init__.py
@@ -1,5 +1,6 @@
 from .BetterPushButton import BetterPushButton
 from .BetterSlider import BetterSlider
+from .CheckableComboBox import CheckableComboBox
 from .colorutils import wavelengthToHex
 from .dialogtools import askYesNoQuestion, askForFilePath, askForFolderPath, askForTextInput
 from .EmbeddedNapari import EmbeddedNapari

[…]
