### 2021-07-26 753a9b5
Fix double newlines in saved scripts on Windows
