### 2021-07-26 e948dfc
Fix load/save scan API methods using incorrect working directory
