### 2021-07-26 9c6ec3e
Fix setScanButtonChecked not having any effect
