### 2021-07-26 cf4944f
Clarify in documentation that device names must be unique
