### 2021-07-26 62d1022
Rename some NidaqManager signals
