### 2021-07-26 e79e4de
Fix scan laser preset not being saved in HDF5 files
