### 2021-07-26 ff39e2a
Change sleep duration for ScanOnce to match ScanLapse
