### 2021-07-26 2c11bcf
Remove "Close Window" and "Exit" entries from napari menu bar
