+++ imswitch/imcommon/view/guitools/EmbeddedNapari.py
@@ -1,10 +1,23 @@
 import napari
+from napari.utils.translations import trans
+from qtpy import QtWidgets
 
 
 class EmbeddedNapari(napari.Viewer):
     def __init__(self, *args, show=False, **kwargs):
         super().__init__(*args, show=show, **kwargs)
+
+        # Make menu bar not native
         self.window._qt_window.menuBar().setNativeMenuBar(False)
+
+        # Remove unwanted menu bar items
+        menuChildren = self.window._qt_window.findChildren(QtWidgets.QAction)
+        for menuChild in menuChildren:
+            try:
+                if menuChild.text() in [trans._('Close Window'), trans._('Exit')]:
+                    self.window.file_menu.removeAction(menuChild)
+            except Exception:
+                pass
 
     def getWidget(self):
         return self.window._qt_window
