### 2021-07-26 8b65756
Make critical napari viewer layers non-deletable (#54)
