### 2021-07-26 d39f848
Allow excluding TTL devices from scan and unlock such lasers (#51)
