### 2021-07-26 dfccec0
Remove limitation that SpecFrames mode not supported with 2+ detectors
