### 2021-07-26 5a62a67
Improve NidaqManager error handling
