+++ imswitch/imcontrol/model/managers/NidaqManager.py
@@ -3,6 +3,7 @@ Created on Tue Apr  7 16:46:33 2020
 
 """
 import operator
+import traceback
 import warnings
 
 import nidaqmx
@@ -19,6 +20,8 @@ class NidaqManager(SignalInterface):
     sigScanBuilt = Signal(object)  # (scanInfoDict)
     sigScanStarted = Signal()
     sigScanDone = Signal()
+    
+    sigScanBuildFailed = Signal()
 
     def __init__(self, setupInfo):
         super().__init__()
@@ -239,7 +242,7 @@ class NidaqManager(SignalInterface):
                     self.busy = False
 
     def runScan(self, signalDic, scanInfoDict):
-        """ Function assuming that the user wants to run a full scan with a stage 
+        """ Function assuming that the user wants to run a full scan with a stage
         controlled by analog voltage outputs and a cycle of TTL pulses continuously
         running. """
         if not self.busy:
@@ -247,105 +250,113 @@ class NidaqManager(SignalInterface):
             self.signalSent = False
             print('Create nidaq scan...')
 
-            # TODO: fill this
-            stageDic = signalDic['scanSignalsDict']
-            ttlDic = signalDic['TTLCycleSignalsDict']
-            AOTargetChanPairs = self.__makeSortedTargets('analogChannel')
+            try:
+                # TODO: fill this
+                stageDic = signalDic['scanSignalsDict']
+                ttlDic = signalDic['TTLCycleSignalsDict']
+                AOTargetChanPairs = self.__makeSortedTargets('analogChannel')
 
-            AOsignals = []
-            AOchannels = []
+                AOsignals = []
+                AOchannels = []
 
-            for pair in AOTargetChanPairs:
-                try:
-                    signal = stageDic[pair[0]]
-                    channel = pair[1]
-                    AOsignals.append(signal)
-                    AOchannels.append(channel)
-                except:
-                    pass
+                for pair in AOTargetChanPairs:
+                    try:
+                        signal = stageDic[pair[0]]
+                        channel = pair[1]
+                        AOsignals.append(signal)
+                        AOchannels.append(channel)
+                    except:
+                        pass
 
-            DOTargetChanPairs = self.__makeSortedTargets('digitalLine')
+                DOTargetChanPairs = self.__makeSortedTargets('digitalLine')
 
-            DOsignals = []
-            DOlines = []
+                DOsignals = []
+                DOlines = []
 
-            for pair in DOTargetChanPairs:
-                try:
-                    signal = ttlDic[pair[0]]
-                    line = pair[1]
-                    DOsignals.append(signal)
-                    DOlines.append(line)
-                except:
-                    pass
+                for pair in DOTargetChanPairs:
+                    try:
+                        signal = ttlDic[pair[0]]
+                        line = pair[1]
+                        DOsignals.append(signal)
+                        DOlines.append(line)
+                    except:
+                        pass
 
-            if len(AOsignals) < 1 and len(DOsignals) < 1:
+                if len(AOsignals) < 1 and len(DOsignals) < 1:
+                    raise NidaqManagerError('No signals to send')
+
+                # create task waiters and change constants for beginning scan
+                self.aoTaskWaiter = WaitThread()
+                self.doTaskWaiter = WaitThread()
+                if self.__timerCounterChannel is not None :
+                    self.timerTaskWaiter = WaitThread()
+                    # create timer counter output task, to control the acquisition timing (1 MHz)
+                    sampsInScan = np.int(len(AOsignals[0] if len(AOsignals) > 0 else DOsignals[0]) * 10)
+                    self.timerTask = self.__createChanCOTask(
+                        'TimerTask', channel=self.__timerCounterChannel, rate=1e6, sampsInScan=sampsInScan,
[…]
