### 2021-07-26 9b3cb53
Improve monkeypatch for napari layer removal
The previous method caused weird errors in certain situations, for example on Ubuntu with Python 3.9 after merging the shortcuts branch.
