### 2021-07-26 a68956b
Remove laser preset shortcuts according to Xavi's suggestion
