### 2021-07-26 eb89fc4
Fix load parameters from HDF5 file shortcut to conflicting with napari
