### 2021-07-26 5879e73
Change shortcuts to use Ctrl
