### 2021-07-26 1b48d66
Fix FFT not using correct starting image limits
