### 2021-07-27 kasasxav/ImSwitch#73 @uellue
Examples trigger error in fresh install without connected hardware
**Bug description**

Selecting the `example_monalisa.json` or `example_sted.json` in a fresh install without connected hardware triggers an error.

**How to reproduce**

1. Install current master of ImSwitch with `pip install -e .` from a Git clone
2. Wipe `Documents\ImSwitch` if present
3. Start `imswitch`
4. Select `example_monalisa.json` or `example_sted.json`

**Expected behavior**

The application works in a demo version or shows a concise error message on what is wrong and what the user can do to fix it.

**Screenshots**

![image](https://user-images.githubusercontent.com/25689052/127155327-b47aa140-5974-4fcb-8d02-ff00f2f75cb0.png)

**Environment**
 - OS: "Windows 10"
 - ImSwitch version: current master at this time
 - How you downloaded ImSwitch: `git clone ...` and `pip install -e`

**Additional context**
Add any other context about the problem here, if applicable.
