### 2021-07-27 kasasxav/ImSwitch#59 @uellue
Hi @stafak, thank you for the explanation and working on a fix! When I tried to test it, I ran into #73, unfortunately.
