### 2021-07-27 #3394 @uellue
@kasasxav Thank you for the documentation updates, a great improvement! I've checked the corresponding checkboxes.

When I wanted to test the fix for the "window problem" https://github.com/kasasxav/ImSwitch/issues/59 I ran into a new issue: https://github.com/kasasxav/ImSwitch/issues/73

Do you test these examples in your CI pipeline, actually? From my personal experience it is well worth the effort to automate checks for all examples to make sure the first impression for new users is good. :-)
