### 2021-07-27 a6782f1
Fix some NI-DAQmx errors not being changed to warnings (#73)
