### 2021-07-27 kasasxav/ImSwitch#73 @stafak
Whoops, looks like I missed a thing when I changed the error handling recently. Should be fixed by the commit above (available on master).
