### 2021-07-27 #3394 @stafak
Thank you @uellue! :) Sorry about the regression in kasasxav/ImSwitch#73, I think I've fixed it now.

We do test a setup similar to `example_monalisa.json` in our CI pipeline, but you're right that it would be good to run the tests for all examples. (However, in this case I believe we didn't catch the error for a different reason. From the error message, it looks like it happened specifically because NI-DAQmx drivers are installed on your computer but the devices are not connected, but the CI pipeline environment doesn't have the drivers installed at all.)
