### 2021-07-27 kasasxav/ImSwitch#59 @uellue
@stafak fix confirmed! For some reason the GUI element scaling is much better now, too. 🚀
