### 2021-07-27 #3394 @uellue
@stafak @kasasxav Thank you for fixing https://github.com/kasasxav/ImSwitch/issues/73 so quickly! Everything seems to work now. :-)

@Kevin-Mattheus-Moerman that concludes my review and I cordially recommend this paper for publication. 💯
