### 2021-07-27 3ce8c04
Fix laser "on" button being disabled by default
