### 2021-07-28 870dc6e
Proper path joining for correction pattern search
