### 2021-07-28 42b158e
Replace usage of deprecated python/numpy/pyqtgraph functionality
