+++ imswitch/imcontrol/model/managers/DetectorsManager.py
@@ -1,4 +1,3 @@
-from collections import Sequence
 from time import sleep
 
 import numpy as np
