+++ imswitch/imreconstruct/view/DataEditDialog.py
@@ -23,7 +23,7 @@ class DataEditDialog(QtWidgets.QDialog):
         self.imgVb = imageWidget.addViewBox(row=0, col=0)
         self.imgVb.setMouseMode(pg.ViewBox.PanMode)
         self.img = pg.ImageItem(axisOrder='row-major')
-        self.img.translate(-0.5, -0.5)
+        self.img.setTransform(self.img.transform().translate(-0.5, -0.5))
         self.imgVb.addItem(self.img)
         self.imgVb.setAspectLocked(True)
         self.imgHist = pg.HistogramLUTItem(image=self.img)
@@ -58,7 +58,7 @@ class DataEditDialog(QtWidgets.QDialog):
         self.dfVb = dfWidget.addViewBox(row=0, col=0)
         self.dfVb.setMouseMode(pg.ViewBox.PanMode)
         self.df = pg.ImageItem(axisOrder='row-major')
-        self.df.translate(-0.5, -0.5)
+        self.df.setTransform(self.df.transform().translate(-0.5, -0.5))
         self.dfVb.addItem(self.df)
         self.dfVb.setAspectLocked(True)
         self.dfHist = pg.HistogramLUTItem(image=self.df)
