### 2021-07-28 #3394 @beniroquai
@Kevin-Mattheus-Moerman I have also finished my test on including external hardware successfully. Very neat software!
