+++ imswitch/imcontrol/model/signaldesigners/PointScanTTLCycleDesigner.py
@@ -33,29 +33,34 @@ class PointScanTTLCycleDesigner(TTLCycleDesigner):
         else:
             signal_dict = {}
 
-            sample_rate = setupInfo.scan.sampleRate
+            # sample_rate = setupInfo.scan.sampleRate
             targets = parameterDict['target_device']
-            samples_pixel = parameterDict['sequence_time'] * sample_rate
-            pixels_line = scanInfoDict['pixels_line']
+            # samples_pixel = parameterDict['sequence_time'] * sample_rate
+            # pixels_line = scanInfoDict['pixels_line']
             samples_line = scanInfoDict['scan_samples_line']
             samples_total = scanInfoDict['scan_samples_total']
 
-            onepad_extraon = 2  # extra ON to make sure the laser is on at the start and end of the line (due to rise/fall time) (if it is ON there initially)
+            # extra ON to make sure the laser is on at the start and end of the line (due to
+            # rise/fall time) (if it is ON there initially)
+            onepad_extraon = 2
             zeropad_syncdelay = 0  # extra delay to sync with actual galvo positions
-            zeropad_lineflyback = scanInfoDict['scan_samples_period'] - scanInfoDict['scan_samples_line'] - onepad_extraon
+            zeropad_lineflyback = (scanInfoDict['scan_samples_period'] -
+                                   scanInfoDict['scan_samples_line'] -
+                                   onepad_extraon)
             zeropad_initpos = scanInfoDict['scan_throw_initpos']
             zeropad_settling = scanInfoDict['scan_throw_settling']
             zeropad_start = scanInfoDict['scan_throw_startzero']
             zeropad_startacc = scanInfoDict['scan_throw_startacc']
-            zeropad_finalpos = scanInfoDict['scan_throw_finalpos']
+            # zeropad_finalpos = scanInfoDict['scan_throw_finalpos']
             # Tile and pad TTL signals according to fast axis scan parameters
             for i, target in enumerate(targets):
                 signal_line = np.zeros(samples_line, dtype='bool')
-                for j, (start, end) in enumerate(zip(parameterDict['TTL_start'][i], parameterDict['TTL_end'][i])):
-                    start = start*1e3
-                    end = end*1e3
+                for j, (start, end) in enumerate(zip(parameterDict['TTL_start'][i],
+                                                     parameterDict['TTL_end'][i])):
+                    start = start * 1e3
+                    end = end * 1e3
                     start_on = min(np.int(np.round(start * samples_line)), samples_line)
-                    end_on =  min(np.int(np.round(end * samples_line)), samples_line)
+                    end_on = min(np.int(np.round(end * samples_line)), samples_line)
                     signal_line[start_on:end_on] = True
 
                 if signal_line[0]:
@@ -63,17 +68,24 @@ class PointScanTTLCycleDesigner(TTLCycleDesigner):
                 else:
                     signal_line = np.append(np.zeros(onepad_extraon, dtype='bool'), signal_line)
                 signal_period = np.append(signal_line, np.zeros(zeropad_lineflyback, dtype='bool'))
-                #TODO: # only do 2D-scan for now, fix for 3D-scan
-                signal = np.tile(signal_period, scanInfoDict['n_lines'] - 1)  # all lines except last
-                signal = np.append(signal, signal_line)  # add last line (does without flyback)
+                # TODO: # only do 2D-scan for now, fix for 3D-scan
 
-                signal = np.append(np.zeros(zeropad_syncdelay, dtype='bool'), signal)  # pad a delay for synchronizing scan position with TTL
-                signal = np.append(np.zeros(zeropad_startacc, dtype='bool'), signal)  # pad first line acceleration
[…]
