+++ imswitch/imcontrol/model/interfaces/hamamatsu_mock.py
@@ -75,11 +75,11 @@ class MockHamamatsu(Driver):
         self.s = Q_(1, 's')
 
         # Open the camera.
-#        self.camera_handle = ctypes.c_void_p(0)
-#        self.checkStatus(dcam.dcam_open(ctypes.byref(self.camera_handle),
-#                                        ctypes.c_int32(self.camera_id),
-#                                        None),
-#                         "dcam_open")
+        # self.camera_handle = ctypes.c_void_p(0)
+        # self.checkStatus(dcam.dcam_open(ctypes.byref(self.camera_handle),
+        #                                 ctypes.c_int32(self.camera_id),
+        #                                 None),
+        #                  "dcam_open")
         # Get camera properties.
         self.properties = {'Name': 'MOCK Hamamatsu',
                            'exposure_time': 10,  # * self.s,
@@ -265,21 +265,21 @@ class MockHamamatsu(Driver):
 
         # Some values are not changeable while the acquisition is running
         if (self.mock_acquisiton_running and
-            (property_name == 'subarray_vpos' or property_name == 'subarray_hpos' or
-             property_name == 'subarray_vsize' or property_name == 'subarray_hsize')):
+                (property_name == 'subarray_vpos' or property_name == 'subarray_hpos' or
+                 property_name == 'subarray_vsize' or property_name == 'subarray_hsize')):
             raise Exception('Value not changeable while acquisition is running')
 
         # If the value is text, figure out what the
         # corresponding numerical property value is.
 
         self.properties[property_name] = property_value
-#        print(property_name, 'set to:', self.properties[property_name])
-#            if (property_value in text_values):
-#                property_value = float(text_values[property_value])
-#            else:
-#                print(" unknown property text value:", property_value, "for",
-#                      property_name)
-#                return False
+        # print(property_name, 'set to:', self.properties[property_name])
+        #     if (property_value in text_values):
+        #         property_value = float(text_values[property_value])
+        #     else:
+        #         print(" unknown property text value:", property_value, "for",
+        #               property_name)
+        #         return False
         return property_value
 
     # setSubArrayMode
