+++ imswitch/imcontrol/model/managers/lasers/LaserManager.py
@@ -107,4 +107,3 @@ class LaserManager(ABC):
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
-

[…]
