### 2021-08-02 7181541
Add single file timelapse option (#53)
