+++ imswitch/imreconstruct/model/DataObj.py
@@ -6,13 +6,14 @@ import tifffile as tiff
 
 
 class DataObj:
-    def __init__(self, name, *, path=None, file=None):
+    def __init__(self, name, *, path=None, file=None, datasetName=None):
         self.name = name
         self.dataPath = path
         self.darkFrame = None
         self._meanData = None
         self._file = file
         self._data = None
+        self._datasetName = datasetName
         self._attrs = None
 
     @property
@@ -21,7 +22,7 @@ class DataObj:
             return self._data
 
         if isinstance(self._file, h5py.File):
-            self._data = np.array(self._file.get('data')[:])
+            self._data = np.array(self._file.get(self._datasetName)[:])
         elif isinstance(self._file, tiff.TiffFile):
             self._data = self._file.asarray()
 
@@ -42,13 +43,17 @@ class DataObj:
         return self.data is not None
 
     @property
+    def datasetName(self):
+        return self._datasetName
+
+    @property
     def numFrames(self):
         return np.shape(self.data)[0] if self.data is not None else None
 
     def checkAndLoadData(self):
         if not self.dataLoaded:
             try:
-                self._file = loadFromPath(self.dataPath)
+                self._file, self._datasetName = self._loadFromPath(self.dataPath, self._datasetName)
                 if self.data is not None:
                     print('Data loaded')
             except Exception:
@@ -76,19 +81,35 @@ class DataObj:
         return self._meanData
 
 
-def loadFromPath(path):
-    path = os.path.abspath(path)
-    try:
-        ext = os.path.splitext(path)[1]
-        if ext in ['.hdf5', '.hdf']:
-            return h5py.File(path, 'r')
-        elif ext in ['.tiff', '.tif']:
-            return tiff.TiffFile(path)
-        else:
-            raise ValueError(f'Unsupported file extension "{ext}"')
-    except Exception:
-        print('Error while loading data')
-        return None
+    @staticmethod
+    def _loadFromPath(path, datasetName=None):
+        path = os.path.abspath(path)
+        try:
+            ext = os.path.splitext(path)[1]
+            if ext in ['.hdf5', '.hdf']:
+                file = h5py.File(path, 'r')
+                if len(file) < 1:
+                    raise RuntimeError('File does not contain any datasets')
+                elif len(file) > 1 and datasetName is None:
+                    raise RuntimeError('File contains multiple datasets')
+
+                if datasetName is None:
+                    datasetName = list(file.keys())[0]
+
+                return file, datasetName
+            elif ext in ['.tiff', '.tif']:
+                return tiff.TiffFile(path), None
+            else:
+                raise ValueError(f'Unsupported file extension "{ext}"')
+        except Exception:
+            print('Error while loading data')
+            return None, None
+
+    def __eq__(self, other):
+        return (self.name == other.name and
+                self.dataPath == other.dataPath and
+                self._file.filename == other._file.filename and
+                self.datasetName == other.datasetName)
 
 
 # Copyright (C) 2020, 2021 TestaLab
