+++ imswitch/imcontrol/view/widgets/RecordingWidget.py
@@ -9,8 +9,7 @@ from .basewidgets import Widget
 
 
 class RecordingWidget(Widget):
-    """ Widget to control image or sequence recording.
-    Recording only possible when liveview active. """
+    """ Widget to control image or sequence recording. """
 
     sigDetectorModeChanged = QtCore.Signal()
     sigDetectorSpecificChanged = QtCore.Signal()
@@ -59,31 +58,33 @@ class RecordingWidget(Widget):
         self.recButton.setCheckable(True)
         self.recButton.setSizePolicy(QtWidgets.QSizePolicy.Preferred,
                                      QtWidgets.QSizePolicy.Expanding)
+
         # Number of frames and measurement timing
         modeTitle = QtWidgets.QLabel('<strong>Recording mode</strong>')
         modeTitle.setTextFormat(QtCore.Qt.RichText)
+
         self.specifyFrames = QtWidgets.QRadioButton('Number of frames')
+        self.currentFrame = QtWidgets.QLabel('0 /')
+        self.currentFrame.setAlignment((QtCore.Qt.AlignRight |
+                                        QtCore.Qt.AlignVCenter))
+        self.numExpositionsEdit = QtWidgets.QLineEdit('100')
+
         self.specifyTime = QtWidgets.QRadioButton('Time (s)')
+        self.currentTime = QtWidgets.QLabel('0 / ')
+        self.currentTime.setAlignment((QtCore.Qt.AlignRight |
+                                       QtCore.Qt.AlignVCenter))
+        self.timeToRec = QtWidgets.QLineEdit('1')
+
         self.recScanOnceBtn = QtWidgets.QRadioButton('Scan once')
+
         self.recScanLapseBtn = QtWidgets.QRadioButton('Time-lapse scan')
         self.currentLapse = QtWidgets.QLabel('0 / ')
         self.timeLapseEdit = QtWidgets.QLineEdit('5')
         self.freqLabel = QtWidgets.QLabel('Freq [s]')
         self.freqEdit = QtWidgets.QLineEdit('0')
-        self.stepSizeLabel = QtWidgets.QLabel('Step size [um]')
-        self.stepSizeEdit = QtWidgets.QLineEdit('0.05')
+        self.singleFileLapseBox = QtWidgets.QCheckBox('Save scans in a single file')
+
         self.untilSTOPbtn = QtWidgets.QRadioButton('Run until STOP')
-        self.timeToRec = QtWidgets.QLineEdit('1')
-        self.currentTime = QtWidgets.QLabel('0 / ')
-        self.currentTime.setAlignment((QtCore.Qt.AlignRight |
-                                       QtCore.Qt.AlignVCenter))
-        self.currentFrame = QtWidgets.QLabel('0 /')
-        self.currentFrame.setAlignment((QtCore.Qt.AlignRight |
-                                        QtCore.Qt.AlignVCenter))
-        self.numExpositionsEdit = QtWidgets.QLineEdit('100')
-        self.tRemaining = QtWidgets.QLabel()
-        self.tRemaining.setAlignment((QtCore.Qt.AlignCenter |
-                                      QtCore.Qt.AlignVCenter))
 
         self.saveModeLabel = QtWidgets.QLabel('<strong>Save mode:</strong>')
         self.saveModeList = QtWidgets.QComboBox()
@@ -112,7 +113,6 @@ class RecordingWidget(Widget):
         recGrid.addWidget(QtWidgets.QLabel('Detector to capture'), gridRow, 0)
         recGrid.addWidget(self.detectorModeList, gridRow, 1, 1, 4)
         gridRow += 1
-
[…]

[…]
