### 2021-08-02 714b383
Add single file multi-detector recording option (#53)
