### 2021-08-02 85c0bd6
Remove redundant getUniqueName function
