+++ imswitch/imcommon/model/filetools.py
@@ -1,42 +1,0 @@
-import glob
-import os
-
-
-# Check for same name conflict
-def getUniqueName(name):
-    name, ext = os.path.splitext(name)
-
-    n = 1
-    while glob.glob(name + ".*"):
-        if n > 1:
-            name = name.replace('_{}'.format(n - 1), '_{}'.format(n))
-        else:
-            name = insertSuffix(name, '_{}'.format(n))
-        n += 1
-
-    return ''.join((name, ext))
-
-
-def insertSuffix(filename, suffix, newExt=None):
-    names = os.path.splitext(filename)
-    if newExt is None:
-        return names[0] + suffix + names[1]
-    else:
-        return names[0] + suffix + newExt
-
-
-# Copyright (C) 2017 Federico Barabas
-# This file is part of Tormenta.
-#
-# Tormenta is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# Tormenta is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
