+++ imswitch/imreconstruct/model/DataObj.py
@@ -6,7 +6,7 @@ import tifffile as tiff
 
 
 class DataObj:
-    def __init__(self, name, *, path=None, file=None, datasetName=None):
+    def __init__(self, name, datasetName, *, path=None, file=None):
         self.name = name
         self.dataPath = path
         self.darkFrame = None
@@ -53,7 +53,7 @@ class DataObj:
     def checkAndLoadData(self):
         if not self.dataLoaded:
             try:
-                self._file, self._datasetName = self._loadFromPath(self.dataPath, self._datasetName)
+                self._file, self._datasetName = DataObj._open(self.dataPath, self._datasetName)
                 if self.data is not None:
                     print('Data loaded')
             except Exception:
@@ -80,9 +80,21 @@ class DataObj:
 
         return self._meanData
 
+    @staticmethod
+    def getDatasetNames(path):
+        file, _ = DataObj._open(path, allowMultipleDatasets=True)
+        try:
+            if isinstance(file, h5py.File):
+                return list(file.keys())
+            elif isinstance(file, tiff.TiffFile):
+                return ['default']
+            else:
+                raise ValueError(f'Unsupported file type "{type(file).__name__}"')
+        finally:
+            file.close()
 
     @staticmethod
-    def _loadFromPath(path, datasetName=None):
+    def _open(path, datasetName=None, allowMultipleDatasets=False):
         path = os.path.abspath(path)
         try:
             ext = os.path.splitext(path)[1]
@@ -90,10 +102,10 @@ class DataObj:
                 file = h5py.File(path, 'r')
                 if len(file) < 1:
                     raise RuntimeError('File does not contain any datasets')
-                elif len(file) > 1 and datasetName is None:
+                elif len(file) > 1 and datasetName is None and not allowMultipleDatasets:
                     raise RuntimeError('File contains multiple datasets')
 
-                if datasetName is None:
+                if datasetName is None and not allowMultipleDatasets:
                     datasetName = list(file.keys())[0]
 
                 return file, datasetName
@@ -106,9 +118,14 @@ class DataObj:
             return None, None
 
     def __eq__(self, other):
+        try:
+            sameFile = self._file == other.file or self._file.filename == other._file.filename
+        except AttributeError:
+            sameFile = False
+
         return (self.name == other.name and
                 self.dataPath == other.dataPath and
-                self._file.filename == other._file.filename and
+                sameFile and
                 self.datasetName == other.datasetName)
 
 
