### 2021-08-02 7c02403
Support for multi-dataset files in the metadata loading feature
Also changed attrs to be saved per dataset.
