### 2021-08-03 a25a7a7
Clarify that save mode is only for recordings
