### 2021-08-03 6a37b3c
Fix snap not working before turning on LiveView
