### 2021-08-04 d8d6022
Fix detector list in recording widget not working with 1 detector setups
