### 2021-08-04 c978598
Don't send scan signals through CommunicationChannel in cont. laser mode
