+++ imswitch/imcontrol/controller/controllers/ScanController.py
@@ -45,9 +45,11 @@ class ScanController(SuperScanController):
 
         # Connect NidaqManager signals
         self._master.nidaqManager.sigScanBuilt.connect(
-            lambda _, deviceList: self._commChannel.sigScanBuilt.emit(deviceList)
+            lambda _, deviceList: self.emitScanSignal(self._commChannel.sigScanBuilt, deviceList)
         )
-        self._master.nidaqManager.sigScanStarted.connect(self._commChannel.sigScanStarted)
+        self._master.nidaqManager.sigScanStarted.connect(
+            lambda: self.emitScanSignal(self._commChannel.sigScanStarted)
+        )
         self._master.nidaqManager.sigScanDone.connect(self.scanDone)
         self._master.nidaqManager.sigScanBuildFailed.connect(self.scanFailed)
 
@@ -187,21 +189,21 @@ class ScanController(SuperScanController):
             return
 
         if not sigScanStartingEmitted:
-            self._commChannel.sigScanStarting.emit()
+            self.emitScanSignal(self._commChannel.sigScanStarting)
         self._master.nidaqManager.runScan(self.signalDic, self.scanInfoDict)
 
     def scanDone(self):
         print('Scan done')
         if not self._widget.isContLaserMode() and not self._widget.continuousCheckEnabled():
             self._widget.setScanButtonChecked(False)
-            self._commChannel.sigScanEnded.emit()
+            self.emitScanSignal(self._commChannel.sigScanEnded)
         else:
             print('Repeat scan')
             self.runScanAdvanced(sigScanStartingEmitted=True)
 
     def scanFailed(self):
         self._widget.setScanButtonChecked(False)
-        self._commChannel.sigScanEnded.emit()
+        self.emitScanSignal(self._commChannel.sigScanEnded)
 
     def getParameters(self):
         if self.settingParameters:
@@ -283,6 +285,10 @@ class ScanController(SuperScanController):
                 ) if isLaser else '#ffffff'
             )
         self._widget.plotSignalGraph(areas, signals, colors, sampleRate)
+
+    def emitScanSignal(self, signal, *args):
+        if not self._widget.isContLaserMode():  # Don't emit in cont. laser mode
+            signal.emit(*args)
 
     def attrChanged(self, key, value):
         if self.settingAttr or len(key) != 2:
