### 2021-08-04 443d325
Rename confusing "continuousCheck" variable
