### 2021-08-04 d222051
Disable cont. laser pulses mode and repeating in recording scans
