### 2021-08-04 1bba63c
Fix duplicate datasets when saving multi-dataset recordings to RAM
Appeared in 3dbaae0c57cba5816ae5941e1a486f452da30ec9.
