+++ imswitch/imreconstruct/controller/ImRecMainViewController.py
@@ -37,7 +37,7 @@ class ImRecMainViewController(ImRecWidgetController):
         self._signalExtractor = SignalExtractor()
         self._patternFinder = PatternFinder()
 
-        self._currentData = None
+        self._currentDataObj = None
         self._pattern = self._widget.getPatternParams()
         self._settingPatternParams = False
         self._scanParDict = {
@@ -62,7 +62,12 @@ class ImRecMainViewController(ImRecWidgetController):
         self._widget.sigSetSaveFolder.connect(self.setSaveFolder)
 
         self._widget.sigReconstuctCurrent.connect(self.reconstructCurrent)
-        self._widget.sigReconstructMulti.connect(self.reconstructMulti)
+        self._widget.sigReconstructMultiConsolidated.connect(
+            lambda: self.reconstructMulti(consolidate=True)
+        )
+        self._widget.sigReconstructMultiIndividual.connect(
+            lambda: self.reconstructMulti(consolidate=False)
+        )
         self._widget.sigQuickLoadData.connect(self.quickLoadData)
         self._widget.sigUpdate.connect(lambda: self.updateScanParams(applyOnCurrentRecon=True))
 
@@ -92,10 +97,10 @@ class ImRecMainViewController(ImRecWidgetController):
 
     def findPattern(self):
         print('Find pattern clicked')
-        if self._currentData is None:
+        if self._currentDataObj is None:
             return
 
-        meanData = self._currentData.getMeanData()
+        meanData = self._currentDataObj.getMeanData()
         if len(meanData) < 1:
             return
 
@@ -165,19 +170,19 @@ class ImRecMainViewController(ImRecWidgetController):
                     return
 
             name = os.path.split(dataPath)[1]
-            if self._currentData is not None:
-                self._currentData.checkAndUnloadData()
-            self._currentData = DataObj(name, datasetToLoad, path=dataPath)
-            self._currentData.checkAndLoadData()
-            if self._currentData.dataLoaded:
-                self._commChannel.sigCurrentDataChanged.emit(self._currentData)
+            if self._currentDataObj is not None:
+                self._currentDataObj.checkAndUnloadData()
+            self._currentDataObj = DataObj(name, datasetToLoad, path=dataPath)
+            self._currentDataObj.checkAndLoadData()
+            if self._currentDataObj.dataLoaded:
+                self._commChannel.sigCurrentDataChanged.emit(self._currentDataObj)
                 print('Data loaded')
                 self._widget.raiseCurrentDataDock()
             else:
                 pass
 
     def currentDataChanged(self, dataObj):
-        self._currentData = dataObj
+        self._currentDataObj = dataObj
 
         # Update scan params based on new data
         # TODO: What if the attribute names change in imcontrol?
@@ -216,7 +221,7 @@ class ImRecMainViewController(ImRecWidgetController):
 
         self.updateScanParams()
 
-    def extractData(self):
+    def extractData(self, data):
         fwhmNm = self._widget.getFwhmNm()
         bgModelling = self._widget.getBgModelling()
         if bgModelling == 'Constant':
@@ -236,63 +241,64 @@ class ImRecMainViewController(ImRecWidgetController):
         device = self._widget.getComputeDevice()
         pattern = self._pattern
         if device == 'CPU' or device == 'GPU':
-            coeffs = self._signalExtractor.extractSignal(
-                self._currentData.data, sigmas, pattern, device.lower()
-            )
+            coeffs = self._signalExtractor.extractSignal(data, sigmas, pattern, device.lower())
         else:
             raise ValueError(f'Invalid device "{device}" specified; must be either "CPU" or "GPU"')
 
         return coeffs
 
     def reconstructCurrent(self):
[…]
