+++ imswitch/imreconstruct/view/ImRecMainView.py
@@ -19,7 +19,8 @@ class ImRecMainView(QtWidgets.QMainWindow):
     sigSetSaveFolder = QtCore.Signal()
 
     sigReconstuctCurrent = QtCore.Signal()
-    sigReconstructMulti = QtCore.Signal()
+    sigReconstructMultiConsolidated = QtCore.Signal()
+    sigReconstructMultiIndividual = QtCore.Signal()
     sigQuickLoadData = QtCore.Signal()
     sigUpdate = QtCore.Signal()
 
@@ -68,7 +69,8 @@ class ImRecMainView(QtWidgets.QMainWindow):
 
         btnFrame = BtnFrame()
         btnFrame.sigReconstuctCurrent.connect(self.sigReconstuctCurrent)
-        btnFrame.sigReconstructMulti.connect(self.sigReconstructMulti)
+        btnFrame.sigReconstructMultiConsolidated.connect(self.sigReconstructMultiConsolidated)
+        btnFrame.sigReconstructMultiIndividual.connect(self.sigReconstructMultiIndividual)
         btnFrame.sigQuickLoadData.connect(self.sigQuickLoadData)
         btnFrame.sigUpdate.connect(self.sigUpdate)
 
@@ -143,8 +145,8 @@ class ImRecMainView(QtWidgets.QMainWindow):
     def raiseMultiDataDock(self):
         self.multiDataDock.raiseDock()
 
-    def addNewData(self, reconObj):
-        self.reconstructionWidget.addNewData(reconObj)
+    def addNewData(self, reconObj, name):
+        self.reconstructionWidget.addNewData(reconObj, name)
 
     def getMultiDatas(self):
         dataList = self.multiDataFrame.dataList
@@ -233,29 +235,41 @@ class ReconParTree(ParameterTree):
 
 class BtnFrame(QtWidgets.QFrame):
     sigReconstuctCurrent = QtCore.Signal()
-    sigReconstructMulti = QtCore.Signal()
+    sigReconstructMultiConsolidated = QtCore.Signal()
+    sigReconstructMultiIndividual = QtCore.Signal()
     sigQuickLoadData = QtCore.Signal()
     sigUpdate = QtCore.Signal()
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
-        reconCurrBtn = BetterPushButton('Reconstruct current')
-        reconCurrBtn.clicked.connect(self.sigReconstuctCurrent)
-        reconMultiBtn = BetterPushButton('Reconstruct multidata')
-        reconMultiBtn.clicked.connect(self.sigReconstructMulti)
-        quickLoadDataBtn = BetterPushButton('Quick load data')
-        quickLoadDataBtn.clicked.connect(self.sigQuickLoadData)
-        updateBtn = BetterPushButton('Update reconstruction')
-        updateBtn.clicked.connect(self.sigUpdate)
+        self.reconCurrBtn = BetterPushButton('Reconstruct current')
+        self.reconCurrBtn.clicked.connect(self.sigReconstuctCurrent)
+        self.quickLoadDataBtn = BetterPushButton('Quick load data')
+        self.quickLoadDataBtn.clicked.connect(self.sigQuickLoadData)
+        self.updateBtn = BetterPushButton('Update reconstruction')
+        self.updateBtn.clicked.connect(self.sigUpdate)
+
+        self.reconMultiBtn = QtWidgets.QToolButton()
+        self.reconMultiBtn.setSizePolicy(
+            QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Preferred)
+        )
[…]
