### 2021-08-05 c753e98
Add "quick load data" menu bar item to imreconstruct
