### 2021-08-05 8af1c97
Fix issues when recording to same filename with different save modes
