### 2021-08-05 07eb19a
Perform number of frames check in reconstructMulti as well
