### 2021-08-05 3bcf871
Add option to save all reconstruction/coefficient objects (#34)
