+++ imswitch/imreconstruct/controller/ImRecMainViewController.py
@@ -57,7 +57,9 @@ class ImRecMainViewController(ImRecWidgetController):
         self._commChannel.sigScanParamsUpdated.connect(self.scanParamsUpdated)
 
         self._widget.sigSaveReconstruction.connect(lambda: self.saveCurrent('reconstruction'))
+        self._widget.sigSaveReconstructionAll.connect(lambda: self.saveAll('reconstruction'))
         self._widget.sigSaveCoeffs.connect(lambda: self.saveCurrent('coefficients'))
+        self._widget.sigSaveCoeffsAll.connect(lambda: self.saveAll('coefficients'))
         self._widget.sigSetDataFolder.connect(self.setDataFolder)
         self._widget.sigSetSaveFolder.connect(self.setSaveFolder)
 
@@ -303,66 +305,92 @@ class ImRecMainViewController(ImRecWidgetController):
             correctedData[i, :, :] = data[i, :, :] * c
         return correctedData
 
-    def saveCurrent(self, dataType=None):
-        """Saves the reconstructed image from self.reconstructor to specified
-        destination"""
-        if dataType:
-            saveName = guitools.askForFilePath(self._widget,
-                                               caption='Save File',
-                                               defaultFolder=self._saveFolder,
-                                               nameFilter='*.tiff', isSaving=True)
+    def saveCurrent(self, dataType):
+        """ Saves the reconstructed image or coefficeints from the current
+        ReconObj to a user-specified destination. """
 
-            if saveName:
-                if dataType == 'reconstruction':
-                    reconObj = self.reconstructionController.getActiveReconObj()
-                    scanParDict = reconObj.getScanParams()
-                    vxsizec = int(float(
-                        scanParDict['step_sizes'][scanParDict['dimensions'].index(
-                            self._widget.r_l_text
-                        )]
-                    ))
-                    vxsizer = int(float(
-                        scanParDict['step_sizes'][scanParDict['dimensions'].index(
-                            self._widget.u_d_text
-                        )]
-                    ))
-                    vxsizez = int(float(
-                        reconObj.scanParDict['step_sizes'][scanParDict['dimensions'].index(
-                            self._widget.b_f_text
-                        )]
-                    ))
-                    dt = int(float(
-                        scanParDict['step_sizes'][scanParDict['dimensions'].index(
-                            self._widget.timepoints_text
-                        )]
-                    ))
+        filePath = guitools.askForFilePath(self._widget,
+                                           caption=f'Save {dataType}',
+                                           defaultFolder=self._saveFolder or self._dataFolder,
+                                           nameFilter='*.tiff', isSaving=True)
 
-                    print(f'Trying to save to: {saveName}, Vx size: {vxsizec, vxsizer, vxsizez},'
-                          f' dt: {dt}')
-                    # Reconstructed image
-                    reconstrData = copy.deepcopy(reconObj.getReconstruction())
-                    reconstrData = reconstrData[:, 0, :, :, :, :]
-                    reconstrData = np.swapaxes(reconstrData, 1, 2)
-                    tiff.imwrite(saveName, reconstrData,
-                                 imagej=True, resolution=(1 / vxsizec, 1 / vxsizer),
-                                 metadata={'spacing': vxsizez, 'unit': 'nm', 'axes': 'TZCYX'})
-                elif dataType == 'coefficients':
-                    reconObj = self.reconstructionController.getActiveReconObj()
-                    coeffs = copy.deepcopy(reconObj.getCoeffs())
-                    print('Shape of coeffs = ', coeffs.shape)
-                    try:
[…]
