### 2021-08-05 45db54b
Use napari's dimension features instead of custom sliders (#70)
