+++ imswitch/imreconstruct/view/ReconstructionView.py
@@ -9,7 +9,7 @@ class ReconstructionView(QtWidgets.QFrame):
 
     # Signals
     sigItemSelected = QtCore.Signal()
-    sigImgSliceChanged = QtCore.Signal(int, int, int, int)
+    sigAxisStepChanged = QtCore.Signal(tuple)
     sigViewChanged = QtCore.Signal()
 
     # Methods
@@ -19,48 +19,12 @@ class ReconstructionView(QtWidgets.QFrame):
         # Image Widget
         guitools.addNapariGrayclipColormap()
         self.napariViewer = guitools.EmbeddedNapari()
+        self.napariViewer.dims.events.connect(self.dimsChanged)
         guitools.NapariUpdateLevelsWidget.addToViewer(self.napariViewer)
 
         self.imgLayer = self.napariViewer.add_image(
             np.zeros((1, 1)), rgb=False, name='Reconstruction', colormap='grayclip', protected=True
         )
-
-        # Slider and edit box for choosing slice
-        sliceLabel = QtWidgets.QLabel('Slice # ')
-        self.sliceNum = QtWidgets.QLabel('0')
-        self.sliceSlider = QtWidgets.QSlider(QtCore.Qt.Horizontal, self)
-        self.sliceSlider.setMinimum(0)
-        self.sliceSlider.setMaximum(0)
-        self.sliceSlider.setTickInterval(5)
-        self.sliceSlider.setSingleStep(1)
-        self.sliceSlider.valueChanged[int].connect(self.sliceSliderMoved)
-
-        baseLabel = QtWidgets.QLabel('Base # ')
-        self.baseNum = QtWidgets.QLabel('0')
-        self.baseSlider = QtWidgets.QSlider(QtCore.Qt.Horizontal, self)
-        self.baseSlider.setMinimum(0)
-        self.baseSlider.setMaximum(0)
-        self.baseSlider.setTickInterval(5)
-        self.baseSlider.setSingleStep(1)
-        self.baseSlider.valueChanged[int].connect(self.baseSliderMoved)
-
-        timeLabel = QtWidgets.QLabel('Time point # ')
-        self.timeNum = QtWidgets.QLabel('0')
-        self.timeSlider = QtWidgets.QSlider(QtCore.Qt.Horizontal, self)
-        self.timeSlider.setMinimum(0)
-        self.timeSlider.setMaximum(0)
-        self.timeSlider.setTickInterval(5)
-        self.timeSlider.setSingleStep(1)
-        self.timeSlider.valueChanged[int].connect(self.timeSliderMoved)
-
-        datasetLabel = QtWidgets.QLabel('Dataset # ')
-        self.datasetNum = QtWidgets.QLabel('0')
-        self.datasetSlider = QtWidgets.QSlider(QtCore.Qt.Horizontal, self)
-        self.datasetSlider.setMinimum(0)
-        self.datasetSlider.setMaximum(0)
-        self.datasetSlider.setTickInterval(5)
-        self.datasetSlider.setSingleStep(1)
-        self.datasetSlider.valueChanged[int].connect(self.datasetSliderMoved)
 
         # Button group for choosing view
         self.chooseViewGroup = QtWidgets.QButtonGroup()
@@ -102,21 +66,9 @@ class ReconstructionView(QtWidgets.QFrame):
 
         self.setLayout(layout)
 
-        layout.addWidget(self.napariViewer.get_widget(), 0, 0, 2, 1)
+        layout.addWidget(self.napariViewer.get_widget(), 0, 0, 4, 1)
         layout.addWidget(self.chooseViewBox, 0, 1, 1, 2)
[…]
