### 2021-08-06 b043f18
Use qdarkstyle's palette where possible in custom stylesheets
