### 2021-08-06 310b83a
Move RAM usage indicator to status bar
