+++ imswitch/imcommon/view/MultiModuleWindow.py
@@ -1,3 +1,4 @@
+from qdarkstyle import DarkPalette
 from qtpy import QtCore, QtWidgets
 
 from .AboutDialog import AboutDialog
@@ -27,6 +28,27 @@ class MultiModuleWindow(QtWidgets.QMainWindow):
         # Add tabs
         self.moduleTabs = QtWidgets.QTabWidget()
         self.moduleTabs.setTabPosition(QtWidgets.QTabWidget.TabPosition.West)
+
+        # RAM usage bar
+        self.memBarLabel = QtWidgets.QLabel('Current RAM usage:')
+        self.memBarLabel.setAlignment(QtCore.Qt.AlignRight | QtCore.Qt.AlignVCenter)
+        self.memBar = QtWidgets.QProgressBar()
+        self.memBar.setMaximum(100)  # Percentage
+        self.memBar.setStyleSheet(
+            f'min-width: 320px;'
+            f'border: 1px solid {DarkPalette.COLOR_BACKGROUND_5};'
+            f'background-color: {DarkPalette.COLOR_BACKGROUND_1};'
+        )
+
+        self.memBarContainer = QtWidgets.QWidget()
+        self.memBarContainer.setStyleSheet('background-color: transparent;')
+        self.memBarContainerLayout = QtWidgets.QHBoxLayout()
+        self.memBarContainerLayout.setContentsMargins(0, 2, 6, 4)
+        self.memBarContainerLayout.addWidget(self.memBarLabel, 1)
+        self.memBarContainerLayout.addWidget(self.memBar)
+        self.memBarContainer.setLayout(self.memBarContainerLayout)
+
+        self.statusBar().addPermanentWidget(self.memBarContainer)
 
         # Display loading screen until show(showLoadingScreen=False) is called
         loadingLabel = QtWidgets.QLabel('<h1>Starting ImSwitch…</h1>')
@@ -66,8 +88,11 @@ class MultiModuleWindow(QtWidgets.QMainWindow):
                 self.moduleTabs.setCurrentIndex(i)
                 return
 
-    def setLoadingProgress(self, progressFraction):
+    def updateLoadingProgress(self, progressFraction):
         self.loadingProgressBar.setValue(progressFraction * 100)
+
+    def updateRAMUsage(self, usageFraction):
+        self.memBar.setValue(round(usageFraction * 100))
 
     def showPickModulesDialogBlocking(self):
         result = self.pickModulesDialog.exec_()
