### 2021-08-06 1893e5d
Reduce time between program start and loading screen showing up
