+++ imswitch/imcommon/view/guitools/EmbeddedNapari.py
@@ -1,51 +1,0 @@
-import napari
-from napari.utils.translations import trans
-from qtpy import QtWidgets
-
-
-class EmbeddedNapari(napari.Viewer):
-    """ Napari viewer to be embedded in non-napari windows. Also includes a
-    feature to protect certain layers from being removed when added using
-    the add_image method. """
-
-    def __init__(self, *args, show=False, **kwargs):
-        super().__init__(*args, show=show, **kwargs)
-
-        # Monkeypatch layer removal methods
-        oldDelitemIndices = self.layers._delitem_indices
-
-        def newDelitemIndices(key):
-            indices = oldDelitemIndices(key)
-            for index in indices[:]:
-                layer = index[0][index[1]]
-                if hasattr(layer, 'protected') and layer.protected:
-                    indices.remove(index)
-            return indices
-
-        self.layers._delitem_indices = newDelitemIndices
-
-        # Make menu bar not native
-        self.window._qt_window.menuBar().setNativeMenuBar(False)
-
-        # Remove unwanted menu bar items
-        menuChildren = self.window._qt_window.findChildren(QtWidgets.QAction)
-        for menuChild in menuChildren:
-            try:
-                if menuChild.text() in [trans._('Close Window'), trans._('Exit')]:
-                    self.window.file_menu.removeAction(menuChild)
-            except Exception:
-                pass
-
-    def add_image(self, *args, protected=False, **kwargs):
-        result = super().add_image(*args, **kwargs)
-
-        if isinstance(result, list):
-            for layer in result:
-                layer.protected = protected
-        else:
-            result.protected = protected
-
-        return result
-
-    def get_widget(self):
-        return self.window._qt_window
