+++ imswitch/imcommon/view/guitools/__init__.py
@@ -1,18 +1,8 @@
 from .BetterPushButton import BetterPushButton
 from .BetterSlider import BetterSlider
 from .CheckableComboBox import CheckableComboBox
-from .EmbeddedNapari import EmbeddedNapari
 from .FloatSlider import FloatSlider
-from .colorutils import wavelengthToHex
 from .dialogtools import askYesNoQuestion, askForFilePath, askForFolderPath, askForTextInput
 from .imagetools import bestLevels, minmaxLevels
-from .naparitools import (
-    addNapariGrayclipColormap, NapariUpdateLevelsWidget, NapariShiftWidget, VispyROIVisual,
-    VispyLineVisual, VispyGridVisual, VispyCrosshairVisual, VispyScatterVisual
-)
-from .pyqtgraphtools import (
-    setPGBestImageLimits, PGGrid, PGTwoColorGrid, PGCrosshair, PGROI, PGCropROI, SumpixelsGraph,
-    ProjectionGraph
-)
 from .stylesheet import getBaseStyleSheet
 from .texttools import ordinalSuffix
