### 2021-08-06 1457f2d
Make creation of imcontrol widgets cleaner
