+++ imswitch/imcontrol/view/ImConMainView.py
@@ -46,7 +46,7 @@ class ImConMainView(QtWidgets.QMainWindow):
         self.setCentralWidget(self.cwidget)
 
         # Dock area
-        rightDocks = {
+        rightDockInfos = {
             'FocusLock': _DockInfo(name='Focus Lock', yPosition=0),
             'SLM': _DockInfo(name='SLM', yPosition=0),
             'Laser': _DockInfo(name='Laser Control', yPosition=0),
@@ -59,14 +59,14 @@ class ImConMainView(QtWidgets.QMainWindow):
             'ULenses': _DockInfo(name='uLenses Tool', yPosition=3),
             'FFT': _DockInfo(name='FFT Tool', yPosition=3)
         }
-        leftDocks = {
+        leftDockInfos = {
             'Settings': _DockInfo(name='Detector Settings', yPosition=0),
             'View': _DockInfo(name='Image Controls', yPosition=1),
             'Recording': _DockInfo(name='Recording', yPosition=2),
             'Console': _DockInfo(name='Console', yPosition=3)
         }
         otherDockKeys = ['Image']
-        allDockKeys = list(rightDocks.keys()) + list(leftDocks.keys()) + otherDockKeys
+        allDockKeys = list(rightDockInfos.keys()) + list(leftDockInfos.keys()) + otherDockKeys
 
         dockArea = DockArea()
         enabledDockKeys = viewSetupInfo.availableWidgets
@@ -75,63 +75,24 @@ class ImConMainView(QtWidgets.QMainWindow):
         elif enabledDockKeys is True:
             enabledDockKeys = allDockKeys
 
-        prevRightDock = None
-        prevRightDockYPosition = -1
-        def addRightDock(widgetKey, dockInfo):
-            nonlocal prevRightDock, prevRightDockYPosition
-            self.docks[widgetKey] = Dock(dockInfo.name, size=(1, 1))
-            widget = getattr(widgets, f'{widgetKey}Widget')
-            if issubclass(widget, guitools.NapariBaseWidget):
-                self.widgets[widgetKey] = self.factory.createWidget(
-                    widget, self.widgets['Image'].napariViewer
-                )
-            else:
-                self.widgets[widgetKey] = self.factory.createWidget(
-                    getattr(widgets, f'{widgetKey}Widget')
-                )
-            self.docks[widgetKey].addWidget(self.widgets[widgetKey])
-            if prevRightDock is None:
-                dockArea.addDock(self.docks[widgetKey])
-            elif dockInfo.yPosition > prevRightDockYPosition:
-                dockArea.addDock(self.docks[widgetKey], 'bottom', prevRightDock)
-            else:
-                dockArea.addDock(self.docks[widgetKey], 'above', prevRightDock)
-            prevRightDock = self.docks[widgetKey]
-            prevRightDockYPosition = dockInfo.yPosition
-
         if 'Image' in enabledDockKeys:
             self.docks['Image'] = Dock('Image Display', size=(1, 1))
             self.widgets['Image'] = self.factory.createWidget(widgets.ImageWidget)
             self.docks['Image'].addWidget(self.widgets['Image'])
+            self.factory.setArgument('napariViewer', self.widgets['Image'].napariViewer)
 
-        for widgetKey, dockInfo in rightDocks.items():
-            if widgetKey in enabledDockKeys:
-                addRightDock(widgetKey, dockInfo)
+        rightDocks = self._addDocks(
+            {k: v for k, v in rightDockInfos.items() if k in enabledDockKeys},
+            dockArea, 'right'
+        )
 
         if 'Image' in enabledDockKeys:
             dockArea.addDock(self.docks['Image'], 'left')
-        
[…]
