### 2021-08-06 49887d6
Restore uLenses to use VisPy visuals (due to napari points being slow)
