### 2021-08-06 916839e
Prepare NapariHybridWidget for future features
