### 2021-08-10 1ef8337
Fix QDarkStyle minimum version being too low
