+++ setup.cfg
@@ -33,7 +33,7 @@ install_requires =
     PyQt5 >= 5.15.2
     pyqtgraph >= 0.12.1
     pyserial >= 3.5
-    QDarkStyle >= 2.8
+    QDarkStyle >= 3
     QScintilla >= 2.12
     qtpy >= 1.9
     requests >= 2.25
