### 2021-08-10 88da593
Initial work on improved logging
Currently only applied on HamamatsuManager, Cobolt0601Manager, lantzlasers.py
