+++ imswitch/imcommon/model/__init__.py
@@ -1,4 +1,5 @@
 from .SharedAttributes import SharedAttributes
 from .VFileCollection import DataItem, VFileCollection
 from .api import APIExport, generateAPI
+from .logging import initLogger
 from .shortcut import shortcut, generateShortcuts
