+++ imswitch/imcontrol/model/managers/lasers/Cobolt0601LaserManager.py
@@ -1,5 +1,6 @@
 from lantz import Q_
 
+from imswitch.imcommon.model import initLogger
 from .LantzLaserManager import LantzLaserManager
 
 
@@ -14,6 +15,8 @@ class Cobolt0601LaserManager(LantzLaserManager):
     """
 
     def __init__(self, laserInfo, name, **_lowLevelManagers):
+        self.__logger = initLogger(self, __name__, instanceName=name)
+
         super().__init__(laserInfo, name, isBinary=False, valueUnits='mW', valueDecimals=0,
                          driver='cobolt.cobolt0601.Cobolt0601_f2', **_lowLevelManagers)
 
@@ -38,12 +41,12 @@ class Cobolt0601LaserManager(LantzLaserManager):
             powerQ = self._laser.power_sp * self._numLasers
             self._laser.enter_mod_mode()
             self._setModPower(powerQ)
-            print('Entered digital modulation mode')
-            print(f'Modulation mode is: {self._laser.mod_mode}')
+            self.__logger.debug('Entered digital modulation mode')
+            self.__logger.debug(f'Modulation mode is: {self._laser.mod_mode}')
         else:
             self._laser.digital_mod = False
             self._laser.query('cp')
-            print('Exited digital modulation mode')
+            self.__logger.debug('Exited digital modulation mode')
 
         self._digitalMod = active
 
@@ -52,7 +55,7 @@ class Cobolt0601LaserManager(LantzLaserManager):
 
     def _setModPower(self, power):
         self._laser.power_mod = power / self._numLasers
-        print(f'Set digital modulation mode power to: {power}')
+        self.__logger.debug(f'Set digital modulation mode power to: {power}')
 
 
 # Copyright (C) 2020, 2021 TestaLab
