+++ setup.cfg
@@ -19,6 +19,7 @@ packages = find:
 include_package_data = True
 python_requires = >=3.7
 install_requires =
+    coloredlogs >= 15
     colour-science >= 0.3
     dataclasses-json >= 0.5
     dotmap >= 1.3
