### 2021-08-11 53b9d7b
Slightly rework initLogger
