### 2021-08-11 87c5de8
Change prints to logging where not already done
