+++ imswitch/imcontrol/controller/MasterController.py
@@ -12,7 +12,6 @@ class MasterController:
     """
 
     def __init__(self, setupInfo, commChannel, moduleCommChannel):
-        print('init master controller')
         self.__setupInfo = setupInfo
         self.__commChannel = commChannel
         self.__moduleCommChannel = moduleCommChannel
