### 2021-08-11 ee79e15
Remove old hamamatsu test program
