+++ imswitch/imcontrol/model/interfaces/hamamatsu.py
@@ -891,86 +891,6 @@ class HamamatsuCameraMR(HamamatsuCamera):
 
 
 #
-# Testing.
-#
-
-if __name__ == "__main__":
-    initDcam()
-
-    print("found:", n_cameras, "cameras")
-    if (n_cameras > 0):
-
-        hcam = HamamatsuCameraMR(0)
-        print(hcam.setPropertyValue("defect_correct_mode", 1))
-        print("camera 0 model:", hcam.getModelInfo(0))
-
-        # List support properties.
-        if 1:
-            print("Supported properties:")
-            props = hcam.getProperties()
-            for i, id_name in enumerate(sorted(props.keys())):
-                p_value, p_type = hcam.getPropertyValue(id_name)
-                p_readable, p_writable = hcam.getPropertyRW(id_name)
-                read_write = ""
-                if p_readable:
-                    read_write += "read"
-                if p_writable:
-                    read_write += ", write"
-                print("  ", i, ")", id_name, " = ", p_value, " type is:", p_type, ",", read_write)
-                text_values = hcam.getPropertyText(id_name)
-                if len(text_values) > 0:
-                    print("          option / value")
-                    for key in sorted(text_values, key=text_values.get):
-                        print("         ", key, "/", text_values[key])
-
-        # Test setting & getting some parameters.
-        if 0:
-            print(hcam.setPropertyValue("exposure_time", 0.001))
-
-            # print hcam.setPropertyValue("subarray_hsize", 2048)
-            # print hcam.setPropertyValue("subarray_vsize", 2048)
-            print(hcam.setPropertyValue("subarray_hpos", 512))
-            print(hcam.setPropertyValue("subarray_vpos", 512))
-            print(hcam.setPropertyValue("subarray_hsize", 1024))
-            print(hcam.setPropertyValue("subarray_vsize", 1024))
-
-            print(hcam.setPropertyValue("binning", "1x1"))
-            print(hcam.setPropertyValue("readout_speed", 2))
-
-            hcam.setSubArrayMode()
-            # hcam.startAcquisition()
-            # hcam.stopAcquisition()
-
-            params = ["internal_frame_rate",
-                      "timing_readout_time",
-                      "exposure_time"]
-
-            #                      "image_height",
-            #                      "image_width",
-            #                      "image_framebytes",
-            #                      "buffer_framebytes",
-            #                      "buffer_rowbytes",
-            #                      "buffer_top_offset_bytes",
-            #                      "subarray_hsize",
-            #                      "subarray_vsize",
-            #                      "binning"]
-            for param in params:
-                print(param, hcam.getPropertyValue(param)[0])
-
-        # Test acquisition.
-        if 0:
-            hcam.startAcquisition()
-            cnt = 1
-            for i in range(300):
-                (frames, dims) = hcam.getFrames()
-                for aframe in frames:
-                    print(cnt, aframe[0:5])
-                    cnt += 1
-
-            hcam.stopAcquisition()
-
-
-#
 # The MIT License
 #
 # Copyright (c) 2013 Zhuang Lab, Harvard University
