### 2021-08-11 2749561
Fix script termination not checking correctly if a scrípt is running
