### 2021-08-13 #3394 @Kevin-Mattheus-Moerman
@openjournals/dev can you see what is going on here? Paper compiles fine but the reference checking fails.
