### 2021-08-13 #3394 @Kevin-Mattheus-Moerman
@kasasxav I proofread the paper and it looks good. The references should be checked (see my attempt above :point_up_2:) but other than that I think it looks in good shape. At this point can you work on the steps below:

- [ ] Can you confirm what the version tag for this publication should be? Is the software still at `v1.0.0` or has it moved on?
- [ ] Can you archive a copy of the software on ZENODO? You can do this manually or follow these automated steps: https://guides.github.com/activities/citable-code/
- [ ] Once you have created the archive can you check and edit the meta-data where needed so it matches the paper meta-data. In particular make sure the title and author list are the same. Also the archived version should carry the same license. Note you can add orcid links to authors to the ZENODO listing too.
- [ ] Once the ZENODO archive is created and updated in terms of meta-data, please report back here with the archive DOI.

Thanks
