### 2021-08-13 57a0960
Remove empty fields from EPICS reference in paper bibtex
