+++ paper/paper.bib
@@ -134,10 +134,7 @@ year = {2021}
 @article{epics,
 title = {EPICS architecture},
 author = {Dalesio, L R and Kozubal, A J and Kraimer, M R},
-doi = {},
-url = {https://www.osti.gov/biblio/6110347}, journal = {},
-number = ,
-volume = ,
+url = {https://www.osti.gov/biblio/6110347},
 place = {United States},
 year = {1991},
 month = {1}
