+++ docs/changelog.rst
@@ -3,6 +3,22 @@ Changelog
 *********
 
 
+1.2.0
+=====
+
+Highlights:
+
+- Saving multi-detector and timelapse recordings in a single file is now supported (#53)
+- Selecting specific detectors to record is now supported (#52)
+- It is now possible to edit values/on-off-state of non-involved lasers during scanning (#51)
+- The image reconstruction module now allows reconstructing all loaded data files (e.g. multi-file timelapses) into a single reconstruction (#34)
+- The documentation has been improved (#61, #63)
+- Fixed the SLM widget causing crashes on macOS (#57)
+- Fixed the module picker being empty in standalone Windows bundles (#55)
+
+A list of all code changes is available on GitHub: https://github.com/kasasxav/ImSwitch/compare/v1.1.0...v1.2.0
+
+
 1.1.0
 =====
 
