### 2021-08-13 #3394 @kasasxav
Thank you @Kevin-Mattheus-Moerman for the instructions and for your time. We would also like to thank the reviewers @untzag, @beniroquai, @uellue for your time and suggestions, it's truly exciting how the software has improved through this process!

The software version is now v1.2.0 and the Zenodo DOI 10.5281/zenodo.5196462 (https://zenodo.org/record/5196462#.YRaSXC0RoUs).
